Distribution of ADL (overall and by sex)

Distribution of ADL (overall and by sex)

Usage

chart_adl(
  x,
  save_chart = TRUE,
  filename = paste(tempdir(), "chart", sep = "/")
)

Arguments

x

Indicators dataset produced by create_op_all

save_chart

Logical. Should chart be saved? Default is TRUE.

filename

Prefix to add to output chart filename or a directory path to save output to instead of working directory. Defaults to a path to a temporary directory and a filename starting with chart. Ignored if save_chart is FALSE.

Value

Bar plot of ADL in PNG format saved in current working directory or in a specified directory if filename is a path unless when save_chart is FALSE in which case chart is shown in current graphic device.

Examples

  # Create chart using indicators.ALL dataset
  chart_adl(x = indicators.ALL)
#> pdf 
#>   2