createOP

Create older people indicators dataframe from survey survey data collected using the standard RAM-OP questionnaire.

createOP(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

Three dataframes of Older People indicators each with 138 variables:

indicators.ALL

Indicators dataframe for all respondents

indicators.MALES

Indicators dataframe for male respondents

indicators.FEMALES

Indicators dataframe for female repsondents

Indicators

Demographic

psu

Primary sampling unit

resp1

Respondent is SUBJECT

resp2

Respondent is FAMILY CARER

resp3

Respondent is OTHER CARER

resp4

Respondent is OTHER

age

Age of respondent (years)

ageGrp1

Age of respondent is between 50 and 59 years

ageGrp2

Age of respondent is between 50 and 59 years

ageGrp3

Age of respondent is between 50 and 59 years

ageGrp4

Age of respondent is between 50 and 59 years

ageGrp5

Age of respondent is between 50 and 59 years

sex1

Male

sex2

Female

marital1

Marital status = SINGLE

marital2

Marital status = MARRIED

marital3

Marital status = LIVING TOGETHER

marital4

Marital status = DIVORCED

marital5

Marital status = SEPARATED

marital6

Marital status = OTHER

alone

Respondent lives alone

Dietary intake indicators

These dietary intake indicators have been purpose-built for older people but the basic approach used is described in:

Kennedy G, Ballard T, Dop M C (2011). Guidelines for Measuring Household and Individual Dietary Diversity. Rome, FAO http://www.fao.org/docrep/014/i1983e/i1983e00.htm

and extended to include indicators of probable adequate intake of a number of nutrients / micronutrients.

MF

Meal frequenct

DDS

Dietary Diversity Score (count of 11 groups)

FG01

Cereals

FG02

Roots and tubers

FG03

Fruits and vegetables

FG04

All meat

FG05

Eggs

FG06

Fish

FG07

Legumes, nuts and seeds

FG08

Milk and milk products

FG09

Fats

FG10

Sugar

FG11

Other

proteinRich

Protein rich foods

pProtein

Protein rich plant sources of protein

aProtein

Protein rich animal sources of protein

pVitA

Plant sources of vitamin A

aVitA

Animal sources of vitamin A

xVitA

Any source of vitamin A

ironRich

Iron rich foods

caRich

Calcium rich foods

znRich

Zinc rich foods

vitB1

Vitamin B1-rich foods

vitB2

Vitamin B2-rich foods

vitB3

Vitamin B3-rich foods

vitB6

Vitamin B6-rich foods

vitB12

Vitamin B12-rich foods

vitBcomplex

Vitamin B1/B2/B3/B6/B12-rich foods

Household Hunger Scale (HHS)

The HHS is described in:

Ballard T, Coates J, Swindale A, Deitchler M (2011). Household Hunger Scale: Indicator Definition and Measurement Guide. Washington DC, FANTA-2 Bridge, FHI 360 https://www.fantaproject.org/monitoring-and-evaluation/household-hunger-scale-hhs

HHS1

Little or no hunger in household

HHS2

Moderate hunger in household

HHS3

Severe hunger in household

Katz "Index of Independence in Activities of Daily Living" (ADL) score

The Katz ADL score is described in:

Katz S, Ford AB, Moskowitz RW, Jackson BA, Jaffe MW (1963). Studies of illness in the aged. The Index of ADL: a standardized measure of biological and psychosocial function. JAMA, 1963, 185(12):914-9 doi:10.1001/jama.1963.03060120024016

Katz S, Down TD, Cash HR, Grotz, RC (1970). Progress in the development of the index of ADL. The Gerontologist, 10(1), 20-30 doi:10.1093/geront/10.4_Part_1.274

Katz S (1983). Assessing self-maintenance: Activities of daily living, mobility and instrumental activities of daily living. JAGS, 31(12), 721-726 doi:10.1111/j.1532-5415.1983.tb03391.x

ADL01

Bathing

ADL02

Dressing

ADL03

Toileting

ADL04

Transferring (mobility)

ADL05

Continence

ADL06

Feeding

scoreADL

ADL Score

classADL1

Severity of dependence 1

classADL2

Severity of dependence 2

classADL3

Severity of dependence 3

hasHelp

Have someone to help with everyday activities

unmetNeed

Need help but has no helper

K6 Short form psychological distress score

The K6 score is described in:

Kessler RC, Andrews G, Colpe LJ, Hiripi E, Mroczek, DK, Normand SLT, et al. (2002). Short screening scales to monitor population prevalences and trends in non-specific psychological distress. Psychological Medicine, 32(6), 959–976 doi:10.1017/S0033291702006074

K6

K6 score

K6Case

K6 score > 12 (in serious psychological distress)

Brief Community Screening Instrument for Dementia (CSID)

The CSID dementia screening tool is described in:

Prince M, et al. (2010). A brief dementia screener suitable for use by non-specialists in resource poor settings - The cross-cultural derivation and validation of the brief Community Screening Instrument for Dementia. International Journal of Geriatric Psychiatry, 26(9), 899–907 doi:10.1002/gps.2622

DS

Probable dementia by CSID screen

Health and health-seeking indicators

H1

Chronic condition

H2

Takes drugs regularly for chronic condition

H31

No drugs available

H32

Too expensive / no money

H33

Too old to look for care

H34

Use traditional medicine

H35

Drugs don't help

H36

No-one to help me

H37

No need

H38

Other

H39

No reason given

H4

Recent disease episode

H5

Accessed care for recent disease episode

H61

No drugs available

H62

Too expensive / no money

H63

Too old to look for care

H64

Use traditional medicine

H65

Drugs don't help

H66

No-one to help me

H67

No need

H68

Other

H69

No reason given

Income and income sources

M1

Has a personal income

M2A

Agriculture / fishing / livestock

M2B

Wages / salary

M2C

Sale of charcoal / bricks / &c.

M2D

Trading (e.g. market or shop)

M2E

Investments

M2F

Spending savings / sale of assets

M2G

Charity

M2H

Cash transfer / Social security

M2I

Other

Water, sanitation and hygiene (WASH) indicators

These are a (core) subset of indicators from:

WHO / UNICEF (2006). Core Questions on Drinking-water and Sanitation for Household Surveys. Geneva, WHO / UNICEF http://www.who.int/water_sanitation_health/monitoring/household_surveys/en/

W1

Improved source of drinking water

W2

Safe drinking water (improved source OR adequate treatment)

W3

Improved sanitation facility

W4

Improved non-shared sanitation facility

Anthropometry and screening

MUAC

Mid-upper arm circumference (mm)

oedema

Bilateral pitting oedema (may not be nutritional)

screened

Either MUAC or oedema checked previously

Visual impairment by "Tumbling E" method

The "Tumbling E" method is described in:

Taylor HR (1978). Applying new design principles to the construction of an illiterate E Chart. Am J Optom & Physiol Optics 55:348

poorVA

Poor visual acuity (correct in < 3 of 4 tests)

Miscellaneous indicators

chew

Problems chewing food

food

Anyone in HH receives a ration

NFRI

Anyone in HH received non-food relief item(s) in previous month

Washington Group on Disability

See:

http://www.washingtongroup-disability.com https://www.cdc.gov/nchs/washington_group/wg_documents.htm

for details.

wgVisionD0

Vision domain 0

wgVisionD1

Vision domain 1

wgVisionD2

Vision domain 2

wgVisionD3

Vision domain 3

wgHearingD0

Hearing domain 0

wgHearingD1

Hearing domain 1

wgHearingD2

Hearing domain 2

wgHearingD3

Hearing domain 3

wgMobilityD0

Mobility domain 0

wgMobilityD1

Mobility domain 1

wgMobilityD2

Mobility domain 2

wgMobilityD3

Mobility domain 3

wgRememberingD0

Remembering domain 0

wgRememberingD1

Remembering domain 1

wgRememberingD2

Remembering domain 2

wgRememberingD3

Remembering domain 3

wgSelfCareD0

Self-care domain 0

wgSelfCareD1

Self-care domain 1

wgSelfCareD2

Self-care domain 2

wgSelfCareD3

Self-care domain 3

wgCommunicatingD0

Communication domain 0

wgCommunicatingD1

Communication domain 1

wgCommunicatingD2

Communication domain 2

wgCommunicatingD3

Communication domain 3

wgP0

Overall 0

wgP1

Overall 1

wgP2

Overall 2

wgP3

Overall 3

wgPM

Any disability

Examples

# Create indicators dataset from RAM-OP survey data collected from
# Addis Ababa, Ethiopia
createOP(testSVY)

#>     psu resp1 resp2 resp3 resp4 age ageGrp1 ageGrp2 ageGrp3 ageGrp4 ageGrp5
#> 1   201     1     0     0     0  67       0       1       0       0       0
#> 2   201     1     0     0     0  74       0       0       1       0       0
#> 3   201     1     0     0     0  60       0       1       0       0       0
#> 4   201     1     0     0     0  60       0       1       0       0       0
#> 5   201     1     0     0     0  85       0       0       0       1       0
#> 6   201     0     1     0     0  86       0       0       0       1       0
#> 7   201     1     0     0     0  80       0       0       0       1       0
#>     sex1 sex2 marital1 marital2 marital3 marital4 marital5 marital6 alone MF
#> 1      0    1        0        0        0        0        1        0     0  3
#> 2      1    0        0        1        0        0        0        0     0  3
#> 3      1    0        0        1        0        0        0        0     0  2
#> 4      0    1        0        1        0        0        0        0     0  3
#> 5      0    1        0        0        0        0        1        0     0  3
#> 6      1    0        0        0        0        0        1        0     1  4
#> 7      1    0        0        0        0        0        1        0     0  3
#>     DDS FG01 FG02 FG03 FG04 FG05 FG06 FG07 FG08 FG09 FG10 FG11 proteinRich
#> 1     6    1    1    1    0    0    1    1    0    0    0    1           1
#> 2     7    1    1    1    0    0    0    1    0    1    1    1           1
#> 3     2    0    0    0    0    0    1    0    0    0    0    1           0
#> 4     3    1    1    0    0    0    0    1    0    0    0    0           1
#> 5     5    1    0    1    0    0    1    0    0    0    1    1           0
#> 6     5    1    1    0    0    0    1    1    0    0    0    1           1
#> 7     6    1    1    1    0    0    1    1    0    0    0    1           1
#>     pProtein aProtein pVitA aVitA xVitA ironRich caRich znRich vitB1 vitB2
#> 1          1        0     1     0     1        1      0      1     1     1
#> 2          1        0     1     0     1        1      0      1     1     1
#> 3          0        0     0     0     0        1      0      1     1     1
#> 4          1        0     0     0     0        0      0      1     1     1
#> 5          0        0     1     0     1        1      0      1     1     1
#> 6          1        0     0     0     0        1      0      1     1     1
#> 7          1        0     1     0     1        1      0      1     1     1
#>     vitB3 vitB6 vitB12 vitBcomplex HHS1 HHS2 HHS3 ADL01 ADL02 ADL03 ADL04 ADL05
#> 1       1     1      1           1    1    0    0     1     1     1     1     1
#> 2       1     1      0           0    1    0    0     1     1     1     1     0
#> 3       1     1      1           1    1    0    0     1     1     1     1     1
#> 4       1     1      0           0    1    0    0     1     1     1     1     1
#> 5       1     1      1           1    1    0    0     1     1     1     1     0
#> 6       1     1      1           1    0    1    0     1     1     1     1     1
#> 7       1     1      1           1    1    0    0     1     1     1     1     1
#>     ADL06 scoreADL classADL1 classADL2 classADL3 hasHelp unmetNeed K6 K6Case DS
#> 1       1        6         1         0         0       1         0  8      0  0
#> 2       1        5         1         0         0       1         0 18      1  0
#> 3       1        6         1         0         0       0         0 13      1  0
#> 4       1        6         1         0         0       0         0 14      1  1
#> 5       1        5         1         0         0       0         1 11      0  0
#> 6       1        6         1         0         0       1         0  0      0  0
#> 7       1        6         1         0         0       1         0  9      0  0
#>     H1 H2 H31 H32 H33 H34 H35 H36 H37 H38 H39 H4 H5 H61 H62 H63 H64 H65 H66 H67
#> 1    1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA  NA  NA  NA  NA
#> 2    1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA  NA  NA  NA  NA
#> 3    0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA  NA  NA  NA  NA
#> 4    0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA  NA  NA  NA  NA
#> 5    0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA  NA  NA  NA  NA
#> 6    0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA  NA  NA  NA  NA
#> 7    1  0   0   0   0   1   0   0   0   0   0  1  0   0   0   0   1   0   0   0
#>     H68 H69 M1 M2A M2B M2C M2D M2E M2F M2G M2H M2I W1 W2 W3 W4 MUAC oedema
#> 1    NA  NA  1   1   0   0   0   0   0   0   1   0  0  0  0  0  263      0
#> 2    NA  NA  1   1   0   0   0   0   0   0   1   0  1  1  0  0  231      1
#> 3    NA  NA  1   1   1   1   0   0   0   0   0   0  0  0  0  0  275      0
#> 4    NA  NA  1   1   1   0   1   0   0   0   0   0  1  1  1  1  264      0
#> 5    NA  NA  1   1   1   0   1   0   0   0   1   0  1  1  0  0  248      0
#> 6    NA  NA  1   0   0   0   0   0   0   0   1   0  0  0  0  0  251      0
#> 7     0   0  1   1   0   0   0   0   0   0   1   0  0  0  0  0  258      0
#>     screened poorVA chew food NFRI wgVisionD0 wgVisionD1 wgVisionD2 wgVisionD3
#> 1          0      0    0    1    0          1          0          0          0
#> 2          0      1    0    0    0          1          0          0          0
#> 3          0      0    0    0    0          1          0          0          0
#> 4          0      0    0    0    0          1          0          0          0
#> 5          0      0    1    0    0          1          0          0          0
#> 6          0      0    1    0    0          1          0          0          0
#> 7          0      1    1    0    0          1          0          0          0
#>     wgHearingD0 wgHearingD1 wgHearingD2 wgHearingD3 wgMobilityD0 wgMobilityD1
#> 1             1           0           0           0            1            0
#> 2             1           0           0           0            1            0
#> 3             1           0           0           0            1            0
#> 4             1           0           0           0            1            0
#> 5             1           0           0           0            1            0
#> 6             1           0           0           0            1            0
#> 7             1           0           0           0            1            0
#>     wgMobilityD2 wgMobilityD3 wgRememberingD0 wgRememberingD1 wgRememberingD2
#> 1              0            0               1               0               0
#> 2              0            0               1               0               0
#> 3              0            0               1               0               0
#> 4              0            0               1               0               0
#> 5              0            0               1               0               0
#> 6              0            0               1               0               0
#> 7              0            0               1               0               0
#>     wgRememberingD3 wgSelfCareD0 wgSelfCareD1 wgSelfCareD2 wgSelfCareD3
#> 1                 0            1            0            0            0
#> 2                 0            1            0            0            0
#> 3                 0            1            0            0            0
#> 4                 0            1            0            0            0
#> 5                 0            1            0            0            0
#> 6                 0            1            0            0            0
#> 7                 0            1            0            0            0
#>     wgCommunicatingD0 wgCommunicatingD1 wgCommunicatingD2 wgCommunicatingD3
#> 1                   1                 0                 0                 0
#> 2                   1                 0                 0                 0
#> 3                   1                 0                 0                 0
#> 4                   1                 0                 0                 0
#> 5                   1                 0                 0                 0
#> 6                   1                 0                 0                 0
#> 7                   1                 0                 0                 0
#>     wgP0 wgP1 wgP2 wgP3 wgPM
#> 1      1    0    0    0    0
#> 2      1    0    0    0    0
#> 3      1    0    0    0    0
#> 4      1    0    0    0    0
#> 5      1    0    0    0    0
#> 6      1    0    0    0    0
#> 7      1    0    0    0    0
#>  [ reached getOption("max.print") -- omitted 185 rows ]