Create older people indicators dataframe on activities of daily living from survey data collected using the standard RAM-OP questionnaire

Create older people indicators dataframe on activities of daily living from survey data collected using the standard RAM-OP questionnaire

Usage

create_op_adl(svy)

Arguments

svy

A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of older people indicators on activities of daily living

Katz "Index of Independence in Activities of Daily Living" (ADL) score

The Katz ADL score is described in:

Katz S, Ford AB, Moskowitz RW, Jackson BA, Jaffe MW (1963). Studies of illness in the aged. The Index of ADL: a standardized measure of biological and psychosocial function. JAMA, 1963, 185(12):914-9 doi:10.1001/jama.1963.03060120024016

Katz S, Down TD, Cash HR, Grotz, RC (1970). Progress in the development of the index of ADL. The Gerontologist, 10(1), 20-30 doi:10.1093/geront/10.4_Part_1.274

Katz S (1983). Assessing self-maintenance: Activities of daily living, mobility and instrumental activities of daily living. JAGS, 31(12), 721-726 doi:10.1111/j.1532-5415.1983.tb03391.x

ADL01

Bathing

ADL02

Dressing

ADL03

Toileting

ADL04

Transferring (mobility)

ADL05

Continence

ADL06

Feeding

scoreADL

ADL Score

classADL1

Severity of dependence 1

classADL2

Severity of dependence 2

classADL3

Severity of dependence 3

hasHelp

Have someone to help with everyday activities

unmetNeed

Need help but has no helper

Author

Mark Myatt

Examples

# Create activities of daily living indicators dataset from RAM-OP survey
# data collected from Addis Ababa, Ethiopia
create_op_adl(testSVY)
#> # A tibble: 192 × 15
#>      psu  sex1  sex2 ADL01 ADL02 ADL03 ADL04 ADL05 ADL06 score…¹ class…² class…³
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>   <dbl>   <dbl>   <dbl>
#>  1   201     0     1     1     1     1     1     1     1       6       1       0
#>  2   201     1     0     1     1     1     1     0     1       5       1       0
#>  3   201     1     0     1     1     1     1     1     1       6       1       0
#>  4   201     0     1     1     1     1     1     1     1       6       1       0
#>  5   201     0     1     1     1     1     1     0     1       5       1       0
#>  6   201     1     0     1     1     1     1     1     1       6       1       0
#>  7   201     1     0     1     1     1     1     1     1       6       1       0
#>  8   201     0     1     0     1     1     0     0     1       3       0       1
#>  9   201     1     0     1     1     1     1     0     1       5       1       0
#> 10   201     0     1     1     1     1     1     1     1       6       1       0
#> # … with 182 more rows, 3 more variables: classADL3 <dbl>, hasHelp <dbl>,
#> #   unmetNeed <dbl>, and abbreviated variable names ¹​scoreADL, ²​classADL1,
#> #   ³​classADL2