create_op_adl_females

Create female older people indicators dataframe for activities of daily living from survey data collected using the standard RAM-OP questionnaire

create_op_adl_females(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of female older people indicators on activities of daily living

Examples

# Create activities of daily living indicators dataset from RAM-OP survey
# data collected from Addis Ababa, Ethiopia
create_op_adl_females(testSVY)

#>     psu sex1 sex2 ADL01 ADL02 ADL03 ADL04 ADL05 ADL06 scoreADL classADL1
#> 1   201    0    1     1     1     1     1     1     1        6         1
#> 4   201    0    1     1     1     1     1     1     1        6         1
#> 5   201    0    1     1     1     1     1     0     1        5         1
#> 8   201    0    1     0     1     1     0     0     1        3         0
#> 10  201    0    1     1     1     1     1     1     1        6         1
#> 11  201    0    1     1     1     1     1     1     1        6         1
#> 12  201    0    1     1     1     1     0     0     1        4         0
#> 15  202    0    1     1     1     1     1     1     1        6         1
#> 16  202    0    1     1     1     1     0     1     1        5         1
#> 17  202    0    1     1     1     1     1     1     1        6         1
#> 18  202    0    1     1     1     1     1     0     1        5         1
#> 22  202    0    1     1     1     1     1     1     1        6         1
#> 23  202    0    1     1     1     1     1     1     1        6         1
#> 24  202    0    1     1     1     1     1     0     1        5         1
#> 25  203    0    1     1     1     1     1     1     1        6         1
#> 26  203    0    1     1     1     1     1     1     1        6         1
#> 28  203    0    1     1     1     1     1     1     1        6         1
#> 29  203    0    1     1     1     1     1     1     1        6         1
#> 30  203    0    1     1     1     1     1     1     1        6         1
#> 31  203    0    1     1     1     1     1     1     1        6         1
#> 32  203    0    1     1     1     1     1     0     1        5         1
#> 37  204    0    1     1     1     1     1     0     1        5         1
#> 38  204    0    1     1     1     1     1     1     1        6         1
#> 39  204    0    1     0     0     0     1     1     1        3         0
#> 41  204    0    1     0     1     1     1     1     1        5         1
#> 44  204    0    1     1     1     1     1     1     1        6         1
#> 45  204    0    1     1     1     1     1     1     1        6         1
#> 46  204    0    1     1     1     1     0     1     1        5         1
#> 47  204    0    1     1     1     1     0     1     1        5         1
#> 48  204    0    1     1     1     1     1     1     1        6         1
#> 49  205    0    1     1     1     1     1     1     1        6         1
#> 50  205    0    1     1     1     1     1     1     1        6         1
#> 51  205    0    1     1     1     1     1     1     1        6         1
#> 52  205    0    1     1     1     1     1     0     1        5         1
#> 53  205    0    1     1     1     1     1     1     1        6         1
#> 55  205    0    1     1     1     1     1     1     1        6         1
#> 56  205    0    1     1     1     1     1     1     1        6         1
#> 57  205    0    1     1     1     1     1     1     1        6         1
#> 58  205    0    1     1     1     1     1     0     1        5         1
#> 60  205    0    1     1     1     1     1     1     1        6         1
#> 62  206    0    1     1     1     1     1     1     1        6         1
#> 63  206    0    1     1     1     1     1     1     1        6         1
#> 64  206    0    1     1     1     1     0     1     1        5         1
#> 67  206    0    1     1     1     1     1     1     1        6         1
#> 72  206    0    1     1     1     1     1     1     1        6         1
#> 74  207    0    1     1     1     1     1     1     1        6         1
#> 76  207    0    1     1     1     1     1     1     1        6         1
#> 78  207    0    1     1     1     1     1     1     1        6         1
#> 79  207    0    1     1     1     1     1     1     1        6         1
#> 80  207    0    1     1     1     1     1     0     1        5         1
#> 82  207    0    1     1     1     1     1     1     1        6         1
#> 83  207    0    1     1     1     1     1     1     1        6         1
#> 85  208    0    1     1     1     1     1     1     1        6         1
#> 86  208    0    1     1     1     1     1     0     1        5         1
#> 88  208    0    1     1     1     1     1     1     1        6         1
#> 89  208    0    1     1     1     1     1     0     1        5         1
#> 91  208    0    1     1     1     1     1     1     1        6         1
#> 93  208    0    1     1     1     1     1     0     1        5         1
#> 94  208    0    1     1     1     1     1     1     1        6         1
#> 96  208    0    1     1     1     1     1     0     1        5         1
#> 97  209    0    1     1     1     1     1     0     1        5         1
#> 98  209    0    1     1     1     1     1     0     1        5         1
#> 99  209    0    1     1     1     1     1     1     1        6         1
#> 102 209    0    1     1     1     1     1     0     1        5         1
#> 109 210    0    1     1     1     1     1     1     1        6         1
#> 110 210    0    1     1     1     1     1     1     1        6         1
#> 111 210    0    1     1     1     1     1     1     1        6         1
#> 113 210    0    1     1     1     1     1     1     1        6         1
#> 115 210    0    1     1     1     1     1     1     1        6         1
#> 117 210    0    1     1     1     1     1     0     1        5         1
#> 122 211    0    1     1     1     1     1     0     1        5         1
#> 124 211    0    1     1     1     1     1     1     1        6         1
#> 125 211    0    1     1     1     1     1     1     1        6         1
#> 128 211    0    1     1     1     1     1     1     1        6         1
#> 129 211    0    1     1     1     1     1     1     1        6         1
#> 130 211    0    1     1     1     1     1     0     1        5         1
#> 131 211    0    1     1     1     1     1     0     1        5         1
#> 132 211    0    1     1     1     1     1     1     1        6         1
#> 133 212    0    1     1     1     1     1     1     1        6         1
#> 135 212    0    1     1     1     1     1     1     1        6         1
#> 137 212    0    1     1     1     1     1     1     1        6         1
#> 139 212    0    1     1     1     1     1     1     1        6         1
#> 140 212    0    1     1     1     1     1     1     1        6         1
#> 144 212    0    1     1     1     1     1     0     1        5         1
#> 146 213    0    1     1     1     1     1     1     1        6         1
#> 147 213    0    1     1     1     1     1     1     1        6         1
#> 148 213    0    1     1     1     1     1     1     1        6         1
#> 149 213    0    1     1     1     1     1     0     1        5         1
#> 153 213    0    1     1     1     1     1     1     1        6         1
#> 154 213    0    1     1     1     1     1     1     1        6         1
#> 155 213    0    1     1     1     1     1     1     1        6         1
#> 156 213    0    1     1     1     1     1     0     1        5         1
#> 157 214    0    1     1     1     1     1     1     1        6         1
#> 159 214    0    1     1     1     1     1     1     1        6         1
#> 162 214    0    1     1     1     1     1     0     1        5         1
#> 164 214    0    1     1     1     1     1     1     1        6         1
#> 166 214    0    1     1     1     1     1     0     1        5         1
#> 167 214    0    1     1     1     1     1     1     1        6         1
#> 168 214    0    1     1     1     1     1     0     1        5         1
#> 169 215    0    1     1     1     1     1     1     1        6         1
#> 170 215    0    1     1     1     1     1     1     1        6         1
#> 171 215    0    1     1     1     1     1     1     1        6         1
#> 175 215    0    1     1     1     1     1     1     1        6         1
#> 177 215    0    1     1     1     1     1     0     1        5         1
#> 178 215    0    1     1     1     1     1     0     1        5         1
#> 182 216    0    1     1     1     1     1     1     1        6         1
#> 184 216    0    1     1     1     1     1     1     1        6         1
#> 185 216    0    1     1     1     1     1     1     1        6         1
#> 186 216    0    1     1     1     1     1     1     1        6         1
#> 188 216    0    1     1     1     1     1     1     1        6         1
#> 190 216    0    1     1     1     1     1     1     1        6         1
#> 191 216    0    1     1     1     1     1     1     1        6         1
#> 192 216    0    1     1     1     1     1     0     1        5         1
#>     classADL2 classADL3 hasHelp unmetNeed
#> 1           0         0       1         0
#> 4           0         0       0         0
#> 5           0         0       0         1
#> 8           1         0       1         0
#> 10          0         0       1         0
#> 11          0         0       1         0
#> 12          1         0       1         0
#> 15          0         0       1         0
#> 16          0         0       0         1
#> 17          0         0       1         0
#> 18          0         0       1         0
#> 22          0         0       0         0
#> 23          0         0       0         0
#> 24          0         0       1         0
#> 25          0         0       1         0
#> 26          0         0       0         0
#> 28          0         0       1         0
#> 29          0         0       1         0
#> 30          0         0       0         0
#> 31          0         0       0         0
#> 32          0         0       1         0
#> 37          0         0       1         0
#> 38          0         0       1         0
#> 39          1         0       0         1
#> 41          0         0       1         0
#> 44          0         0       1         0
#> 45          0         0       1         0
#> 46          0         0       1         0
#> 47          0         0       1         0
#> 48          0         0       0         0
#> 49          0         0       1         0
#> 50          0         0       0         0
#> 51          0         0       0         0
#> 52          0         0       1         0
#> 53          0         0       1         0
#> 55          0         0       1         0
#> 56          0         0       1         0
#> 57          0         0       1         0
#> 58          0         0       1         0
#> 60          0         0       0         0
#> 62          0         0       0         0
#> 63          0         0       1         0
#> 64          0         0       1         0
#> 67          0         0       0         0
#> 72          0         0       1         0
#> 74          0         0       1         0
#> 76          0         0       1         0
#> 78          0         0       1         0
#> 79          0         0       0         0
#> 80          0         0       1         0
#> 82          0         0       1         0
#> 83          0         0       1         0
#> 85          0         0       1         0
#> 86          0         0       1         0
#> 88          0         0       0         0
#> 89          0         0       1         0
#> 91          0         0       0         0
#> 93          0         0       1         0
#> 94          0         0       1         0
#> 96          0         0       0         1
#> 97          0         0       0         1
#> 98          0         0       1         0
#> 99          0         0       1         0
#> 102         0         0       1         0
#> 109         0         0       0         0
#> 110         0         0       1         0
#> 111         0         0       1         0
#> 113         0         0       0         0
#> 115         0         0       1         0
#> 117         0         0       1         0
#> 122         0         0       1         0
#> 124         0         0       0         0
#> 125         0         0       1         0
#> 128         0         0       0         0
#> 129         0         0       0         0
#> 130         0         0       0         1
#> 131         0         0       0         1
#> 132         0         0       0         0
#> 133         0         0       1         0
#> 135         0         0       0         0
#> 137         0         0       1         0
#> 139         0         0       1         0
#> 140         0         0       1         0
#> 144         0         0       1         0
#> 146         0         0       0         0
#> 147         0         0       0         0
#> 148         0         0       0         0
#> 149         0         0       0         1
#> 153         0         0       0         0
#> 154         0         0       1         0
#> 155         0         0       1         0
#> 156         0         0       1         0
#> 157         0         0       1         0
#> 159         0         0       1         0
#> 162         0         0       0         1
#> 164         0         0       0         0
#> 166         0         0       1         0
#> 167         0         0       1         0
#> 168         0         0       1         0
#> 169         0         0       0         0
#> 170         0         0       0         0
#> 171         0         0       1         0
#> 175         0         0       0         0
#> 177         0         0       1         0
#> 178         0         0       0         1
#> 182         0         0       1         0
#> 184         0         0       1         0
#> 185         0         0       0         0
#> 186         0         0       1         0
#> 188         0         0       0         0
#> 190         0         0       1         0
#> 191         0         0       1         0
#> 192         0         0       1         0