Create male older people indicators dataframe for activities of daily living from survey data collected using the standard RAM-OP questionnaire

Create male older people indicators dataframe for activities of daily living from survey data collected using the standard RAM-OP questionnaire

create_op_adl_males(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of male older people indicators on activities of daily living

Examples

# Create activities of daily living indicators dataset from RAM-OP survey
# data collected from Addis Ababa, Ethiopia
create_op_adl_males(testSVY)

#>     psu sex1 sex2 ADL01 ADL02 ADL03 ADL04 ADL05 ADL06 scoreADL classADL1
#> 2   201    1    0     1     1     1     1     0     1        5         1
#> 3   201    1    0     1     1     1     1     1     1        6         1
#> 6   201    1    0     1     1     1     1     1     1        6         1
#> 7   201    1    0     1     1     1     1     1     1        6         1
#> 9   201    1    0     1     1     1     1     0     1        5         1
#> 13  202    1    0     1     1     1     1     1     1        6         1
#> 14  202    1    0     1     1     1     1     1     1        6         1
#> 19  202    1    0     1     1     1     0     1     1        5         1
#> 20  202    1    0     1     1     1     1     1     1        6         1
#> 21  202    1    0     1     1     1     1     1     1        6         1
#> 27  203    1    0     1     1     1     1     1     1        6         1
#> 33  203    1    0     1     1     1     1     1     1        6         1
#> 34  203    1    0     1     1     1     1     0     1        5         1
#> 35  203    1    0     1     1     1     1     1     1        6         1
#> 36  203    1    0     1     1     1     1     1     1        6         1
#> 40  204    1    0     1     1     1     1     1     1        6         1
#> 42  204    1    0     1     1     1     1     1     1        6         1
#> 43  204    1    0     1     1     1     1     0     1        5         1
#> 54  205    1    0     1     1     1     1     1     1        6         1
#> 59  205    1    0     1     1     1     1     0     1        5         1
#> 61  206    1    0     1     1     1     1     1     1        6         1
#> 65  206    1    0     1     1     1     1     1     1        6         1
#> 66  206    1    0     1     1     1     1     0     1        5         1
#> 68  206    1    0     1     1     1     1     1     1        6         1
#> 69  206    1    0     1     1     1     1     1     1        6         1
#> 70  206    1    0     1     1     1     1     0     1        5         1
#> 71  206    1    0     1     1     1     1     1     1        6         1
#> 73  207    1    0     1     1     1     1     1     1        6         1
#> 75  207    1    0     1     1     1     1     1     1        6         1
#> 77  207    1    0     1     1     1     1     1     1        6         1
#> 81  207    1    0     1     1     1     1     1     1        6         1
#> 84  207    1    0     1     1     1     1     1     1        6         1
#> 87  208    1    0     1     1     1     1     1     1        6         1
#> 90  208    1    0     1     1     1     1     1     1        6         1
#> 92  208    1    0     1     1     1     1     0     1        5         1
#> 95  208    1    0     0     1     1     1     1     1        5         1
#> 100 209    1    0     1     1     1     1     1     1        6         1
#> 101 209    1    0     1     1     1     1     1     1        6         1
#> 103 209    1    0     1     1     1     1     1     1        6         1
#> 104 209    1    0     1     1     1     1     1     1        6         1
#> 105 209    1    0     1     1     1     1     1     1        6         1
#> 106 209    1    0     1     1     1     1     1     1        6         1
#> 107 209    1    0     1     1     1     1     0     1        5         1
#> 108 209    1    0     1     1     1     1     1     1        6         1
#> 112 210    1    0     0     0     0     0     1     1        2         0
#> 114 210    1    0     1     1     1     1     1     1        6         1
#> 116 210    1    0     1     1     1     1     1     1        6         1
#> 118 210    1    0     1     1     1     1     1     1        6         1
#> 119 210    1    0     1     1     1     1     1     1        6         1
#> 120 210    1    0     1     1     1     1     1     1        6         1
#> 121 211    1    0     1     1     1     1     1     1        6         1
#> 123 211    1    0     1     1     1     1     0     1        5         1
#> 126 211    1    0     1     1     1     1     1     1        6         1
#> 127 211    1    0     1     1     1     1     1     1        6         1
#> 134 212    1    0     1     1     1     1     0     1        5         1
#> 136 212    1    0     1     1     1     1     1     1        6         1
#> 138 212    1    0     1     1     1     1     0     1        5         1
#> 141 212    1    0     1     1     1     1     1     1        6         1
#> 142 212    1    0     1     1     1     1     1     1        6         1
#> 143 212    1    0     1     1     1     1     1     1        6         1
#> 145 213    1    0     1     1     1     1     1     1        6         1
#> 150 213    1    0     1     1     1     1     0     1        5         1
#> 151 213    1    0     1     1     1     1     1     1        6         1
#> 152 213    1    0     1     1     1     1     1     1        6         1
#> 158 214    1    0     1     1     1     1     1     1        6         1
#> 160 214    1    0     1     1     1     1     1     1        6         1
#> 161 214    1    0     1     1     1     1     1     1        6         1
#> 163 214    1    0     1     1     1     1     1     1        6         1
#> 165 214    1    0     1     1     1     1     1     1        6         1
#> 172 215    1    0     1     1     1     1     1     1        6         1
#> 173 215    1    0     1     1     1     1     0     1        5         1
#> 174 215    1    0     1     1     1     1     0     1        5         1
#> 176 215    1    0     1     1     1     1     1     1        6         1
#> 179 215    1    0     1     1     1     1     1     1        6         1
#> 180 215    1    0     1     1     1     1     1     1        6         1
#> 181 216    1    0     1     1     1     1     0     1        5         1
#> 183 216    1    0     1     1     1     1     1     1        6         1
#> 187 216    1    0     1     1     1     1     1     1        6         1
#> 189 216    1    0     0     0     0     0     0     0        0         0
#>     classADL2 classADL3 hasHelp unmetNeed
#> 2           0         0       1         0
#> 3           0         0       0         0
#> 6           0         0       1         0
#> 7           0         0       1         0
#> 9           0         0       0         1
#> 13          0         0       1         0
#> 14          0         0       1         0
#> 19          0         0       1         0
#> 20          0         0       1         0
#> 21          0         0       1         0
#> 27          0         0       1         0
#> 33          0         0       1         0
#> 34          0         0       1         0
#> 35          0         0       1         0
#> 36          0         0       1         0
#> 40          0         0       0         0
#> 42          0         0       1         0
#> 43          0         0       0         1
#> 54          0         0       1         0
#> 59          0         0       1         0
#> 61          0         0       0         0
#> 65          0         0       0         0
#> 66          0         0       1         0
#> 68          0         0       1         0
#> 69          0         0       1         0
#> 70          0         0       0         1
#> 71          0         0       1         0
#> 73          0         0       1         0
#> 75          0         0       0         0
#> 77          0         0       1         0
#> 81          0         0       1         0
#> 84          0         0       0         0
#> 87          0         0       1         0
#> 90          0         0       0         0
#> 92          0         0       0         1
#> 95          0         0       1         0
#> 100         0         0       0         0
#> 101         0         0       0         0
#> 103         0         0       0         0
#> 104         0         0       0         0
#> 105         0         0       0         0
#> 106         0         0       1         0
#> 107         0         0       0         1
#> 108         0         0       1         0
#> 112         0         1       1         0
#> 114         0         0       1         0
#> 116         0         0       0         0
#> 118         0         0       0         0
#> 119         0         0       0         0
#> 120         0         0       1         0
#> 121         0         0       0         0
#> 123         0         0       1         0
#> 126         0         0       1         0
#> 127         0         0       0         0
#> 134         0         0       0         1
#> 136         0         0       0         0
#> 138         0         0       1         0
#> 141         0         0       1         0
#> 142         0         0       1         0
#> 143         0         0       1         0
#> 145         0         0       0         0
#> 150         0         0       0         1
#> 151         0         0       1         0
#> 152         0         0       0         0
#> 158         0         0       1         0
#> 160         0         0       1         0
#> 161         0         0       1         0
#> 163         0         0       1         0
#> 165         0         0       0         0
#> 172         0         0       0         0
#> 173         0         0       0         1
#> 174         0         0       0         1
#> 176         0         0       1         0
#> 179         0         0       0         0
#> 180         0         0       0         0
#> 181         0         0       0         1
#> 183         0         0       0         0
#> 187         0         0       1         0
#> 189         0         1       0         1