Create older people indicators dataframe from survey data collected using the standard RAM-OP questionnaire.

Create older people indicators dataframe from survey data collected using the standard RAM-OP questionnaire.

create_op_all(
  svy,
  indicators = c("demo", "food", "hunger", "disability", "adl", "mental", "dementia",
    "health", "income", "wash", "anthro", "oedema", "screening", "visual", "misc"),
  gender = c("m", "f")
)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

indicators
A character vector of indicator names

gender
Either an "m" for male or "f" for female; Whether to report indicators for males or females. If unspecified (default), both are reported.

Value

A tibble of older people indicators

Examples

create_op_all(svy = testSVY)

#> # A tibble: 192 x 138
#>      psu  sex1  sex2 resp1 resp2 resp3 resp4   age ageGrp1 ageGrp2 ageGrp3
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <int>   <dbl>   <dbl>   <dbl>
#>  1   201     0     1     1     0     0     0    67       0       1       0
#>  2   201     1     0     1     0     0     0    74       0       0       1
#>  3   201     1     0     1     0     0     0    60       0       1       0
#>  4   201     0     1     1     0     0     0    60       0       1       0
#>  5   201     0     1     1     0     0     0    85       0       0       0
#>  6   201     1     0     0     1     0     0    86       0       0       0
#>  7   201     1     0     1     0     0     0    80       0       0       0
#>  8   201     0     1     1     0     0     0    60       0       1       0
#>  9   201     1     0     1     0     0     0    62       0       1       0
#> 10   201     0     1     1     0     0     0    72       0       0       1
#> # … with 182 more rows, and 127 more variables: ageGrp4 <dbl>, ageGrp5 <dbl>,
#> #   marital1 <dbl>, marital2 <dbl>, marital3 <dbl>, marital4 <dbl>,
#> #   marital5 <dbl>, marital6 <dbl>, alone <dbl>, MF <dbl>, DDS <dbl>,
#> #   FG01 <dbl>, FG02 <dbl>, FG03 <dbl>, FG04 <dbl>, FG05 <dbl>, FG06 <dbl>,
#> #   FG07 <dbl>, FG08 <dbl>, FG09 <dbl>, FG10 <dbl>, FG11 <dbl>,
#> #   proteinRich <dbl>, pProtein <dbl>, aProtein <dbl>, pVitA <dbl>,
#> #   aVitA <dbl>, xVitA <dbl>, ironRich <dbl>, caRich <dbl>, znRich <dbl>,
#> #   vitB1 <dbl>, vitB2 <dbl>, vitB3 <dbl>, vitB6 <dbl>, vitB12 <dbl>,
#> #   vitBcomplex <dbl>, HHS1 <dbl>, HHS2 <dbl>, HHS3 <dbl>, wgVisionD0 <dbl>,
#> #   wgVisionD1 <dbl>, wgVisionD2 <dbl>, wgVisionD3 <dbl>, wgHearingD0 <dbl>,
#> #   wgHearingD1 <dbl>, wgHearingD2 <dbl>, wgHearingD3 <dbl>,
#> #   wgMobilityD0 <dbl>, wgMobilityD1 <dbl>, wgMobilityD2 <dbl>,
#> #   wgMobilityD3 <dbl>, wgRememberingD0 <dbl>, wgRememberingD1 <dbl>,
#> #   wgRememberingD2 <dbl>, wgRememberingD3 <dbl>, wgSelfCareD0 <dbl>,
#> #   wgSelfCareD1 <dbl>, wgSelfCareD2 <dbl>, wgSelfCareD3 <dbl>,
#> #   wgCommunicatingD0 <dbl>, wgCommunicatingD1 <dbl>, wgCommunicatingD2 <dbl>,
#> #   wgCommunicatingD3 <dbl>, wgP0 <dbl>, wgP1 <dbl>, wgP2 <dbl>, wgP3 <dbl>,
#> #   wgPM <dbl>, ADL01 <dbl>, ADL02 <dbl>, ADL03 <dbl>, ADL04 <dbl>,
#> #   ADL05 <dbl>, ADL06 <dbl>, scoreADL <dbl>, classADL1 <dbl>, classADL2 <dbl>,
#> #   classADL3 <dbl>, hasHelp <dbl>, unmetNeed <dbl>, K6 <dbl>, K6Case <dbl>,
#> #   DS <dbl>, H1 <dbl>, H2 <dbl>, H31 <dbl>, H32 <dbl>, H33 <dbl>, H34 <dbl>,
#> #   H35 <dbl>, H36 <dbl>, H37 <dbl>, H38 <dbl>, H39 <dbl>, H4 <dbl>, H5 <dbl>,
#> #   H61 <dbl>, H62 <dbl>, H63 <dbl>, …