Create female older people indicators dataframe for anthropometry from survey data collected using the standard RAM-OP questionnaire

Create female older people indicators dataframe for anthropometry from survey data collected using the standard RAM-OP questionnaire

create_op_anthro_females(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of female older people indicators on anthropometry

Examples

# Create anthropometry indicators dataset from RAM-OP survey data collected
# from Addis Ababa, Ethiopia
create_op_anthro_females(testSVY)

#>     psu sex1 sex2 MUAC
#> 1   201    0    1  263
#> 4   201    0    1  264
#> 5   201    0    1  248
#> 8   201    0    1  221
#> 10  201    0    1  227
#> 11  201    0    1  262
#> 12  201    0    1  209
#> 15  202    0    1  277
#> 16  202    0    1  276
#> 17  202    0    1  256
#> 18  202    0    1  328
#> 22  202    0    1  244
#> 23  202    0    1  263
#> 24  202    0    1  240
#> 25  203    0    1  213
#> 26  203    0    1  194
#> 28  203    0    1  291
#> 29  203    0    1  301
#> 30  203    0    1  315
#> 31  203    0    1  266
#> 32  203    0    1  203
#> 37  204    0    1  317
#> 38  204    0    1  253
#> 39  204    0    1  293
#> 41  204    0    1  236
#> 44  204    0    1  288
#> 45  204    0    1  269
#> 46  204    0    1  342
#> 47  204    0    1  311
#> 48  204    0    1  244
#> 49  205    0    1  380
#> 50  205    0    1  341
#> 51  205    0    1  229
#> 52  205    0    1  252
#> 53  205    0    1  294
#> 55  205    0    1  263
#> 56  205    0    1  232
#> 57  205    0    1  310
#> 58  205    0    1  257
#> 60  205    0    1  231
#> 62  206    0    1  305
#> 63  206    0    1  244
#> 64  206    0    1  227
#> 67  206    0    1  308
#> 72  206    0    1  239
#> 74  207    0    1  365
#> 76  207    0    1  234
#> 78  207    0    1  267
#> 79  207    0    1  244
#> 80  207    0    1  273
#> 82  207    0    1  301
#> 83  207    0    1  286
#> 85  208    0    1  292
#> 86  208    0    1  289
#> 88  208    0    1  257
#> 89  208    0    1  293
#> 91  208    0    1  272
#> 93  208    0    1  267
#> 94  208    0    1  329
#> 96  208    0    1  264
#> 97  209    0    1  241
#> 98  209    0    1  287
#> 99  209    0    1  223
#> 102 209    0    1  286
#> 109 210    0    1  232
#> 110 210    0    1  394
#> 111 210    0    1  312
#> 113 210    0    1  262
#> 115 210    0    1  252
#> 117 210    0    1  286
#> 122 211    0    1  258
#> 124 211    0    1  247
#> 125 211    0    1  366
#> 128 211    0    1  221
#> 129 211    0    1  322
#> 130 211    0    1  307
#> 131 211    0    1  333
#> 132 211    0    1  204
#> 133 212    0    1  256
#> 135 212    0    1  243
#> 137 212    0    1  267
#> 139 212    0    1  274
#> 140 212    0    1  303
#> 144 212    0    1  195
#> 146 213    0    1  250
#> 147 213    0    1  210
#> 148 213    0    1  293
#> 149 213    0    1  236
#> 153 213    0    1  306
#> 154 213    0    1  292
#> 155 213    0    1  285
#> 156 213    0    1  257
#> 157 214    0    1  325
#> 159 214    0    1  256
#> 162 214    0    1  251
#> 164 214    0    1  234
#> 166 214    0    1  243
#> 167 214    0    1  306
#> 168 214    0    1  289
#> 169 215    0    1  266
#> 170 215    0    1  302
#> 171 215    0    1  307
#> 175 215    0    1  336
#> 177 215    0    1  262
#> 178 215    0    1  294
#> 182 216    0    1  272
#> 184 216    0    1  273
#> 185 216    0    1  308
#> 186 216    0    1  315
#> 188 216    0    1  338
#> 190 216    0    1  230
#> 191 216    0    1  361
#> 192 216    0    1  278