Create female older people indicators dataframe for demography and situation from survey data collected using the standard RAM-OP questionnaire

Create female older people indicators dataframe for demography and situation from survey data collected using the standard RAM-OP questionnaire

create_op_demo_females(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of female older people indicators on demography and situation

Examples

# Create demography and situation indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_demo_females(testSVY)

#>     psu resp1 resp2 resp3 resp4 age ageGrp1 ageGrp2 ageGrp3 ageGrp4 ageGrp5
#> 1   201     1     0     0     0  67       0       1       0       0       0
#> 4   201     1     0     0     0  60       0       1       0       0       0
#> 5   201     1     0     0     0  85       0       0       0       1       0
#> 8   201     1     0     0     0  60       0       1       0       0       0
#> 10  201     1     0     0     0  72       0       0       1       0       0
#> 11  201     1     0     0     0  80       0       0       0       1       0
#> 12  201     0     1     0     0  87       0       0       0       1       0
#> 15  202     1     0     0     0  71       0       0       1       0       0
#> 16  202     1     0     0     0  61       0       1       0       0       0
#> 17  202     1     0     0     0  65       0       1       0       0       0
#> 18  202     1     0     0     0  65       0       1       0       0       0
#> 22  202     0     1     0     0  61       0       1       0       0       0
#> 23  202     1     0     0     0  82       0       0       0       1       0
#> 24  202     1     0     0     0  82       0       0       0       1       0
#> 25  203     1     0     0     0  74       0       0       1       0       0
#> 26  203     1     0     0     0  65       0       1       0       0       0
#> 28  203     1     0     0     0  84       0       0       0       1       0
#> 29  203     1     0     0     0  67       0       1       0       0       0
#> 30  203     1     0     0     0  73       0       0       1       0       0
#> 31  203     0     1     0     0  80       0       0       0       1       0
#> 32  203     1     0     0     0  89       0       0       0       1       0
#> 37  204     1     0     0     0  69       0       1       0       0       0
#> 38  204     0     0     0     1  75       0       0       1       0       0
#> 39  204     0     1     0     0  90       0       0       0       0       1
#> 41  204     1     0     0     0  61       0       1       0       0       0
#> 44  204     1     0     0     0  91       0       0       0       0       1
#> 45  204     1     0     0     0  90       0       0       0       0       1
#> 46  204     1     0     0     0  64       0       1       0       0       0
#> 47  204     0     1     0     0  80       0       0       0       1       0
#> 48  204     1     0     0     0  80       0       0       0       1       0
#> 49  205     1     0     0     0  65       0       1       0       0       0
#> 50  205     1     0     0     0  65       0       1       0       0       0
#> 51  205     1     0     0     0  78       0       0       1       0       0
#> 52  205     1     0     0     0  68       0       1       0       0       0
#> 53  205     1     0     0     0  60       0       1       0       0       0
#> 55  205     0     1     0     0  80       0       0       0       1       0
#> 56  205     1     0     0     0  60       0       1       0       0       0
#> 57  205     1     0     0     0  61       0       1       0       0       0
#> 58  205     1     0     0     0  75       0       0       1       0       0
#> 60  205     1     0     0     0  70       0       0       1       0       0
#> 62  206     0     1     0     0  72       0       0       1       0       0
#> 63  206     1     0     0     0  83       0       0       0       1       0
#> 64  206     1     0     0     0  67       0       1       0       0       0
#> 67  206     0     0     1     0  65       0       1       0       0       0
#> 72  206     1     0     0     0  69       0       1       0       0       0
#> 74  207     1     0     0     0  62       0       1       0       0       0
#> 76  207     1     0     0     0  83       0       0       0       1       0
#> 78  207     1     0     0     0  67       0       1       0       0       0
#> 79  207     1     0     0     0  66       0       1       0       0       0
#> 80  207     1     0     0     0  60       0       1       0       0       0
#> 82  207     1     0     0     0  60       0       1       0       0       0
#> 83  207     1     0     0     0 101       0       0       0       0       1
#> 85  208     0     1     0     0  62       0       1       0       0       0
#> 86  208     1     0     0     0  66       0       1       0       0       0
#> 88  208     1     0     0     0  70       0       0       1       0       0
#> 89  208     1     0     0     0  72       0       0       1       0       0
#> 91  208     1     0     0     0  75       0       0       1       0       0
#> 93  208     1     0     0     0  85       0       0       0       1       0
#> 94  208     1     0     0     0  63       0       1       0       0       0
#> 96  208     0     1     0     0  65       0       1       0       0       0
#> 97  209     1     0     0     0  80       0       0       0       1       0
#> 98  209     1     0     0     0  83       0       0       0       1       0
#> 99  209     1     0     0     0  88       0       0       0       1       0
#> 102 209     1     0     0     0  80       0       0       0       1       0
#> 109 210     1     0     0     0  62       0       1       0       0       0
#> 110 210     1     0     0     0  68       0       1       0       0       0
#> 111 210     1     0     0     0  70       0       0       1       0       0
#> 113 210     1     0     0     0  63       0       1       0       0       0
#> 115 210     1     0     0     0  95       0       0       0       0       1
#> 117 210     1     0     0     0  75       0       0       1       0       0
#> 122 211     1     0     0     0  70       0       0       1       0       0
#> 124 211     0     1     0     0  60       0       1       0       0       0
#> 125 211     1     0     0     0  78       0       0       1       0       0
#> 128 211     1     0     0     0  78       0       0       1       0       0
#> 129 211     1     0     0     0  70       0       0       1       0       0
#> 130 211     1     0     0     0  80       0       0       0       1       0
#> 131 211     1     0     0     0  62       0       1       0       0       0
#> 132 211     1     0     0     0  68       0       1       0       0       0
#> 133 212     0     1     0     0  85       0       0       0       1       0
#> 135 212     1     0     0     0  65       0       1       0       0       0
#> 137 212     1     0     0     0  75       0       0       1       0       0
#> 139 212     1     0     0     0  87       0       0       0       1       0
#> 140 212     1     0     0     0  80       0       0       0       1       0
#> 144 212     1     0     0     0  88       0       0       0       1       0
#> 146 213     1     0     0     0  70       0       0       1       0       0
#> 147 213     1     0     0     0  90       0       0       0       0       1
#> 148 213     1     0     0     0  60       0       1       0       0       0
#> 149 213     1     0     0     0  68       0       1       0       0       0
#> 153 213     0     1     0     0  93       0       0       0       0       1
#> 154 213     1     0     0     0  85       0       0       0       1       0
#> 155 213     1     0     0     0  70       0       0       1       0       0
#> 156 213     1     0     0     0  64       0       1       0       0       0
#> 157 214     1     0     0     0  61       0       1       0       0       0
#> 159 214     1     0     0     0  65       0       1       0       0       0
#> 162 214     0     1     0     0  60       0       1       0       0       0
#> 164 214     1     0     0     0  69       0       1       0       0       0
#> 166 214     1     0     0     0  67       0       1       0       0       0
#> 167 214     1     0     0     0  68       0       1       0       0       0
#> 168 214     1     0     0     0  63       0       1       0       0       0
#> 169 215     1     0     0     0  70       0       0       1       0       0
#> 170 215     1     0     0     0  65       0       1       0       0       0
#> 171 215     0     0     1     0  60       0       1       0       0       0
#> 175 215     1     0     0     0  63       0       1       0       0       0
#> 177 215     1     0     0     0  63       0       1       0       0       0
#> 178 215     1     0     0     0  60       0       1       0       0       0
#> 182 216     1     0     0     0  67       0       1       0       0       0
#> 184 216     1     0     0     0  67       0       1       0       0       0
#> 185 216     1     0     0     0  62       0       1       0       0       0
#> 186 216     1     0     0     0  61       0       1       0       0       0
#> 188 216     1     0     0     0  60       0       1       0       0       0
#> 190 216     0     0     1     0  80       0       0       0       1       0
#> 191 216     1     0     0     0  65       0       1       0       0       0
#> 192 216     1     0     0     0  67       0       1       0       0       0
#>     sex1 sex2 marital1 marital2 marital3 marital4 marital5 marital6 alone
#> 1      0    1        0        0        0        0        1        0     0
#> 4      0    1        0        1        0        0        0        0     0
#> 5      0    1        0        0        0        0        1        0     0
#> 8      0    1        0        0        0        0        1        0     0
#> 10     0    1        0        0        0        0        1        0     0
#> 11     0    1        0        0        0        0        1        0     0
#> 12     0    1        0        0        0        0        1        0     0
#> 15     0    1        0        0        0        0        1        0     1
#> 16     0    1        1        0        0        0        0        0     0
#> 17     0    1        1        0        0        0        0        0     0
#> 18     0    1        0        1        0        0        0        0     0
#> 22     0    1        0        1        0        0        0        0     0
#> 23     0    1        0        0        0        0        1        0     0
#> 24     0    1        0        0        0        0        1        0     0
#> 25     0    1        0        0        0        0        1        0     0
#> 26     0    1        0        0        0        0        1        0     0
#> 28     0    1        0        0        0        0        1        0     0
#> 29     0    1        0        0        1        0        0        0     0
#> 30     0    1        0        0        0        0        1        0     0
#> 31     0    1        0        0        0        0        1        0     0
#> 32     0    1        0        0        0        0        1        0     0
#> 37     0    1        0        0        0        0        1        0     0
#> 38     0    1        0        0        0        0        1        0     0
#> 39     0    1        0        0        0        0        1        0     1
#> 41     0    1        1        0        0        0        0        0     0
#> 44     0    1        0        0        0        0        1        0     0
#> 45     0    1        0        0        0        0        1        0     0
#> 46     0    1        0        0        0        0        1        0     0
#> 47     0    1        0        0        0        0        1        0     0
#> 48     0    1        0        0        0        0        1        0     1
#> 49     0    1        0        1        0        0        0        0     0
#> 50     0    1        0        0        0        0        1        0     0
#> 51     0    1        0        0        0        0        1        0     0
#> 52     0    1        0        0        1        0        0        0     0
#> 53     0    1        1        0        0        0        0        0     1
#> 55     0    1        0        0        0        0        1        0     0
#> 56     0    1        0        0        0        0        1        0     0
#> 57     0    1        0        0        0        0        1        0     0
#> 58     0    1        0        0        0        0        1        0     0
#> 60     0    1        0        0        0        1        0        0     1
#> 62     0    1        0        0        1        0        0        0     0
#> 63     0    1        0        0        0        0        1        0     0
#> 64     0    1        0        0        0        1        0        0     0
#> 67     0    1        0        1        0        0        0        0     0
#> 72     0    1        0        0        0        0        1        0     0
#> 74     0    1        1        0        0        0        0        0     1
#> 76     0    1        0        0        0        0        1        0     0
#> 78     0    1        0        0        1        0        0        0     0
#> 79     0    1        0        0        0        0        1        0     0
#> 80     0    1        0        0        0        0        1        0     0
#> 82     0    1        0        0        0        0        1        0     0
#> 83     0    1        0        0        0        0        1        0     0
#> 85     0    1        0        0        0        1        0        0     0
#> 86     0    1        0        1        0        0        0        0     0
#> 88     0    1        0        0        0        0        1        0     0
#> 89     0    1        0        0        0        0        1        0     0
#> 91     0    1        0        0        0        0        1        0     0
#> 93     0    1        0        0        0        0        1        0     1
#> 94     0    1        0        1        0        0        0        0     0
#> 96     0    1        0        0        0        0        1        0     0
#> 97     0    1        0        0        0        0        1        0     0
#> 98     0    1        0        0        0        0        1        0     0
#> 99     0    1        0        0        1        0        0        0     0
#> 102    0    1        0        0        0        0        1        0     0
#> 109    0    1        0        0        0        1        0        0     0
#> 110    0    1        0        0        1        0        0        0     0
#> 111    0    1        0        0        0        0        1        0     0
#> 113    0    1        0        1        0        0        0        0     0
#> 115    0    1        0        0        0        0        1        0     0
#> 117    0    1        0        0        0        0        1        0     0
#> 122    0    1        0        0        0        0        1        0     0
#> 124    0    1        0        0        0        1        0        0     0
#> 125    0    1        0        0        0        0        1        0     1
#> 128    0    1        0        0        0        0        1        0     1
#> 129    0    1        0        0        0        0        1        0     0
#> 130    0    1        0        0        0        0        1        0     0
#> 131    0    1        0        0        1        0        0        0     0
#> 132    0    1        0        0        0        0        1        0     0
#> 133    0    1        0        0        0        0        1        0     1
#> 135    0    1        0        0        0        0        1        0     0
#> 137    0    1        0        0        0        0        1        0     0
#> 139    0    1        0        1        0        0        0        0     0
#> 140    0    1        0        0        0        0        1        0     1
#> 144    0    1        0        0        0        0        1        0     0
#> 146    0    1        0        1        0        0        0        0     0
#> 147    0    1        0        0        0        0        1        0     0
#> 148    0    1        0        0        0        0        1        0     0
#> 149    0    1        0        0        0        0        1        0     0
#> 153    0    1        0        0        0        0        1        0     1
#> 154    0    1        0        0        0        0        1        0     0
#> 155    0    1        0        0        0        0        1        0     0
#> 156    0    1        0        0        0        0        1        0     0
#> 157    0    1        0        1        0        0        0        0     0
#> 159    0    1        0        0        0        0        1        0     0
#> 162    0    1        0        0        0        0        1        0     0
#> 164    0    1        0        0        0        0        1        0     0
#> 166    0    1        0        0        0        0        1        0     0
#> 167    0    1        0        0        0        0        1        0     0
#> 168    0    1        0        0        0        0        1        0     0
#> 169    0    1        0        0        0        0        1        0     0
#> 170    0    1        0        0        1        0        0        0     0
#> 171    0    1        0        1        0        0        0        0     0
#> 175    0    1        0        0        0        0        1        0     0
#> 177    0    1        0        1        0        0        0        0     0
#> 178    0    1        0        1        0        0        0        0     1
#> 182    0    1        0        1        0        0        0        0     0
#> 184    0    1        1        0        0        0        0        0     0
#> 185    0    1        0        0        0        0        1        0     0
#> 186    0    1        0        0        0        0        1        0     0
#> 188    0    1        0        0        0        0        1        0     0
#> 190    0    1        0        0        0        0        1        0     0
#> 191    0    1        0        1        0        0        0        0     0
#> 192    0    1        0        0        0        0        1        0     1