create_op_disability_males

Create male older people indicators dataframe for disability from survey data collected using the standard RAM-OP questionnaire

create_op_disability_males(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of male older people indicators on disability

Examples

# Create disability indicators dataset from RAM-OP survey data collected
# from Addis Ababa, Ethiopia
create_op_disability_males(testSVY)

#>     psu sex1 sex2 wgVisionD0 wgVisionD1 wgVisionD2 wgVisionD3 wgHearingD0
#> 2   201    1    0          1          0          0          0           1
#> 3   201    1    0          1          0          0          0           1
#> 6   201    1    0          1          0          0          0           1
#> 7   201    1    0          1          0          0          0           1
#> 9   201    1    0          1          0          0          0           1
#> 13  202    1    0          1          0          0          0           1
#> 14  202    1    0          1          0          0          0           1
#> 19  202    1    0          1          0          0          0           1
#> 20  202    1    0          1          0          0          0           1
#> 21  202    1    0          1          0          0          0           1
#> 27  203    1    0          1          0          0          0           1
#> 33  203    1    0          1          0          0          0           1
#> 34  203    1    0          1          0          0          0           1
#> 35  203    1    0          1          0          0          0           1
#> 36  203    1    0          1          0          0          0           1
#> 40  204    1    0          1          0          0          0           1
#> 42  204    1    0          1          0          0          0           1
#> 43  204    1    0          1          0          0          0           1
#> 54  205    1    0          1          0          0          0           1
#> 59  205    1    0          1          0          0          0           1
#> 61  206    1    0          1          0          0          0           1
#> 65  206    1    0          1          0          0          0           1
#> 66  206    1    0          1          0          0          0           1
#> 68  206    1    0          1          0          0          0           1
#> 69  206    1    0          1          0          0          0           1
#> 70  206    1    0          1          0          0          0           1
#> 71  206    1    0          1          0          0          0           1
#> 73  207    1    0          1          0          0          0           1
#> 75  207    1    0          1          0          0          0           1
#> 77  207    1    0          1          0          0          0           1
#> 81  207    1    0          1          0          0          0           1
#> 84  207    1    0          1          0          0          0           1
#> 87  208    1    0          1          0          0          0           1
#> 90  208    1    0          1          0          0          0           1
#> 92  208    1    0          1          0          0          0           1
#> 95  208    1    0          1          0          0          0           1
#> 100 209    1    0          1          0          0          0           1
#> 101 209    1    0          1          0          0          0           1
#> 103 209    1    0          1          0          0          0           1
#> 104 209    1    0          1          0          0          0           1
#> 105 209    1    0          1          0          0          0           1
#> 106 209    1    0          1          0          0          0           1
#> 107 209    1    0          1          0          0          0           1
#> 108 209    1    0          1          0          0          0           1
#> 112 210    1    0          1          0          0          0           1
#> 114 210    1    0          1          0          0          0           1
#> 116 210    1    0          1          0          0          0           1
#> 118 210    1    0          1          0          0          0           1
#> 119 210    1    0          1          0          0          0           1
#> 120 210    1    0          1          0          0          0           1
#> 121 211    1    0          1          0          0          0           1
#> 123 211    1    0          1          0          0          0           1
#> 126 211    1    0          1          0          0          0           1
#> 127 211    1    0          1          0          0          0           1
#> 134 212    1    0          1          0          0          0           1
#> 136 212    1    0          1          0          0          0           1
#> 138 212    1    0          1          0          0          0           1
#> 141 212    1    0          1          0          0          0           1
#> 142 212    1    0          1          0          0          0           1
#> 143 212    1    0          1          0          0          0           1
#> 145 213    1    0          1          0          0          0           1
#> 150 213    1    0          1          0          0          0           1
#> 151 213    1    0          1          0          0          0           1
#> 152 213    1    0          1          0          0          0           1
#> 158 214    1    0          1          0          0          0           1
#> 160 214    1    0          1          0          0          0           1
#> 161 214    1    0          1          0          0          0           1
#> 163 214    1    0          1          0          0          0           1
#> 165 214    1    0          1          0          0          0           1
#> 172 215    1    0          1          0          0          0           1
#> 173 215    1    0          1          0          0          0           1
#> 174 215    1    0          1          0          0          0           1
#> 176 215    1    0          1          0          0          0           1
#> 179 215    1    0          1          0          0          0           1
#> 180 215    1    0          1          0          0          0           1
#> 181 216    1    0          1          0          0          0           1
#> 183 216    1    0          1          0          0          0           1
#> 187 216    1    0          1          0          0          0           1
#> 189 216    1    0          1          0          0          0           1
#>     wgHearingD1 wgHearingD2 wgHearingD3 wgMobilityD0 wgMobilityD1 wgMobilityD2
#> 2             0           0           0            1            0            0
#> 3             0           0           0            1            0            0
#> 6             0           0           0            1            0            0
#> 7             0           0           0            1            0            0
#> 9             0           0           0            1            0            0
#> 13            0           0           0            1            0            0
#> 14            0           0           0            1            0            0
#> 19            0           0           0            1            0            0
#> 20            0           0           0            1            0            0
#> 21            0           0           0            1            0            0
#> 27            0           0           0            1            0            0
#> 33            0           0           0            1            0            0
#> 34            0           0           0            1            0            0
#> 35            0           0           0            1            0            0
#> 36            0           0           0            1            0            0
#> 40            0           0           0            1            0            0
#> 42            0           0           0            1            0            0
#> 43            0           0           0            1            0            0
#> 54            0           0           0            1            0            0
#> 59            0           0           0            1            0            0
#> 61            0           0           0            1            0            0
#> 65            0           0           0            1            0            0
#> 66            0           0           0            1            0            0
#> 68            0           0           0            1            0            0
#> 69            0           0           0            1            0            0
#> 70            0           0           0            1            0            0
#> 71            0           0           0            1            0            0
#> 73            0           0           0            1            0            0
#> 75            0           0           0            1            0            0
#> 77            0           0           0            1            0            0
#> 81            0           0           0            1            0            0
#> 84            0           0           0            1            0            0
#> 87            0           0           0            1            0            0
#> 90            0           0           0            1            0            0
#> 92            0           0           0            1            0            0
#> 95            0           0           0            1            0            0
#> 100           0           0           0            1            0            0
#> 101           0           0           0            1            0            0
#> 103           0           0           0            1            0            0
#> 104           0           0           0            1            0            0
#> 105           0           0           0            1            0            0
#> 106           0           0           0            1            0            0
#> 107           0           0           0            1            0            0
#> 108           0           0           0            1            0            0
#> 112           0           0           0            1            0            0
#> 114           0           0           0            1            0            0
#> 116           0           0           0            1            0            0
#> 118           0           0           0            1            0            0
#> 119           0           0           0            1            0            0
#> 120           0           0           0            1            0            0
#> 121           0           0           0            1            0            0
#> 123           0           0           0            1            0            0
#> 126           0           0           0            1            0            0
#> 127           0           0           0            1            0            0
#> 134           0           0           0            1            0            0
#> 136           0           0           0            1            0            0
#> 138           0           0           0            1            0            0
#> 141           0           0           0            1            0            0
#> 142           0           0           0            1            0            0
#> 143           0           0           0            1            0            0
#> 145           0           0           0            1            0            0
#> 150           0           0           0            1            0            0
#> 151           0           0           0            1            0            0
#> 152           0           0           0            1            0            0
#> 158           0           0           0            1            0            0
#> 160           0           0           0            1            0            0
#> 161           0           0           0            1            0            0
#> 163           0           0           0            1            0            0
#> 165           0           0           0            1            0            0
#> 172           0           0           0            1            0            0
#> 173           0           0           0            1            0            0
#> 174           0           0           0            1            0            0
#> 176           0           0           0            1            0            0
#> 179           0           0           0            1            0            0
#> 180           0           0           0            1            0            0
#> 181           0           0           0            1            0            0
#> 183           0           0           0            1            0            0
#> 187           0           0           0            1            0            0
#> 189           0           0           0            1            0            0
#>     wgMobilityD3 wgRememberingD0 wgRememberingD1 wgRememberingD2
#> 2              0               1               0               0
#> 3              0               1               0               0
#> 6              0               1               0               0
#> 7              0               1               0               0
#> 9              0               1               0               0
#> 13             0               1               0               0
#> 14             0               1               0               0
#> 19             0               1               0               0
#> 20             0               1               0               0
#> 21             0               1               0               0
#> 27             0               1               0               0
#> 33             0               1               0               0
#> 34             0               1               0               0
#> 35             0               1               0               0
#> 36             0               1               0               0
#> 40             0               1               0               0
#> 42             0               1               0               0
#> 43             0               1               0               0
#> 54             0               1               0               0
#> 59             0               1               0               0
#> 61             0               1               0               0
#> 65             0               1               0               0
#> 66             0               1               0               0
#> 68             0               1               0               0
#> 69             0               1               0               0
#> 70             0               1               0               0
#> 71             0               1               0               0
#> 73             0               1               0               0
#> 75             0               1               0               0
#> 77             0               1               0               0
#> 81             0               1               0               0
#> 84             0               1               0               0
#> 87             0               1               0               0
#> 90             0               1               0               0
#> 92             0               1               0               0
#> 95             0               1               0               0
#> 100            0               1               0               0
#> 101            0               1               0               0
#> 103            0               1               0               0
#> 104            0               1               0               0
#> 105            0               1               0               0
#> 106            0               1               0               0
#> 107            0               1               0               0
#> 108            0               1               0               0
#> 112            0               1               0               0
#> 114            0               1               0               0
#> 116            0               1               0               0
#> 118            0               1               0               0
#> 119            0               1               0               0
#> 120            0               1               0               0
#> 121            0               1               0               0
#> 123            0               1               0               0
#> 126            0               1               0               0
#> 127            0               1               0               0
#> 134            0               1               0               0
#> 136            0               1               0               0
#> 138            0               1               0               0
#> 141            0               1               0               0
#> 142            0               1               0               0
#> 143            0               1               0               0
#> 145            0               1               0               0
#> 150            0               1               0               0
#> 151            0               1               0               0
#> 152            0               1               0               0
#> 158            0               1               0               0
#> 160            0               1               0               0
#> 161            0               1               0               0
#> 163            0               1               0               0
#> 165            0               1               0               0
#> 172            0               1               0               0
#> 173            0               1               0               0
#> 174            0               1               0               0
#> 176            0               1               0               0
#> 179            0               1               0               0
#> 180            0               1               0               0
#> 181            0               1               0               0
#> 183            0               1               0               0
#> 187            0               1               0               0
#> 189            0               1               0               0
#>     wgRememberingD3 wgSelfCareD0 wgSelfCareD1 wgSelfCareD2 wgSelfCareD3
#> 2                 0            1            0            0            0
#> 3                 0            1            0            0            0
#> 6                 0            1            0            0            0
#> 7                 0            1            0            0            0
#> 9                 0            1            0            0            0
#> 13                0            1            0            0            0
#> 14                0            1            0            0            0
#> 19                0            1            0            0            0
#> 20                0            1            0            0            0
#> 21                0            1            0            0            0
#> 27                0            1            0            0            0
#> 33                0            1            0            0            0
#> 34                0            1            0            0            0
#> 35                0            1            0            0            0
#> 36                0            1            0            0            0
#> 40                0            1            0            0            0
#> 42                0            1            0            0            0
#> 43                0            1            0            0            0
#> 54                0            1            0            0            0
#> 59                0            1            0            0            0
#> 61                0            1            0            0            0
#> 65                0            1            0            0            0
#> 66                0            1            0            0            0
#> 68                0            1            0            0            0
#> 69                0            1            0            0            0
#> 70                0            1            0            0            0
#> 71                0            1            0            0            0
#> 73                0            1            0            0            0
#> 75                0            1            0            0            0
#> 77                0            1            0            0            0
#> 81                0            1            0            0            0
#> 84                0            1            0            0            0
#> 87                0            1            0            0            0
#> 90                0            1            0            0            0
#> 92                0            1            0            0            0
#> 95                0            1            0            0            0
#> 100               0            1            0            0            0
#> 101               0            1            0            0            0
#> 103               0            1            0            0            0
#> 104               0            1            0            0            0
#> 105               0            1            0            0            0
#> 106               0            1            0            0            0
#> 107               0            1            0            0            0
#> 108               0            1            0            0            0
#> 112               0            1            0            0            0
#> 114               0            1            0            0            0
#> 116               0            1            0            0            0
#> 118               0            1            0            0            0
#> 119               0            1            0            0            0
#> 120               0            1            0            0            0
#> 121               0            1            0            0            0
#> 123               0            1            0            0            0
#> 126               0            1            0            0            0
#> 127               0            1            0            0            0
#> 134               0            1            0            0            0
#> 136               0            1            0            0            0
#> 138               0            1            0            0            0
#> 141               0            1            0            0            0
#> 142               0            1            0            0            0
#> 143               0            1            0            0            0
#> 145               0            1            0            0            0
#> 150               0            1            0            0            0
#> 151               0            1            0            0            0
#> 152               0            1            0            0            0
#> 158               0            1            0            0            0
#> 160               0            1            0            0            0
#> 161               0            1            0            0            0
#> 163               0            1            0            0            0
#> 165               0            1            0            0            0
#> 172               0            1            0            0            0
#> 173               0            1            0            0            0
#> 174               0            1            0            0            0
#> 176               0            1            0            0            0
#> 179               0            1            0            0            0
#> 180               0            1            0            0            0
#> 181               0            1            0            0            0
#> 183               0            1            0            0            0
#> 187               0            1            0            0            0
#> 189               0            1            0            0            0
#>     wgCommunicatingD0 wgCommunicatingD1 wgCommunicatingD2 wgCommunicatingD3
#> 2                   1                 0                 0                 0
#> 3                   1                 0                 0                 0
#> 6                   1                 0                 0                 0
#> 7                   1                 0                 0                 0
#> 9                   1                 0                 0                 0
#> 13                  1                 0                 0                 0
#> 14                  1                 0                 0                 0
#> 19                  1                 0                 0                 0
#> 20                  1                 0                 0                 0
#> 21                  1                 0                 0                 0
#> 27                  1                 0                 0                 0
#> 33                  1                 0                 0                 0
#> 34                  1                 0                 0                 0
#> 35                  1                 0                 0                 0
#> 36                  1                 0                 0                 0
#> 40                  1                 0                 0                 0
#> 42                  1                 0                 0                 0
#> 43                  1                 0                 0                 0
#> 54                  1                 0                 0                 0
#> 59                  1                 0                 0                 0
#> 61                  1                 0                 0                 0
#> 65                  1                 0                 0                 0
#> 66                  1                 0                 0                 0
#> 68                  1                 0                 0                 0
#> 69                  1                 0                 0                 0
#> 70                  1                 0                 0                 0
#> 71                  1                 0                 0                 0
#> 73                  1                 0                 0                 0
#> 75                  1                 0                 0                 0
#> 77                  1                 0                 0                 0
#> 81                  1                 0                 0                 0
#> 84                  1                 0                 0                 0
#> 87                  1                 0                 0                 0
#> 90                  1                 0                 0                 0
#> 92                  1                 0                 0                 0
#> 95                  1                 0                 0                 0
#> 100                 1                 0                 0                 0
#> 101                 1                 0                 0                 0
#> 103                 1                 0                 0                 0
#> 104                 1                 0                 0                 0
#> 105                 1                 0                 0                 0
#> 106                 1                 0                 0                 0
#> 107                 1                 0                 0                 0
#> 108                 1                 0                 0                 0
#> 112                 1                 0                 0                 0
#> 114                 1                 0                 0                 0
#> 116                 1                 0                 0                 0
#> 118                 1                 0                 0                 0
#> 119                 1                 0                 0                 0
#> 120                 1                 0                 0                 0
#> 121                 1                 0                 0                 0
#> 123                 1                 0                 0                 0
#> 126                 1                 0                 0                 0
#> 127                 1                 0                 0                 0
#> 134                 1                 0                 0                 0
#> 136                 1                 0                 0                 0
#> 138                 1                 0                 0                 0
#> 141                 1                 0                 0                 0
#> 142                 1                 0                 0                 0
#> 143                 1                 0                 0                 0
#> 145                 1                 0                 0                 0
#> 150                 1                 0                 0                 0
#> 151                 1                 0                 0                 0
#> 152                 1                 0                 0                 0
#> 158                 1                 0                 0                 0
#> 160                 1                 0                 0                 0
#> 161                 1                 0                 0                 0
#> 163                 1                 0                 0                 0
#> 165                 1                 0                 0                 0
#> 172                 1                 0                 0                 0
#> 173                 1                 0                 0                 0
#> 174                 1                 0                 0                 0
#> 176                 1                 0                 0                 0
#> 179                 1                 0                 0                 0
#> 180                 1                 0                 0                 0
#> 181                 1                 0                 0                 0
#> 183                 1                 0                 0                 0
#> 187                 1                 0                 0                 0
#> 189                 1                 0                 0                 0
#>     wgP0 wgP1 wgP2 wgP3 wgPM
#> 2      1    0    0    0    0
#> 3      1    0    0    0    0
#> 6      1    0    0    0    0
#> 7      1    0    0    0    0
#> 9      1    0    0    0    0
#> 13     1    0    0    0    0
#> 14     1    0    0    0    0
#> 19     1    0    0    0    0
#> 20     1    0    0    0    0
#> 21     1    0    0    0    0
#> 27     1    0    0    0    0
#> 33     1    0    0    0    0
#> 34     1    0    0    0    0
#> 35     1    0    0    0    0
#> 36     1    0    0    0    0
#> 40     1    0    0    0    0
#> 42     1    0    0    0    0
#> 43     1    0    0    0    0
#> 54     1    0    0    0    0
#> 59     1    0    0    0    0
#> 61     1    0    0    0    0
#> 65     1    0    0    0    0
#> 66     1    0    0    0    0
#> 68     1    0    0    0    0
#> 69     1    0    0    0    0
#> 70     1    0    0    0    0
#> 71     1    0    0    0    0
#> 73     1    0    0    0    0
#> 75     1    0    0    0    0
#> 77     1    0    0    0    0
#> 81     1    0    0    0    0
#> 84     1    0    0    0    0
#> 87     1    0    0    0    0
#> 90     1    0    0    0    0
#> 92     1    0    0    0    0
#> 95     1    0    0    0    0
#> 100    1    0    0    0    0
#> 101    1    0    0    0    0
#> 103    1    0    0    0    0
#> 104    1    0    0    0    0
#> 105    1    0    0    0    0
#> 106    1    0    0    0    0
#> 107    1    0    0    0    0
#> 108    1    0    0    0    0
#> 112    1    0    0    0    0
#> 114    1    0    0    0    0
#> 116    1    0    0    0    0
#> 118    1    0    0    0    0
#> 119    1    0    0    0    0
#> 120    1    0    0    0    0
#> 121    1    0    0    0    0
#> 123    1    0    0    0    0
#> 126    1    0    0    0    0
#> 127    1    0    0    0    0
#> 134    1    0    0    0    0
#> 136    1    0    0    0    0
#> 138    1    0    0    0    0
#> 141    1    0    0    0    0
#> 142    1    0    0    0    0
#> 143    1    0    0    0    0
#> 145    1    0    0    0    0
#> 150    1    0    0    0    0
#> 151    1    0    0    0    0
#> 152    1    0    0    0    0
#> 158    1    0    0    0    0
#> 160    1    0    0    0    0
#> 161    1    0    0    0    0
#> 163    1    0    0    0    0
#> 165    1    0    0    0    0
#> 172    1    0    0    0    0
#> 173    1    0    0    0    0
#> 174    1    0    0    0    0
#> 176    1    0    0    0    0
#> 179    1    0    0    0    0
#> 180    1    0    0    0    0
#> 181    1    0    0    0    0
#> 183    1    0    0    0    0
#> 187    1    0    0    0    0
#> 189    1    0    0    0    0