Create older people indicators for food intake from survey data collected using the standard RAM-OP questionnaire

Create older people indicators for food intake from survey data collected using the standard RAM-OP questionnaire

Usage

create_op_food(svy)

Arguments

svy

A data.frame collected using the standard RAM-OP questionnaire

Value

A dataframe of older people indicators on food intake

Dietary intake indicators

These dietary intake indicators have been purpose-built for older people but the basic approach used is described in:

Kennedy G, Ballard T, Dop M C (2011). Guidelines for Measuring Household and Individual Dietary Diversity. Rome, FAO https://www.fao.org/3/i1983e/i1983e00.htm

and extended to include indicators of probable adequate intake of a number of nutrients / micronutrients.

MF

Meal frequency

DDS

Dietary Diversity Score (count of 11 groups)

FG01

Cereals

FG02

Roots and tubers

FG03

Fruits and vegetables

FG04

All meat

FG05

Eggs

FG06

Fish

FG07

Legumes, nuts and seeds

FG08

Milk and milk products

FG09

Fats

FG10

Sugar

FG11

Other

proteinRich

Protein rich foods

pProtein

Protein rich plant sources of protein

aProtein

Protein rich animal sources of protein

pVitA

Plant sources of vitamin A

aVitA

Animal sources of vitamin A

xVitA

Any source of vitamin A

ironRich

Iron rich foods

caRich

Calcium rich foods

znRich

Zinc rich foods

vitB1

Vitamin B1-rich foods

vitB2

Vitamin B2-rich foods

vitB3

Vitamin B3-rich foods

vitB6

Vitamin B6-rich foods

vitB12

Vitamin B12-rich foods

vitBcomplex

Vitamin B1/B2/B3/B6/B12-rich foods

Author

Mark Myatt

Examples

# Create food intake indicators dataset from RAM-OP survey data collected
# from Addis Ababa, Ethiopia
create_op_food(testSVY)
#> # A tibble: 192 × 31
#>      psu  sex1  sex2    MF   DDS  FG01  FG02  FG03  FG04  FG05  FG06  FG07  FG08
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1   201     0     1     3     6     1     1     1     0     0     1     1     0
#>  2   201     1     0     3     7     1     1     1     0     0     0     1     0
#>  3   201     1     0     2     2     0     0     0     0     0     1     0     0
#>  4   201     0     1     3     3     1     1     0     0     0     0     1     0
#>  5   201     0     1     3     5     1     0     1     0     0     1     0     0
#>  6   201     1     0     4     5     1     1     0     0     0     1     1     0
#>  7   201     1     0     3     6     1     1     1     0     0     1     1     0
#>  8   201     0     1     3     4     1     1     0     0     0     0     1     0
#>  9   201     1     0     2     4     1     0     1     0     0     0     0     0
#> 10   201     0     1     2     6     1     1     1     1     0     0     0     0
#> # ℹ 182 more rows
#> # ℹ 18 more variables: FG09 <dbl>, FG10 <dbl>, FG11 <dbl>, proteinRich <dbl>,
#> #   pProtein <dbl>, aProtein <dbl>, pVitA <dbl>, aVitA <dbl>, xVitA <dbl>,
#> #   ironRich <dbl>, caRich <dbl>, znRich <dbl>, vitB1 <dbl>, vitB2 <dbl>,
#> #   vitB3 <dbl>, vitB6 <dbl>, vitB12 <dbl>, vitBcomplex <dbl>