Create female older people indicators dataframe for food intake from survey data collected using the standard RAM-OP questionnaire

Create female older people indicators dataframe for food intake from survey data collected using the standard RAM-OP questionnaire

create_op_food_females(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of female older people indicators on food intake

Examples

# Create food intake indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_food_females(testSVY)

#>     psu sex1 sex2 MF DDS FG01 FG02 FG03 FG04 FG05 FG06 FG07 FG08 FG09 FG10 FG11
#> 1   201    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 4   201    0    1  3   3    1    1    0    0    0    0    1    0    0    0    0
#> 5   201    0    1  3   5    1    0    1    0    0    1    0    0    0    1    1
#> 8   201    0    1  3   4    1    1    0    0    0    0    1    0    0    0    1
#> 10  201    0    1  2   6    1    1    1    1    0    0    0    0    0    1    1
#> 11  201    0    1  2   4    1    0    1    0    0    0    0    0    1    0    1
#> 12  201    0    1  4   3    1    0    0    0    0    0    0    0    0    1    1
#> 15  202    0    1  2   5    1    1    0    0    0    1    0    0    0    1    1
#> 16  202    0    1  1   5    1    1    1    0    0    1    0    0    0    0    1
#> 17  202    0    1  3   6    1    1    1    0    0    0    1    0    0    1    1
#> 18  202    0    1  1   3    1    0    0    0    0    0    0    0    0    1    1
#> 22  202    0    1  2   4    1    0    1    0    0    0    0    0    0    1    1
#> 23  202    0    1  3   7    1    1    1    0    0    1    0    1    1    0    1
#> 24  202    0    1  3   6    1    1    1    0    0    1    0    0    0    1    1
#> 25  203    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 26  203    0    1  5   4    1    1    0    0    0    0    1    0    0    0    1
#> 28  203    0    1  3   4    1    0    0    0    0    0    1    0    0    1    1
#> 29  203    0    1  4   5    1    0    0    0    0    0    1    0    1    1    1
#> 30  203    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 31  203    0    1  3   5    1    0    0    0    0    0    1    0    1    1    1
#> 32  203    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 37  204    0    1  2   7    1    1    1    0    0    1    1    0    0    1    1
#> 38  204    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 39  204    0    1  2   2    0    1    0    0    0    0    0    0    0    0    1
#> 41  204    0    1  2   5    1    1    0    0    0    1    0    0    0    1    1
#> 44  204    0    1  3   6    1    0    1    0    0    1    1    0    0    1    1
#> 45  204    0    1  3   3    1    0    0    0    0    1    0    0    0    0    1
#> 46  204    0    1  2   7    1    1    1    1    0    0    1    0    0    1    1
#> 47  204    0    1  2   6    1    1    1    0    0    0    1    0    0    1    1
#> 48  204    0    1  3   6    1    1    1    0    0    0    1    0    1    0    1
#> 49  205    0    1  4   7    1    1    1    0    0    0    1    0    1    1    1
#> 50  205    0    1  2   4    1    1    1    0    0    0    0    0    0    0    1
#> 51  205    0    1  2   4    1    1    0    0    0    0    0    0    1    0    1
#> 52  205    0    1  2   5    1    0    1    0    0    0    1    1    0    0    1
#> 53  205    0    1  2   3    1    0    0    0    0    0    0    0    0    1    1
#> 55  205    0    1  4   6    1    1    1    1    0    1    0    0    0    0    1
#> 56  205    0    1  3   4    1    0    1    0    0    0    0    0    0    1    1
#> 57  205    0    1  2   6    1    1    1    0    0    0    1    0    0    1    1
#> 58  205    0    1  3   5    1    0    0    0    0    1    1    0    0    1    1
#> 60  205    0    1  2   6    1    1    1    0    0    0    1    0    0    1    1
#> 62  206    0    1  4   3    0    1    1    0    0    0    0    0    0    0    1
#> 63  206    0    1  2   5    1    1    1    0    0    0    1    0    0    0    1
#> 64  206    0    1  2   4    1    0    1    0    0    0    1    0    0    0    1
#> 67  206    0    1  4   5    1    0    0    0    0    1    0    0    1    1    1
#> 72  206    0    1  3   6    1    1    1    0    0    0    1    0    0    1    1
#> 74  207    0    1  1   2    0    0    1    0    0    0    0    0    0    0    1
#> 76  207    0    1  3   4    1    0    1    0    0    1    0    0    0    0    1
#> 78  207    0    1  3   4    1    1    0    0    0    0    0    0    0    1    1
#> 79  207    0    1  2   4    1    0    0    0    0    1    0    0    0    1    1
#> 80  207    0    1  2   5    1    0    1    0    0    0    1    0    0    1    1
#> 82  207    0    1  1   3    1    0    0    0    0    1    0    0    0    0    1
#> 83  207    0    1  2   3    0    1    1    0    0    0    0    0    0    0    1
#> 85  208    0    1  3   5    1    0    0    0    0    1    0    0    1    1    1
#> 86  208    0    1  4   5    1    1    1    0    0    0    0    0    0    1    1
#> 88  208    0    1  2   4    1    0    0    0    0    1    0    0    0    1    1
#> 89  208    0    1  3   5    1    0    0    0    0    0    1    0    1    1    1
#> 91  208    0    1  2   4    1    1    0    0    0    0    1    0    0    0    1
#> 93  208    0    1  2   5    1    0    0    0    0    1    1    0    0    1    1
#> 94  208    0    1  3   5    1    1    1    0    0    0    0    0    0    1    1
#> 96  208    0    1  4   5    1    0    1    0    0    0    1    0    0    1    1
#> 97  209    0    1  2   6    1    1    0    0    0    1    0    0    1    1    1
#> 98  209    0    1  3   5    1    1    1    0    0    0    0    1    0    0    1
#> 99  209    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 102 209    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 109 210    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 110 210    0    1  3   4    1    0    1    0    0    0    1    0    0    0    1
#> 111 210    0    1  4   8    1    1    1    1    0    1    0    0    1    1    1
#> 113 210    0    1  1   3    1    1    0    0    0    0    1    0    0    0    0
#> 115 210    0    1  2   4    1    0    0    0    0    1    0    0    0    1    1
#> 117 210    0    1  2   4    0    1    1    0    0    0    0    0    0    1    1
#> 122 211    0    1  3   7    1    1    1    0    0    1    1    0    0    1    1
#> 124 211    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 125 211    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 128 211    0    1  2   4    1    0    0    0    1    0    1    0    0    0    1
#> 129 211    0    1  2   5    1    1    0    0    0    0    0    0    1    1    1
#> 130 211    0    1  3   6    1    0    1    0    0    0    1    0    1    1    1
#> 131 211    0    1  3   5    0    1    1    1    0    0    0    0    0    1    1
#> 132 211    0    1  2   3    0    1    1    0    0    0    0    0    0    0    1
#> 133 212    0    1  3   2    1    0    0    0    0    0    0    0    0    0    1
#> 135 212    0    1  3   5    1    1    0    0    0    0    0    0    1    1    1
#> 137 212    0    1  2   4    1    0    1    0    0    0    0    0    0    1    1
#> 139 212    0    1  4   5    1    1    1    0    0    0    0    0    0    1    1
#> 140 212    0    1  1   2    1    0    1    0    0    0    0    0    0    0    0
#> 144 212    0    1  2   4    0    1    1    0    0    0    1    0    1    0    0
#> 146 213    0    1  4   4    1    0    0    0    0    1    0    0    0    1    1
#> 147 213    0    1  3   4    1    0    1    0    0    1    0    0    0    0    1
#> 148 213    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 149 213    0    1  4   4    1    1    0    0    0    0    0    0    0    1    1
#> 153 213    0    1  1   2    1    0    0    0    0    0    0    0    0    0    1
#> 154 213    0    1  4   9    1    1    1    1    0    1    1    0    1    1    1
#> 155 213    0    1  2   3    1    0    0    0    0    1    0    0    0    0    1
#> 156 213    0    1  4   6    1    1    1    0    0    1    1    0    0    0    1
#> 157 214    0    1  4   7    1    1    1    1    0    1    1    0    0    0    1
#> 159 214    0    1  2   3    0    1    0    0    0    0    0    0    0    1    1
#> 162 214    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 164 214    0    1  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 166 214    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 167 214    0    1  3   6    1    0    1    0    1    0    1    0    1    0    1
#> 168 214    0    1  3   5    1    1    1    0    0    0    1    0    0    0    1
#> 169 215    0    1  1   2    0    0    0    0    0    0    0    0    0    1    1
#> 170 215    0    1  4   3    1    0    0    0    0    0    0    0    0    1    1
#> 171 215    0    1  2   4    1    1    0    0    0    0    1    0    0    0    1
#> 175 215    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 177 215    0    1  3   5    1    1    1    0    0    0    0    0    0    1    1
#> 178 215    0    1  3   3    1    0    1    0    0    0    0    0    0    0    1
#> 182 216    0    1  4   7    1    1    1    0    0    1    1    0    1    0    1
#> 184 216    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 185 216    0    1  3   4    1    1    1    0    0    0    0    0    0    0    1
#> 186 216    0    1  3   4    1    0    0    0    0    0    0    0    1    1    1
#> 188 216    0    1  5   5    1    0    1    0    0    0    1    0    0    1    1
#> 190 216    0    1  3   6    0    1    1    0    0    0    0    1    1    1    1
#> 191 216    0    1  2   7    1    1    0    1    0    1    1    0    1    0    1
#> 192 216    0    1  2   3    1    0    1    0    0    0    0    0    0    0    1
#>     proteinRich pProtein aProtein pVitA aVitA xVitA ironRich caRich znRich
#> 1             1        1        0     1     0     1        1      0      1
#> 4             1        1        0     0     0     0        0      0      1
#> 5             0        0        0     1     0     1        1      0      1
#> 8             1        1        0     0     0     0        0      0      1
#> 10            1        0        1     1     0     1        1      0      1
#> 11            0        0        0     1     0     1        1      0      0
#> 12            0        0        0     0     0     0        0      0      0
#> 15            0        0        0     0     0     0        1      0      1
#> 16            0        0        0     0     0     0        1      0      1
#> 17            1        1        0     1     0     1        1      0      1
#> 18            0        0        0     0     0     0        0      0      0
#> 22            0        0        0     1     0     1        1      0      0
#> 23            1        0        1     1     1     1        1      1      1
#> 24            0        0        0     1     0     1        1      0      1
#> 25            0        0        0     1     0     1        1      0      0
#> 26            1        1        0     0     0     0        0      0      1
#> 28            1        1        0     0     0     0        0      0      1
#> 29            1        1        0     0     0     0        0      0      1
#> 30            1        1        0     1     0     1        1      0      1
#> 31            1        1        0     1     0     1        0      0      1
#> 32            0        0        0     1     0     1        1      0      0
#> 37            1        1        0     1     0     1        1      0      1
#> 38            0        0        0     1     0     1        1      0      0
#> 39            0        0        0     0     0     0        0      0      0
#> 41            0        0        0     0     0     0        1      0      1
#> 44            1        1        0     1     0     1        1      0      1
#> 45            0        0        0     0     0     0        1      0      1
#> 46            1        1        1     1     0     1        1      0      1
#> 47            1        1        0     1     0     1        1      0      1
#> 48            1        1        0     1     0     1        1      0      1
#> 49            1        1        0     1     0     1        1      0      1
#> 50            0        0        0     1     0     1        1      0      0
#> 51            0        0        0     1     0     1        0      0      0
#> 52            1        1        1     0     1     1        0      1      1
#> 53            0        0        0     0     0     0        0      0      0
#> 55            1        0        1     1     0     1        1      0      1
#> 56            0        0        0     1     0     1        1      0      0
#> 57            1        1        0     1     0     1        1      0      1
#> 58            1        1        0     0     0     0        1      0      1
#> 60            1        1        0     1     0     1        1      0      1
#> 62            0        0        0     1     0     1        1      0      0
#> 63            1        1        0     1     0     1        1      0      1
#> 64            1        1        0     1     0     1        1      0      1
#> 67            0        0        0     1     0     1        1      0      1
#> 72            1        1        0     1     0     1        1      0      1
#> 74            0        0        0     1     0     1        1      0      0
#> 76            0        0        0     1     0     1        1      0      1
#> 78            0        0        0     0     0     0        0      0      0
#> 79            0        0        0     0     0     0        1      0      1
#> 80            1        1        0     1     0     1        1      0      1
#> 82            0        0        0     0     0     0        1      0      1
#> 83            0        0        0     0     0     0        0      0      0
#> 85            0        0        0     0     0     0        1      0      1
#> 86            0        0        0     1     0     1        1      0      0
#> 88            0        0        0     0     0     0        1      0      1
#> 89            1        1        0     1     0     1        0      0      1
#> 91            1        1        0     0     0     0        0      0      1
#> 93            1        1        0     0     0     0        1      0      1
#> 94            0        0        0     1     0     1        1      0      0
#> 96            1        1        0     1     0     1        1      0      1
#> 97            0        0        0     0     0     0        1      0      1
#> 98            1        0        1     1     1     1        1      1      0
#> 99            0        0        0     1     0     1        0      0      0
#> 102           1        1        0     1     0     1        1      0      1
#> 109           0        0        0     1     0     1        1      0      0
#> 110           1        1        0     1     0     1        1      0      1
#> 111           1        0        1     1     0     1        1      0      1
#> 113           1        1        0     0     0     0        0      0      1
#> 115           0        0        0     0     0     0        1      0      1
#> 117           0        0        0     1     0     1        1      0      0
#> 122           1        1        0     1     0     1        1      0      1
#> 124           0        0        0     1     0     1        1      0      0
#> 125           0        0        0     1     0     1        1      0      0
#> 128           1        1        1     0     1     1        0      0      1
#> 129           0        0        0     1     0     1        0      0      0
#> 130           1        1        0     1     0     1        1      0      1
#> 131           1        0        1     1     0     1        1      0      1
#> 132           0        0        0     1     0     1        1      0      0
#> 133           0        0        0     0     0     0        0      0      0
#> 135           0        0        0     0     0     0        0      0      0
#> 137           0        0        0     1     0     1        1      0      0
#> 139           0        0        0     1     0     1        1      0      0
#> 140           0        0        0     0     0     0        0      0      0
#> 144           1        1        0     1     0     1        1      0      1
#> 146           0        0        0     0     0     0        1      0      1
#> 147           0        0        0     1     0     1        1      0      1
#> 148           0        0        0     1     0     1        1      0      0
#> 149           0        0        0     0     0     0        0      0      0
#> 153           0        0        0     0     0     0        0      0      0
#> 154           1        1        1     1     0     1        1      0      1
#> 155           0        0        0     0     0     0        1      0      1
#> 156           1        1        0     1     0     1        1      0      1
#> 157           1        1        1     1     0     1        1      0      1
#> 159           0        0        0     0     0     0        0      0      0
#> 162           0        0        0     1     0     1        1      0      0
#> 164           0        0        0     0     0     0        0      0      0
#> 166           1        1        0     1     0     1        1      0      1
#> 167           1        1        1     1     1     1        1      0      1
#> 168           1        1        0     1     0     1        1      0      1
#> 169           0        0        0     0     0     0        0      0      0
#> 170           0        0        0     0     0     0        0      0      0
#> 171           1        1        0     0     0     0        0      0      1
#> 175           0        0        0     1     0     1        1      0      0
#> 177           0        0        0     1     0     1        1      0      0
#> 178           0        0        0     1     0     1        1      0      0
#> 182           1        1        0     1     0     1        1      0      1
#> 184           1        1        0     1     0     1        1      0      1
#> 185           0        0        0     1     0     1        1      0      0
#> 186           0        0        0     1     0     1        0      0      0
#> 188           1        1        0     1     0     1        1      0      1
#> 190           1        0        1     1     1     1        1      1      0
#> 191           1        1        1     0     0     0        1      0      1
#> 192           0        0        0     1     0     1        1      0      0
#>     vitB1 vitB2 vitB3 vitB6 vitB12 vitBcomplex
#> 1       1     1     1     1      1           1
#> 4       1     1     1     1      0           0
#> 5       1     1     1     1      1           1
#> 8       1     1     1     1      0           0
#> 10      1     1     1     1      1           1
#> 11      0     1     0     1      0           0
#> 12      0     0     0     0      0           0
#> 15      1     1     1     1      1           1
#> 16      1     1     1     1      1           1
#> 17      1     1     1     1      0           0
#> 18      0     0     0     0      0           0
#> 22      0     1     0     1      0           0
#> 23      1     1     1     1      1           1
#> 24      1     1     1     1      1           1
#> 25      0     1     0     1      0           0
#> 26      1     1     1     1      0           0
#> 28      1     1     1     1      0           0
#> 29      1     1     1     1      0           0
#> 30      1     1     1     1      1           1
#> 31      1     1     1     1      0           0
#> 32      0     1     0     1      0           0
#> 37      1     1     1     1      1           1
#> 38      0     1     0     1      0           0
#> 39      0     0     0     0      0           0
#> 41      1     1     1     1      1           1
#> 44      1     1     1     1      1           1
#> 45      1     1     1     1      1           1
#> 46      1     1     1     1      1           1
#> 47      1     1     1     1      0           0
#> 48      1     1     1     1      0           0
#> 49      1     1     1     1      0           0
#> 50      0     1     0     1      0           0
#> 51      1     0     0     0      0           0
#> 52      1     1     1     1      1           1
#> 53      0     0     0     0      0           0
#> 55      1     1     1     1      1           1
#> 56      0     1     0     1      0           0
#> 57      1     1     1     1      0           0
#> 58      1     1     1     1      1           1
#> 60      1     1     1     1      0           0
#> 62      0     1     0     1      0           0
#> 63      1     1     1     1      0           0
#> 64      1     1     1     1      0           0
#> 67      1     1     1     1      1           1
#> 72      1     1     1     1      0           0
#> 74      0     1     0     1      0           0
#> 76      1     1     1     1      1           1
#> 78      0     0     0     0      0           0
#> 79      1     1     1     1      1           1
#> 80      1     1     1     1      0           0
#> 82      1     1     1     1      1           1
#> 83      0     0     0     1      0           0
#> 85      1     1     1     1      1           1
#> 86      0     1     0     1      0           0
#> 88      1     1     1     1      1           1
#> 89      1     1     1     1      0           0
#> 91      1     1     1     1      0           0
#> 93      1     1     1     1      1           1
#> 94      0     1     0     1      0           0
#> 96      1     1     1     1      0           0
#> 97      1     1     1     1      1           1
#> 98      1     1     0     1      0           0
#> 99      0     0     0     1      0           0
#> 102     1     1     1     1      1           1
#> 109     0     1     0     1      0           0
#> 110     1     1     1     1      0           0
#> 111     1     1     1     1      1           1
#> 113     1     1     1     1      0           0
#> 115     1     1     1     1      1           1
#> 117     0     1     0     1      0           0
#> 122     1     1     1     1      1           1
#> 124     0     1     0     1      0           0
#> 125     0     1     0     1      0           0
#> 128     1     1     1     1      1           1
#> 129     1     0     0     0      0           0
#> 130     1     1     1     1      0           0
#> 131     1     1     1     1      1           1
#> 132     0     1     0     1      0           0
#> 133     0     0     0     0      0           0
#> 135     0     0     0     0      0           0
#> 137     0     1     0     1      0           0
#> 139     0     1     0     1      0           0
#> 140     0     0     0     1      0           0
#> 144     1     1     1     1      0           0
#> 146     1     1     1     1      1           1
#> 147     1     1     1     1      1           1
#> 148     0     1     0     1      0           0
#> 149     0     0     0     0      0           0
#> 153     0     0     0     0      0           0
#> 154     1     1     1     1      1           1
#> 155     1     1     1     1      1           1
#> 156     1     1     1     1      1           1
#> 157     1     1     1     1      1           1
#> 159     0     0     0     0      0           0
#> 162     0     1     0     1      0           0
#> 164     0     0     0     0      0           0
#> 166     1     1     1     1      1           1
#> 167     1     1     1     1      1           1
#> 168     1     1     1     1      0           0
#> 169     0     0     0     0      0           0
#> 170     0     0     0     0      0           0
#> 171     1     1     1     1      0           0
#> 175     0     1     0     1      0           0
#> 177     0     1     0     1      0           0
#> 178     0     1     0     1      0           0
#> 182     1     1     1     1      1           1
#> 184     1     1     1     1      1           1
#> 185     0     1     0     1      0           0
#> 186     1     0     0     0      0           0
#> 188     1     1     1     1      0           0
#> 190     1     1     0     1      1           0
#> 191     1     1     1     1      1           1
#> 192     0     1     0     1      0           0