Create male older people indicators dataframe for food intake from survey data collected using the standard RAM-OP questionnaire

Create male older people indicators dataframe for food intake from survey data collected using the standard RAM-OP questionnaire

create_op_food_males(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of male older people indicators on food intake

Examples

# Create food intake indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_food_males(testSVY)

#>     psu sex1 sex2 MF DDS FG01 FG02 FG03 FG04 FG05 FG06 FG07 FG08 FG09 FG10 FG11
#> 2   201    1    0  3   7    1    1    1    0    0    0    1    0    1    1    1
#> 3   201    1    0  2   2    0    0    0    0    0    1    0    0    0    0    1
#> 6   201    1    0  4   5    1    1    0    0    0    1    1    0    0    0    1
#> 7   201    1    0  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 9   201    1    0  2   4    1    0    1    0    0    0    0    0    0    1    1
#> 13  202    1    0  3   4    1    0    1    0    0    0    1    0    0    0    1
#> 14  202    1    0  4   6    1    1    0    0    0    0    1    0    1    1    1
#> 19  202    1    0  2   6    1    1    1    0    0    1    1    0    0    0    1
#> 20  202    1    0  2   5    1    0    1    0    0    1    1    0    0    0    1
#> 21  202    1    0  2   5    1    0    1    0    0    1    1    0    0    0    1
#> 27  203    1    0  3   8    1    1    1    1    0    1    0    0    1    1    1
#> 33  203    1    0  3   5    1    1    1    0    0    0    0    0    1    0    1
#> 34  203    1    0  3   7    1    1    1    0    0    1    1    0    0    1    1
#> 35  203    1    0  3   4    0    1    1    0    0    1    0    0    0    0    1
#> 36  203    1    0  2   7    1    1    1    0    1    0    1    0    0    1    1
#> 40  204    1    0  2   3    1    0    1    0    0    0    1    0    0    0    0
#> 42  204    1    0  2   6    1    1    0    0    0    1    0    1    0    1    1
#> 43  204    1    0  4   5    1    0    0    0    0    1    0    0    1    1    1
#> 54  205    1    0  2   4    0    1    0    0    0    1    0    0    0    1    1
#> 59  205    1    0  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 61  206    1    0  2   6    1    1    1    1    0    1    0    0    0    0    1
#> 65  206    1    0  4   3    0    0    1    0    0    0    0    0    0    1    1
#> 66  206    1    0  3   4    1    1    0    0    0    0    0    0    0    1    1
#> 68  206    1    0  3   6    1    1    0    0    0    0    1    0    1    1    1
#> 69  206    1    0  4   6    1    0    1    0    1    0    0    0    1    1    1
#> 70  206    1    0  1   1    0    0    0    0    0    0    0    0    0    0    1
#> 71  206    1    0  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 73  207    1    0  3   6    1    1    0    0    0    1    1    0    1    0    1
#> 75  207    1    0  3   4    1    0    1    0    0    1    0    0    1    0    0
#> 77  207    1    0  4   5    1    1    1    0    0    0    0    0    0    1    1
#> 81  207    1    0  3   7    1    1    1    0    0    1    1    0    0    1    1
#> 84  207    1    0  4   7    1    1    1    0    0    1    0    0    1    1    1
#> 87  208    1    0  3   5    1    1    1    0    0    0    1    0    0    0    1
#> 90  208    1    0  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 92  208    1    0  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 95  208    1    0  3   5    1    1    0    0    0    1    1    0    0    0    1
#> 100 209    1    0  2   4    1    0    1    0    0    0    0    0    0    1    1
#> 101 209    1    0  2   3    1    0    0    0    0    1    0    0    0    0    1
#> 103 209    1    0  1   4    1    1    0    0    0    0    1    0    0    0    1
#> 104 209    1    0  4   6    1    1    0    0    0    0    1    0    1    1    1
#> 105 209    1    0  1   4    1    1    0    0    0    0    1    0    0    0    1
#> 106 209    1    0  1   4    1    0    0    0    0    1    0    0    0    1    1
#> 107 209    1    0  3   5    1    0    0    0    0    1    0    0    1    1    1
#> 108 209    1    0  2   3    1    0    1    0    0    0    0    0    0    0    1
#> 112 210    1    0  2   6    1    1    1    0    0    1    1    0    0    0    1
#> 114 210    1    0  4   5    1    1    0    1    0    0    0    0    0    1    1
#> 116 210    1    0  2   4    1    0    0    0    0    0    1    0    0    1    1
#> 118 210    1    0  1   4    1    0    0    0    0    0    0    0    1    1    1
#> 119 210    1    0  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 120 210    1    0  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 121 211    1    0  3   5    1    0    1    0    0    0    1    0    0    1    1
#> 123 211    1    0  2   5    1    0    0    0    0    0    1    0    1    1    1
#> 126 211    1    0  4   7    1    1    1    0    1    1    1    0    0    0    1
#> 127 211    1    0  1   3    0    0    1    0    0    0    0    0    0    1    1
#> 134 212    1    0  3   4    0    1    0    0    0    1    0    0    0    1    1
#> 136 212    1    0  4   7    1    1    1    0    0    1    1    0    0    1    1
#> 138 212    1    0  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 141 212    1    0  3   4    1    0    0    0    0    0    1    0    0    1    1
#> 142 212    1    0  3   6    1    1    1    0    0    0    1    0    1    0    1
#> 143 212    1    0  1   3    1    0    0    0    0    1    0    0    0    0    1
#> 145 213    1    0  2   4    1    1    0    0    0    0    0    0    0    1    1
#> 150 213    1    0  2   3    1    0    1    0    0    0    0    0    0    0    1
#> 151 213    1    0  4   5    1    1    1    0    0    0    0    0    0    1    1
#> 152 213    1    0  5   5    1    0    1    0    0    0    0    0    1    1    1
#> 158 214    1    0  3   6    1    1    1    0    0    1    0    0    1    0    1
#> 160 214    1    0  3   6    1    1    0    0    0    1    0    0    1    1    1
#> 161 214    1    0  4   5    1    1    1    0    0    1    0    0    0    0    1
#> 163 214    1    0  3   5    1    1    1    0    0    1    0    0    0    0    1
#> 165 214    1    0  2   5    1    0    0    0    0    0    1    0    1    1    1
#> 172 215    1    0  3   7    1    1    1    0    0    1    1    0    1    0    1
#> 173 215    1    0  2   3    1    0    0    0    0    1    0    0    0    0    1
#> 174 215    1    0  3   3    1    0    1    0    0    0    0    0    0    0    1
#> 176 215    1    0  4   3    1    0    0    0    0    0    0    0    0    1    1
#> 179 215    1    0  2   3    1    1    0    0    0    0    1    0    0    0    0
#> 180 215    1    0  3   3    1    0    1    0    0    0    0    0    0    0    1
#> 181 216    1    0  3   2    1    0    0    0    0    0    0    0    0    0    1
#> 183 216    1    0  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 187 216    1    0  2   4    1    0    1    0    0    1    0    0    0    0    1
#> 189 216    1    0  2   4    1    0    1    0    0    0    0    0    0    1    1
#>     proteinRich pProtein aProtein pVitA aVitA xVitA ironRich caRich znRich
#> 2             1        1        0     1     0     1        1      0      1
#> 3             0        0        0     0     0     0        1      0      1
#> 6             1        1        0     0     0     0        1      0      1
#> 7             1        1        0     1     0     1        1      0      1
#> 9             0        0        0     1     0     1        1      0      0
#> 13            1        1        0     1     0     1        1      0      1
#> 14            1        1        0     0     0     0        0      0      1
#> 19            1        1        0     1     0     1        1      0      1
#> 20            1        1        0     1     0     1        1      0      1
#> 21            1        1        0     1     0     1        1      0      1
#> 27            1        0        1     1     0     1        1      0      1
#> 33            0        0        0     1     0     1        0      0      0
#> 34            1        1        0     1     0     1        1      0      1
#> 35            0        0        0     1     0     1        1      0      1
#> 36            1        1        1     0     1     1        0      0      1
#> 40            1        1        0     1     0     1        1      0      1
#> 42            1        0        1     0     1     1        1      1      1
#> 43            0        0        0     0     0     0        1      0      1
#> 54            0        0        0     0     0     0        1      0      1
#> 59            0        0        0     1     0     1        0      0      0
#> 61            1        0        1     0     0     0        1      0      1
#> 65            0        0        0     1     0     1        1      0      0
#> 66            0        0        0     0     0     0        0      0      0
#> 68            1        1        0     0     0     0        0      0      1
#> 69            1        0        1     1     1     1        1      0      0
#> 70            0        0        0     0     0     0        0      0      0
#> 71            0        0        0     1     0     1        1      0      0
#> 73            1        1        0     0     0     0        1      0      1
#> 75            0        0        0     1     0     1        1      0      1
#> 77            0        0        0     1     0     1        1      0      0
#> 81            1        1        0     1     0     1        1      0      1
#> 84            0        0        0     1     0     1        1      0      1
#> 87            1        1        0     1     0     1        1      0      1
#> 90            0        0        0     0     0     0        0      0      0
#> 92            0        0        0     0     0     0        0      0      0
#> 95            1        1        0     0     0     0        1      0      1
#> 100           0        0        0     1     0     1        1      0      0
#> 101           0        0        0     0     0     0        1      0      1
#> 103           1        1        0     0     0     0        0      0      1
#> 104           1        1        0     0     0     0        0      0      1
#> 105           1        1        0     0     0     0        0      0      1
#> 106           0        0        0     0     0     0        1      0      1
#> 107           0        0        0     1     0     1        1      0      1
#> 108           0        0        0     1     0     1        1      0      0
#> 112           1        1        0     1     0     1        1      0      1
#> 114           1        0        1     0     0     0        1      0      1
#> 116           1        1        0     0     0     0        0      0      1
#> 118           0        0        0     1     0     1        0      0      0
#> 119           0        0        0     0     0     0        0      0      0
#> 120           0        0        0     1     0     1        0      0      0
#> 121           1        1        0     1     0     1        0      0      1
#> 123           1        1        0     1     0     1        0      0      1
#> 126           1        1        1     1     1     1        1      0      1
#> 127           0        0        0     0     0     0        0      0      0
#> 134           0        0        0     0     0     0        1      0      1
#> 136           1        1        0     1     0     1        1      0      1
#> 138           0        0        0     0     0     0        0      0      0
#> 141           1        1        0     0     0     0        0      0      1
#> 142           1        1        0     1     0     1        1      0      1
#> 143           0        0        0     0     0     0        1      0      1
#> 145           0        0        0     0     0     0        0      0      0
#> 150           0        0        0     1     0     1        1      0      0
#> 151           0        0        0     1     0     1        1      0      0
#> 152           0        0        0     1     0     1        1      0      0
#> 158           0        0        0     1     0     1        1      0      1
#> 160           0        0        0     1     0     1        1      0      1
#> 161           0        0        0     1     0     1        1      0      1
#> 163           0        0        0     1     0     1        1      0      1
#> 165           1        1        0     0     0     0        0      0      1
#> 172           1        1        0     1     0     1        1      0      1
#> 173           0        0        0     0     0     0        1      0      1
#> 174           0        0        0     1     0     1        1      0      0
#> 176           0        0        0     0     0     0        0      0      0
#> 179           1        1        0     0     0     0        0      0      1
#> 180           0        0        0     1     0     1        0      0      0
#> 181           0        0        0     0     0     0        0      0      0
#> 183           0        0        0     1     0     1        1      0      0
#> 187           0        0        0     1     0     1        1      0      1
#> 189           0        0        0     1     0     1        1      0      0
#>     vitB1 vitB2 vitB3 vitB6 vitB12 vitBcomplex
#> 2       1     1     1     1      0           0
#> 3       1     1     1     1      1           1
#> 6       1     1     1     1      1           1
#> 7       1     1     1     1      1           1
#> 9       0     1     0     1      0           0
#> 13      1     1     1     1      0           0
#> 14      1     1     1     1      0           0
#> 19      1     1     1     1      1           1
#> 20      1     1     1     1      1           1
#> 21      1     1     1     1      1           1
#> 27      1     1     1     1      1           1
#> 33      0     0     0     1      0           0
#> 34      1     1     1     1      1           1
#> 35      1     1     1     1      1           1
#> 36      1     1     1     1      1           1
#> 40      1     1     1     1      0           0
#> 42      1     1     1     1      1           1
#> 43      1     1     1     1      1           1
#> 54      1     1     1     1      1           1
#> 59      0     0     0     1      0           0
#> 61      1     1     1     1      1           1
#> 65      0     1     0     1      0           0
#> 66      0     0     0     0      0           0
#> 68      1     1     1     1      0           0
#> 69      1     1     0     1      1           0
#> 70      0     0     0     0      0           0
#> 71      0     1     0     1      0           0
#> 73      1     1     1     1      1           1
#> 75      1     1     1     1      1           1
#> 77      0     1     0     1      0           0
#> 81      1     1     1     1      1           1
#> 84      1     1     1     1      1           1
#> 87      1     1     1     1      0           0
#> 90      0     0     0     0      0           0
#> 92      0     0     0     0      0           0
#> 95      1     1     1     1      1           1
#> 100     0     1     0     1      0           0
#> 101     1     1     1     1      1           1
#> 103     1     1     1     1      0           0
#> 104     1     1     1     1      0           0
#> 105     1     1     1     1      0           0
#> 106     1     1     1     1      1           1
#> 107     1     1     1     1      1           1
#> 108     0     1     0     1      0           0
#> 112     1     1     1     1      1           1
#> 114     1     1     1     1      1           1
#> 116     1     1     1     1      0           0
#> 118     1     0     0     0      0           0
#> 119     0     0     0     0      0           0
#> 120     0     0     0     1      0           0
#> 121     1     1     1     1      0           0
#> 123     1     1     1     1      0           0
#> 126     1     1     1     1      1           1
#> 127     0     0     0     1      0           0
#> 134     1     1     1     1      1           1
#> 136     1     1     1     1      1           1
#> 138     0     0     0     0      0           0
#> 141     1     1     1     1      0           0
#> 142     1     1     1     1      0           0
#> 143     1     1     1     1      1           1
#> 145     0     0     0     0      0           0
#> 150     0     1     0     1      0           0
#> 151     0     1     0     1      0           0
#> 152     1     1     0     1      0           0
#> 158     1     1     1     1      1           1
#> 160     1     1     1     1      1           1
#> 161     1     1     1     1      1           1
#> 163     1     1     1     1      1           1
#> 165     1     1     1     1      0           0
#> 172     1     1     1     1      1           1
#> 173     1     1     1     1      1           1
#> 174     0     1     0     1      0           0
#> 176     0     0     0     0      0           0
#> 179     1     1     1     1      0           0
#> 180     0     0     0     1      0           0
#> 181     0     0     0     0      0           0
#> 183     0     1     0     1      0           0
#> 187     1     1     1     1      1           1
#> 189     0     1     0     1      0           0