Create older people indicators dataframe for health and health-seeking behaviours from survey data collected using the standard RAM-OP questionnaire.

Create older people indicators dataframe for health and health-seeking behaviours from survey data collected using the standard RAM-OP questionnaire.

create_op_health(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of older people indicators on health and health-seeking behaviour

Indicators

Health and health-seeking indicators

H1

Chronic condition

H2

Takes drugs regularly for chronic condition

H31

No drugs available

H32

Too expensive / no money

H33

Too old to look for care

H34

Use traditional medicine

H35

Drugs don't help

H36

No-one to help me

H37

No need

H38

Other

H39

No reason given

H4

Recent disease episode

H5

Accessed care for recent disease episode

H61

No drugs available

H62

Too expensive / no money

H63

Too old to look for care

H64

Use traditional medicine

H65

Drugs don't help

H66

No-one to help me

H67

No need

H68

Other

H69

No reason given

Examples

# Create health and health-seeking behaviour indicators dataset from RAM-OP
# survey data collected from Addis Ababa, Ethiopia
create_op_health(testSVY)

#>     psu sex1 sex2 H1 H2 H31 H32 H33 H34 H35 H36 H37 H38 H39 H4 H5 H61 H62 H63
#> 1   201    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 2   201    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 3   201    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 4   201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 5   201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 6   201    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 7   201    1    0  1  0   0   0   0   1   0   0   0   0   0  1  0   0   0   0
#> 8   201    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 9   201    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 10  201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 11  201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 12  201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 13  202    1    0  1  0   0   1   0   0   0   0   0   0   0  1  0   0   1   0
#> 14  202    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 15  202    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 16  202    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 17  202    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 18  202    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 19  202    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 20  202    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 21  202    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 22  202    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 23  202    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 24  202    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 25  203    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 26  203    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 27  203    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 28  203    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 29  203    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 30  203    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 31  203    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 32  203    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 33  203    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 34  203    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 35  203    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 36  203    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 37  204    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 38  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 39  204    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 40  204    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 41  204    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 42  204    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 43  204    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 44  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 45  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 46  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 47  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 48  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 49  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 50  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 51  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 52  205    0    1  1  0   0   1   0   0   0   0   0   0   0  1  1  NA  NA  NA
#> 53  205    0    1  1  0   0   0   0   0   0   0   0   0   1  1  1  NA  NA  NA
#> 54  205    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 55  205    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 56  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 57  205    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 58  205    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 59  205    1    0  1  0   0   0   0   0   0   0   0   0   1  1  0   0   1   0
#> 60  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 61  206    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 62  206    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 63  206    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 64  206    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 65  206    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 66  206    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 67  206    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 68  206    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 69  206    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 70  206    1    0  1  0   0   1   0   0   0   0   0   0   0  1  1  NA  NA  NA
#> 71  206    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 72  206    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 73  207    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   1   0   0
#> 74  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 75  207    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 76  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 77  207    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 78  207    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 79  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   0   0
#> 80  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 81  207    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 82  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 83  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 84  207    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 85  208    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 86  208    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 87  208    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 88  208    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 89  208    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 90  208    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 91  208    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 92  208    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 93  208    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 94  208    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 95  208    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 96  208    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 97  209    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 98  209    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 99  209    0    1  1  0   0   0   1   0   0   0   0   0   0  1  0   0   1   0
#> 100 209    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 101 209    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 102 209    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 103 209    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 104 209    1    0  1  0   0   1   0   0   0   0   0   0   0  1  0   0   1   0
#> 105 209    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 106 209    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 107 209    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 108 209    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 109 210    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 110 210    0    1  1  0   0   1   0   0   0   0   0   0   0  1  0   0   1   0
#> 111 210    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 112 210    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 113 210    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 114 210    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 115 210    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 116 210    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 117 210    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 118 210    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 119 210    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 120 210    1    0  1  0   0   1   0   0   0   0   0   0   0  1  0   0   1   0
#> 121 211    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 122 211    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 123 211    1    0  1  0   1   0   0   0   0   0   0   0   0  1  1  NA  NA  NA
#> 124 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 125 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 126 211    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 127 211    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   1   0   0
#> 128 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   0   0
#> 129 211    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 130 211    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 131 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 132 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 133 212    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 134 212    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 135 212    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 136 212    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 137 212    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 138 212    1    0  1  0   0   0   0   0   0   0   0   0   1  1  0   0   1   0
#> 139 212    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 140 212    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 141 212    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 142 212    1    0  1  0   1   0   0   0   0   0   0   0   0  1  1  NA  NA  NA
#> 143 212    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 144 212    0    1  1  0   0   0   0   0   0   0   1   0   0  1  0   0   1   0
#> 145 213    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 146 213    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 147 213    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 148 213    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 149 213    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 150 213    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 151 213    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 152 213    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 153 213    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 154 213    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 155 213    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 156 213    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 157 214    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 158 214    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 159 214    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 160 214    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 161 214    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 162 214    0    1  1  0   0   0   1   0   0   0   0   0   0  1  0   0   1   0
#> 163 214    1    0  1  0   0   0   0   1   0   0   0   0   0  1  0   0   1   0
#> 164 214    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 165 214    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 166 214    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 167 214    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 168 214    0    1  1  0   0   1   0   0   0   0   0   0   0  1  0   0   1   0
#> 169 215    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 170 215    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 171 215    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 172 215    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 173 215    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 174 215    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 175 215    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 176 215    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 177 215    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 178 215    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 179 215    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 180 215    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 181 216    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 182 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 183 216    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 184 216    0    1  1  0   1   0   0   0   0   0   0   0   0  1  0   1   0   0
#> 185 216    0    1  1  0   0   1   0   0   0   0   0   0   0  1  0   0   1   0
#> 186 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 187 216    1    0  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 188 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 189 216    1    0  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 190 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 191 216    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 192 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#>     H64 H65 H66 H67 H68 H69
#> 1    NA  NA  NA  NA  NA  NA
#> 2    NA  NA  NA  NA  NA  NA
#> 3    NA  NA  NA  NA  NA  NA
#> 4    NA  NA  NA  NA  NA  NA
#> 5    NA  NA  NA  NA  NA  NA
#> 6    NA  NA  NA  NA  NA  NA
#> 7     1   0   0   0   0   0
#> 8    NA  NA  NA  NA  NA  NA
#> 9    NA  NA  NA  NA  NA  NA
#> 10   NA  NA  NA  NA  NA  NA
#> 11   NA  NA  NA  NA  NA  NA
#> 12   NA  NA  NA  NA  NA  NA
#> 13    0   0   0   0   0   0
#> 14   NA  NA  NA  NA  NA  NA
#> 15   NA  NA  NA  NA  NA  NA
#> 16   NA  NA  NA  NA  NA  NA
#> 17   NA  NA  NA  NA  NA  NA
#> 18   NA  NA  NA  NA  NA  NA
#> 19   NA  NA  NA  NA  NA  NA
#> 20   NA  NA  NA  NA  NA  NA
#> 21   NA  NA  NA  NA  NA  NA
#> 22   NA  NA  NA  NA  NA  NA
#> 23   NA  NA  NA  NA  NA  NA
#> 24   NA  NA  NA  NA  NA  NA
#> 25   NA  NA  NA  NA  NA  NA
#> 26    0   0   0   0   0   0
#> 27   NA  NA  NA  NA  NA  NA
#> 28   NA  NA  NA  NA  NA  NA
#> 29   NA  NA  NA  NA  NA  NA
#> 30   NA  NA  NA  NA  NA  NA
#> 31   NA  NA  NA  NA  NA  NA
#> 32   NA  NA  NA  NA  NA  NA
#> 33   NA  NA  NA  NA  NA  NA
#> 34   NA  NA  NA  NA  NA  NA
#> 35   NA  NA  NA  NA  NA  NA
#> 36    0   0   0   0   0   0
#> 37    0   0   0   0   0   0
#> 38   NA  NA  NA  NA  NA  NA
#> 39   NA  NA  NA  NA  NA  NA
#> 40   NA  NA  NA  NA  NA  NA
#> 41   NA  NA  NA  NA  NA  NA
#> 42   NA  NA  NA  NA  NA  NA
#> 43   NA  NA  NA  NA  NA  NA
#> 44   NA  NA  NA  NA  NA  NA
#> 45    0   0   0   0   0   0
#> 46   NA  NA  NA  NA  NA  NA
#> 47   NA  NA  NA  NA  NA  NA
#> 48   NA  NA  NA  NA  NA  NA
#> 49   NA  NA  NA  NA  NA  NA
#> 50   NA  NA  NA  NA  NA  NA
#> 51   NA  NA  NA  NA  NA  NA
#> 52   NA  NA  NA  NA  NA  NA
#> 53   NA  NA  NA  NA  NA  NA
#> 54   NA  NA  NA  NA  NA  NA
#> 55   NA  NA  NA  NA  NA  NA
#> 56   NA  NA  NA  NA  NA  NA
#> 57   NA  NA  NA  NA  NA  NA
#> 58   NA  NA  NA  NA  NA  NA
#> 59    0   0   0   0   0   0
#> 60   NA  NA  NA  NA  NA  NA
#> 61   NA  NA  NA  NA  NA  NA
#> 62   NA  NA  NA  NA  NA  NA
#> 63   NA  NA  NA  NA  NA  NA
#> 64    0   0   0   0   0   0
#> 65   NA  NA  NA  NA  NA  NA
#> 66   NA  NA  NA  NA  NA  NA
#> 67   NA  NA  NA  NA  NA  NA
#> 68   NA  NA  NA  NA  NA  NA
#> 69   NA  NA  NA  NA  NA  NA
#> 70   NA  NA  NA  NA  NA  NA
#> 71   NA  NA  NA  NA  NA  NA
#> 72   NA  NA  NA  NA  NA  NA
#> 73    0   0   0   0   0   0
#> 74   NA  NA  NA  NA  NA  NA
#> 75   NA  NA  NA  NA  NA  NA
#> 76   NA  NA  NA  NA  NA  NA
#> 77   NA  NA  NA  NA  NA  NA
#> 78    0   0   0   0   0   0
#> 79    0   0   0   0   1   0
#> 80   NA  NA  NA  NA  NA  NA
#> 81   NA  NA  NA  NA  NA  NA
#> 82   NA  NA  NA  NA  NA  NA
#> 83   NA  NA  NA  NA  NA  NA
#> 84   NA  NA  NA  NA  NA  NA
#> 85   NA  NA  NA  NA  NA  NA
#> 86   NA  NA  NA  NA  NA  NA
#> 87   NA  NA  NA  NA  NA  NA
#> 88   NA  NA  NA  NA  NA  NA
#> 89   NA  NA  NA  NA  NA  NA
#> 90   NA  NA  NA  NA  NA  NA
#> 91   NA  NA  NA  NA  NA  NA
#> 92   NA  NA  NA  NA  NA  NA
#> 93   NA  NA  NA  NA  NA  NA
#> 94    0   0   0   0   0   0
#> 95   NA  NA  NA  NA  NA  NA
#> 96   NA  NA  NA  NA  NA  NA
#> 97   NA  NA  NA  NA  NA  NA
#> 98   NA  NA  NA  NA  NA  NA
#> 99    0   0   0   0   0   0
#> 100   0   0   0   0   0   0
#> 101  NA  NA  NA  NA  NA  NA
#> 102  NA  NA  NA  NA  NA  NA
#> 103  NA  NA  NA  NA  NA  NA
#> 104   0   0   0   0   0   0
#> 105   0   0   0   0   0   0
#> 106  NA  NA  NA  NA  NA  NA
#> 107  NA  NA  NA  NA  NA  NA
#> 108  NA  NA  NA  NA  NA  NA
#> 109  NA  NA  NA  NA  NA  NA
#> 110   0   0   0   0   0   0
#> 111  NA  NA  NA  NA  NA  NA
#> 112  NA  NA  NA  NA  NA  NA
#> 113  NA  NA  NA  NA  NA  NA
#> 114  NA  NA  NA  NA  NA  NA
#> 115  NA  NA  NA  NA  NA  NA
#> 116  NA  NA  NA  NA  NA  NA
#> 117  NA  NA  NA  NA  NA  NA
#> 118  NA  NA  NA  NA  NA  NA
#> 119  NA  NA  NA  NA  NA  NA
#> 120   0   0   0   0   0   0
#> 121  NA  NA  NA  NA  NA  NA
#> 122  NA  NA  NA  NA  NA  NA
#> 123  NA  NA  NA  NA  NA  NA
#> 124  NA  NA  NA  NA  NA  NA
#> 125  NA  NA  NA  NA  NA  NA
#> 126  NA  NA  NA  NA  NA  NA
#> 127   0   0   0   0   0   0
#> 128   0   0   1   0   0   0
#> 129  NA  NA  NA  NA  NA  NA
#> 130  NA  NA  NA  NA  NA  NA
#> 131  NA  NA  NA  NA  NA  NA
#> 132  NA  NA  NA  NA  NA  NA
#> 133  NA  NA  NA  NA  NA  NA
#> 134  NA  NA  NA  NA  NA  NA
#> 135  NA  NA  NA  NA  NA  NA
#> 136  NA  NA  NA  NA  NA  NA
#> 137  NA  NA  NA  NA  NA  NA
#> 138   0   0   0   0   0   0
#> 139  NA  NA  NA  NA  NA  NA
#> 140  NA  NA  NA  NA  NA  NA
#> 141  NA  NA  NA  NA  NA  NA
#> 142  NA  NA  NA  NA  NA  NA
#> 143  NA  NA  NA  NA  NA  NA
#> 144   0   0   0   0   0   0
#> 145  NA  NA  NA  NA  NA  NA
#> 146  NA  NA  NA  NA  NA  NA
#> 147  NA  NA  NA  NA  NA  NA
#> 148  NA  NA  NA  NA  NA  NA
#> 149  NA  NA  NA  NA  NA  NA
#> 150  NA  NA  NA  NA  NA  NA
#> 151  NA  NA  NA  NA  NA  NA
#> 152  NA  NA  NA  NA  NA  NA
#> 153  NA  NA  NA  NA  NA  NA
#> 154  NA  NA  NA  NA  NA  NA
#> 155  NA  NA  NA  NA  NA  NA
#> 156  NA  NA  NA  NA  NA  NA
#> 157  NA  NA  NA  NA  NA  NA
#> 158  NA  NA  NA  NA  NA  NA
#> 159  NA  NA  NA  NA  NA  NA
#> 160  NA  NA  NA  NA  NA  NA
#> 161  NA  NA  NA  NA  NA  NA
#> 162   0   0   0   0   0   0
#> 163   0   0   0   0   0   0
#> 164  NA  NA  NA  NA  NA  NA
#> 165  NA  NA  NA  NA  NA  NA
#> 166  NA  NA  NA  NA  NA  NA
#> 167  NA  NA  NA  NA  NA  NA
#> 168   0   0   0   0   0   0
#> 169  NA  NA  NA  NA  NA  NA
#> 170  NA  NA  NA  NA  NA  NA
#> 171  NA  NA  NA  NA  NA  NA
#> 172  NA  NA  NA  NA  NA  NA
#> 173   0   0   0   0   0   0
#> 174  NA  NA  NA  NA  NA  NA
#> 175  NA  NA  NA  NA  NA  NA
#> 176  NA  NA  NA  NA  NA  NA
#> 177  NA  NA  NA  NA  NA  NA
#> 178  NA  NA  NA  NA  NA  NA
#> 179  NA  NA  NA  NA  NA  NA
#> 180   0   0   0   0   0   0
#> 181  NA  NA  NA  NA  NA  NA
#> 182  NA  NA  NA  NA  NA  NA
#> 183  NA  NA  NA  NA  NA  NA
#> 184   0   0   0   0   0   0
#> 185   0   0   0   0   0   0
#> 186  NA  NA  NA  NA  NA  NA
#> 187  NA  NA  NA  NA  NA  NA
#> 188  NA  NA  NA  NA  NA  NA
#> 189  NA  NA  NA  NA  NA  NA
#> 190  NA  NA  NA  NA  NA  NA
#> 191  NA  NA  NA  NA  NA  NA
#> 192  NA  NA  NA  NA  NA  NA