create_op_health_females

Create female older people indicators dataframe for health and health-seeking behaviours from survey data collected using the standard RAM-OP questionnaire

create_op_health_females(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of female older people indicators on health and health-seeking behaviours

Examples

# Create health and health-seeking behaviours indicators dataset from RAM-OP
# survey data collected from Addis Ababa, Ethiopia
create_op_health_females(testSVY)

#>     psu sex1 sex2 H1 H2 H31 H32 H33 H34 H35 H36 H37 H38 H39 H4 H5 H61 H62 H63
#> 1   201    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 4   201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 5   201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 8   201    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 10  201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 11  201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 12  201    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 15  202    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 16  202    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 17  202    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 18  202    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 22  202    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 23  202    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 24  202    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 25  203    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 26  203    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 28  203    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 29  203    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 30  203    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 31  203    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 32  203    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 37  204    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 38  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 39  204    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 41  204    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 44  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 45  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 46  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 47  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 48  204    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 49  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 50  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 51  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 52  205    0    1  1  0   0   1   0   0   0   0   0   0   0  1  1  NA  NA  NA
#> 53  205    0    1  1  0   0   0   0   0   0   0   0   0   1  1  1  NA  NA  NA
#> 55  205    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 56  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 57  205    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 58  205    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 60  205    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 62  206    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 63  206    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 64  206    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 67  206    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 72  206    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 74  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 76  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 78  207    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 79  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   0   0
#> 80  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 82  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 83  207    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 85  208    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 86  208    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 88  208    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 89  208    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 91  208    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 93  208    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 94  208    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   1   0
#> 96  208    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 97  209    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 98  209    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 99  209    0    1  1  0   0   0   1   0   0   0   0   0   0  1  0   0   1   0
#> 102 209    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 109 210    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 110 210    0    1  1  0   0   1   0   0   0   0   0   0   0  1  0   0   1   0
#> 111 210    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 113 210    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 115 210    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 117 210    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 122 211    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 124 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 125 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 128 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  0   0   0   0
#> 129 211    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 130 211    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 131 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 132 211    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 133 212    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 135 212    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 137 212    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 139 212    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 140 212    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 144 212    0    1  1  0   0   0   0   0   0   0   1   0   0  1  0   0   1   0
#> 146 213    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 147 213    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 148 213    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 149 213    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 153 213    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 154 213    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 155 213    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 156 213    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 157 214    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 159 214    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 162 214    0    1  1  0   0   0   1   0   0   0   0   0   0  1  0   0   1   0
#> 164 214    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 166 214    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 167 214    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 168 214    0    1  1  0   0   1   0   0   0   0   0   0   0  1  0   0   1   0
#> 169 215    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 170 215    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 171 215    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 175 215    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 177 215    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 178 215    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 182 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 184 216    0    1  1  0   1   0   0   0   0   0   0   0   0  1  0   1   0   0
#> 185 216    0    1  1  0   0   1   0   0   0   0   0   0   0  1  0   0   1   0
#> 186 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  0 NA  NA  NA  NA
#> 188 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 190 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 191 216    0    1  1  1  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#> 192 216    0    1  0 NA  NA  NA  NA  NA  NA  NA  NA  NA  NA  1  1  NA  NA  NA
#>     H64 H65 H66 H67 H68 H69
#> 1    NA  NA  NA  NA  NA  NA
#> 4    NA  NA  NA  NA  NA  NA
#> 5    NA  NA  NA  NA  NA  NA
#> 8    NA  NA  NA  NA  NA  NA
#> 10   NA  NA  NA  NA  NA  NA
#> 11   NA  NA  NA  NA  NA  NA
#> 12   NA  NA  NA  NA  NA  NA
#> 15   NA  NA  NA  NA  NA  NA
#> 16   NA  NA  NA  NA  NA  NA
#> 17   NA  NA  NA  NA  NA  NA
#> 18   NA  NA  NA  NA  NA  NA
#> 22   NA  NA  NA  NA  NA  NA
#> 23   NA  NA  NA  NA  NA  NA
#> 24   NA  NA  NA  NA  NA  NA
#> 25   NA  NA  NA  NA  NA  NA
#> 26    0   0   0   0   0   0
#> 28   NA  NA  NA  NA  NA  NA
#> 29   NA  NA  NA  NA  NA  NA
#> 30   NA  NA  NA  NA  NA  NA
#> 31   NA  NA  NA  NA  NA  NA
#> 32   NA  NA  NA  NA  NA  NA
#> 37    0   0   0   0   0   0
#> 38   NA  NA  NA  NA  NA  NA
#> 39   NA  NA  NA  NA  NA  NA
#> 41   NA  NA  NA  NA  NA  NA
#> 44   NA  NA  NA  NA  NA  NA
#> 45    0   0   0   0   0   0
#> 46   NA  NA  NA  NA  NA  NA
#> 47   NA  NA  NA  NA  NA  NA
#> 48   NA  NA  NA  NA  NA  NA
#> 49   NA  NA  NA  NA  NA  NA
#> 50   NA  NA  NA  NA  NA  NA
#> 51   NA  NA  NA  NA  NA  NA
#> 52   NA  NA  NA  NA  NA  NA
#> 53   NA  NA  NA  NA  NA  NA
#> 55   NA  NA  NA  NA  NA  NA
#> 56   NA  NA  NA  NA  NA  NA
#> 57   NA  NA  NA  NA  NA  NA
#> 58   NA  NA  NA  NA  NA  NA
#> 60   NA  NA  NA  NA  NA  NA
#> 62   NA  NA  NA  NA  NA  NA
#> 63   NA  NA  NA  NA  NA  NA
#> 64    0   0   0   0   0   0
#> 67   NA  NA  NA  NA  NA  NA
#> 72   NA  NA  NA  NA  NA  NA
#> 74   NA  NA  NA  NA  NA  NA
#> 76   NA  NA  NA  NA  NA  NA
#> 78    0   0   0   0   0   0
#> 79    0   0   0   0   1   0
#> 80   NA  NA  NA  NA  NA  NA
#> 82   NA  NA  NA  NA  NA  NA
#> 83   NA  NA  NA  NA  NA  NA
#> 85   NA  NA  NA  NA  NA  NA
#> 86   NA  NA  NA  NA  NA  NA
#> 88   NA  NA  NA  NA  NA  NA
#> 89   NA  NA  NA  NA  NA  NA
#> 91   NA  NA  NA  NA  NA  NA
#> 93   NA  NA  NA  NA  NA  NA
#> 94    0   0   0   0   0   0
#> 96   NA  NA  NA  NA  NA  NA
#> 97   NA  NA  NA  NA  NA  NA
#> 98   NA  NA  NA  NA  NA  NA
#> 99    0   0   0   0   0   0
#> 102  NA  NA  NA  NA  NA  NA
#> 109  NA  NA  NA  NA  NA  NA
#> 110   0   0   0   0   0   0
#> 111  NA  NA  NA  NA  NA  NA
#> 113  NA  NA  NA  NA  NA  NA
#> 115  NA  NA  NA  NA  NA  NA
#> 117  NA  NA  NA  NA  NA  NA
#> 122  NA  NA  NA  NA  NA  NA
#> 124  NA  NA  NA  NA  NA  NA
#> 125  NA  NA  NA  NA  NA  NA
#> 128   0   0   1   0   0   0
#> 129  NA  NA  NA  NA  NA  NA
#> 130  NA  NA  NA  NA  NA  NA
#> 131  NA  NA  NA  NA  NA  NA
#> 132  NA  NA  NA  NA  NA  NA
#> 133  NA  NA  NA  NA  NA  NA
#> 135  NA  NA  NA  NA  NA  NA
#> 137  NA  NA  NA  NA  NA  NA
#> 139  NA  NA  NA  NA  NA  NA
#> 140  NA  NA  NA  NA  NA  NA
#> 144   0   0   0   0   0   0
#> 146  NA  NA  NA  NA  NA  NA
#> 147  NA  NA  NA  NA  NA  NA
#> 148  NA  NA  NA  NA  NA  NA
#> 149  NA  NA  NA  NA  NA  NA
#> 153  NA  NA  NA  NA  NA  NA
#> 154  NA  NA  NA  NA  NA  NA
#> 155  NA  NA  NA  NA  NA  NA
#> 156  NA  NA  NA  NA  NA  NA
#> 157  NA  NA  NA  NA  NA  NA
#> 159  NA  NA  NA  NA  NA  NA
#> 162   0   0   0   0   0   0
#> 164  NA  NA  NA  NA  NA  NA
#> 166  NA  NA  NA  NA  NA  NA
#> 167  NA  NA  NA  NA  NA  NA
#> 168   0   0   0   0   0   0
#> 169  NA  NA  NA  NA  NA  NA
#> 170  NA  NA  NA  NA  NA  NA
#> 171  NA  NA  NA  NA  NA  NA
#> 175  NA  NA  NA  NA  NA  NA
#> 177  NA  NA  NA  NA  NA  NA
#> 178  NA  NA  NA  NA  NA  NA
#> 182  NA  NA  NA  NA  NA  NA
#> 184   0   0   0   0   0   0
#> 185   0   0   0   0   0   0
#> 186  NA  NA  NA  NA  NA  NA
#> 188  NA  NA  NA  NA  NA  NA
#> 190  NA  NA  NA  NA  NA  NA
#> 191  NA  NA  NA  NA  NA  NA
#> 192  NA  NA  NA  NA  NA  NA