create_op_screening_females

Create female older people indicators dataframe for screening coverage from survey data collected using the standard RAM-OP questionnaire

create_op_screening_females(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of female older people indicators on screening coverage

Examples

# Create screening coverage indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_screening_females(testSVY)

#>     psu sex1 sex2 screened
#> 1   201    0    1        0
#> 4   201    0    1        0
#> 5   201    0    1        0
#> 8   201    0    1        0
#> 10  201    0    1        0
#> 11  201    0    1        0
#> 12  201    0    1        0
#> 15  202    0    1        0
#> 16  202    0    1        0
#> 17  202    0    1        0
#> 18  202    0    1        0
#> 22  202    0    1        0
#> 23  202    0    1        0
#> 24  202    0    1        0
#> 25  203    0    1        0
#> 26  203    0    1        0
#> 28  203    0    1        0
#> 29  203    0    1        0
#> 30  203    0    1        0
#> 31  203    0    1        0
#> 32  203    0    1        0
#> 37  204    0    1        0
#> 38  204    0    1        0
#> 39  204    0    1        0
#> 41  204    0    1        0
#> 44  204    0    1        0
#> 45  204    0    1        0
#> 46  204    0    1        0
#> 47  204    0    1        0
#> 48  204    0    1        0
#> 49  205    0    1        0
#> 50  205    0    1        0
#> 51  205    0    1        0
#> 52  205    0    1        0
#> 53  205    0    1        0
#> 55  205    0    1        1
#> 56  205    0    1        0
#> 57  205    0    1        0
#> 58  205    0    1        0
#> 60  205    0    1        0
#> 62  206    0    1        0
#> 63  206    0    1        0
#> 64  206    0    1        0
#> 67  206    0    1        0
#> 72  206    0    1        0
#> 74  207    0    1        0
#> 76  207    0    1        0
#> 78  207    0    1        0
#> 79  207    0    1        0
#> 80  207    0    1        0
#> 82  207    0    1        0
#> 83  207    0    1        0
#> 85  208    0    1        0
#> 86  208    0    1        0
#> 88  208    0    1        0
#> 89  208    0    1        0
#> 91  208    0    1        0
#> 93  208    0    1        0
#> 94  208    0    1        0
#> 96  208    0    1        0
#> 97  209    0    1        0
#> 98  209    0    1        0
#> 99  209    0    1        0
#> 102 209    0    1        0
#> 109 210    0    1        0
#> 110 210    0    1        0
#> 111 210    0    1        0
#> 113 210    0    1        0
#> 115 210    0    1        0
#> 117 210    0    1        0
#> 122 211    0    1        0
#> 124 211    0    1        0
#> 125 211    0    1        0
#> 128 211    0    1        0
#> 129 211    0    1        0
#> 130 211    0    1        0
#> 131 211    0    1        0
#> 132 211    0    1        0
#> 133 212    0    1        0
#> 135 212    0    1        0
#> 137 212    0    1        0
#> 139 212    0    1        0
#> 140 212    0    1        0
#> 144 212    0    1        0
#> 146 213    0    1        0
#> 147 213    0    1        0
#> 148 213    0    1        0
#> 149 213    0    1        0
#> 153 213    0    1        0
#> 154 213    0    1        1
#> 155 213    0    1        0
#> 156 213    0    1        0
#> 157 214    0    1        0
#> 159 214    0    1        0
#> 162 214    0    1        0
#> 164 214    0    1        0
#> 166 214    0    1        0
#> 167 214    0    1        0
#> 168 214    0    1        0
#> 169 215    0    1        0
#> 170 215    0    1        0
#> 171 215    0    1        1
#> 175 215    0    1        0
#> 177 215    0    1        0
#> 178 215    0    1        0
#> 182 216    0    1        0
#> 184 216    0    1        0
#> 185 216    0    1        0
#> 186 216    0    1        0
#> 188 216    0    1        0
#> 190 216    0    1        0
#> 191 216    0    1        1
#> 192 216    0    1        0