create_op_wash_males

Create male older people indicators dataframe for water, sanitation and hygiene from survey data collected using the standard RAM-OP questionnaire

create_op_wash_males(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of male older people indicators on water, sanitation and hygiene

Examples

# Create water, sanitation and hygiene indicators dataset from RAM-OP survey
# data collected from Addis Ababa, Ethiopia
create_op_wash_males(testSVY)

#>     psu sex1 sex2 W1 W2 W3 W4
#> 2   201    1    0  1  1  0  0
#> 3   201    1    0  0  0  0  0
#> 6   201    1    0  0  0  0  0
#> 7   201    1    0  0  0  0  0
#> 9   201    1    0  1  1  0  0
#> 13  202    1    0  0  0  0  0
#> 14  202    1    0  1  1  1  1
#> 19  202    1    0  0  0  0  0
#> 20  202    1    0  0  1  0  0
#> 21  202    1    0  0  0  0  0
#> 27  203    1    0  1  1  1  1
#> 33  203    1    0  1  1  0  0
#> 34  203    1    0  1  1  0  0
#> 35  203    1    0  0  0  0  0
#> 36  203    1    0  0  0  0  0
#> 40  204    1    0  1  1  0  0
#> 42  204    1    0  1  1  0  0
#> 43  204    1    0  1  1  1  1
#> 54  205    1    0  1  1  0  0
#> 59  205    1    0  0  0  0  0
#> 61  206    1    0  0  0  0  0
#> 65  206    1    0  1  1  1  1
#> 66  206    1    0  1  1  1  1
#> 68  206    1    0  1  1  1  1
#> 69  206    1    0  0  0  0  0
#> 70  206    1    0  1  1  0  0
#> 71  206    1    0  0  0  0  0
#> 73  207    1    0  0  0  0  0
#> 75  207    1    0  0  0  0  0
#> 77  207    1    0  1  1  0  0
#> 81  207    1    0  0  1  0  0
#> 84  207    1    0  1  1  0  0
#> 87  208    1    0  0  1  0  0
#> 90  208    1    0  0  0  1  1
#> 92  208    1    0  1  1  0  0
#> 95  208    1    0  0  0  0  0
#> 100 209    1    0  1  1  0  0
#> 101 209    1    0  0  0  0  0
#> 103 209    1    0  1  1  0  0
#> 104 209    1    0  0  0  0  0
#> 105 209    1    0  0  0  0  0
#> 106 209    1    0  1  1  0  0
#> 107 209    1    0  1  1  1  1
#> 108 209    1    0  1  1  0  0
#> 112 210    1    0  1  1  1  1
#> 114 210    1    0  1  1  1  1
#> 116 210    1    0  1  1  0  0
#> 118 210    1    0  1  1  1  1
#> 119 210    1    0  0  0  0  0
#> 120 210    1    0  0  0  0  0
#> 121 211    1    0  0  0  0  0
#> 123 211    1    0  1  1  0  0
#> 126 211    1    0  1  1  1  1
#> 127 211    1    0  1  1  0  0
#> 134 212    1    0  1  1  1  1
#> 136 212    1    0  1  1  1  1
#> 138 212    1    0  0  0  0  0
#> 141 212    1    0  1  1  1  1
#> 142 212    1    0  0  0  0  0
#> 143 212    1    0  0  0  0  0
#> 145 213    1    0  1  1  1  0
#> 150 213    1    0  0  1  0  0
#> 151 213    1    0  1  1  0  0
#> 152 213    1    0  1  1  1  1
#> 158 214    1    0  1  1  0  0
#> 160 214    1    0  1  1  0  0
#> 161 214    1    0  1  1  1  1
#> 163 214    1    0  0  1  0  0
#> 165 214    1    0  1  1  1  1
#> 172 215    1    0  1  1  0  0
#> 173 215    1    0  0  0  0  0
#> 174 215    1    0  0  0  0  0
#> 176 215    1    0  1  1  1  1
#> 179 215    1    0  1  1  0  0
#> 180 215    1    0  1  1  1  1
#> 181 216    1    0  1  1  0  0
#> 183 216    1    0  1  1  0  0
#> 187 216    1    0  1  1  1  1
#> 189 216    1    0  1  1  1  1