estimateClassic

Function to apply bootstrap to RAM-OP indicators using a classical estimator.

estimateClassic(x, w, statistic = bootClassic, indicators = c("demo",
  "food", "hunger", "adl", "disability", "mental", "dementia", "health",
  "oedema", "screening", "income", "wash", "visual", "misc"),
  params = get_variables(indicators), outputColumns = params,
  replicates = 399)

Arguments

x
Indicators dataset produced by createOP with primary sampling unit (PSU) in column named PSU

w
A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

statistic
A function operating on data in x; fixed to bootClassic function for means

indicators
A character vector of indicator set names to estimate. Indicator set names are demo, food, hunger, disability, adl, mental, dementia, health, income, wash, visual, and misc. Default is all indicator sets.

params
Parameters (named columns in x) passed to the function specified in statistic. This is equivalent to variables corresponding to the indicator sets specified in indicators. The function get_variables() is used to specify these variables.

outputColumns
Names of columns in output data frame. This defaults to values specified in params

replicates
Number of bootstrap replicates

Value

Dataframe of boot estimates using bootClassic mean function

Examples

  #
  test <- estimateClassic(x = indicators.ALL,
                          w = testPSU,
                          replicates = 9)

  test

#>          LCL.ALL      EST.ALL     UCL.ALL    LCL.MALES   EST.MALES   UCL.MALES
#> 1    0.756250000  0.869791667  0.88437500  0.810710808  0.86250000  0.90313848
#> 2    0.073958333  0.104166667  0.17083333  0.032253086  0.08108108  0.11433426
#> 3    0.017708333  0.031250000  0.05312500  0.014457831  0.03846154  0.08698475
#> 4    0.000000000  0.005208333  0.04062500  0.000000000  0.01351351  0.03475783
#> 5   70.128125000 70.968750000 72.27291667 69.161535029 71.26923077 72.53047311
#> 6    0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 7    0.451041667  0.526041667  0.54479167  0.436731542  0.48750000  0.59773910
#> 8    0.211458333  0.244791667  0.28437500  0.171455051  0.25301205  0.39036729
#> 9    0.141666667  0.213541667  0.26354167  0.101424501  0.17073171  0.26076614
#> 10   0.009375000  0.046875000  0.06666667  0.014695820  0.04819277  0.12250695
#> 11   0.322916667  0.385416667  0.48854167  1.000000000  1.00000000  1.00000000
#> 12   0.511458333  0.614583333  0.67708333  0.000000000  0.00000000  0.00000000
#> 13   0.002083333  0.046875000  0.05520833  0.000000000  0.01282051  0.03578515
#> 14   0.209375000  0.296875000  0.35416667  0.518255300  0.60256410  0.67445394
#> 15   0.077083333  0.109375000  0.16458333  0.060537836  0.14457831  0.18999073
#> 16   0.025000000  0.067708333  0.08229167  0.053846154  0.08433735  0.12006098
#> 17   0.457291667  0.489583333  0.57083333  0.106039373  0.15662651  0.23379673
#> 18   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 19   0.100000000  0.135416667  0.18333333  0.063247863  0.15384615  0.25800764
#> 20   2.429166667  2.567708333  2.84375000  2.277804878  2.54054054  2.66734826
#> 21   4.313541667  4.593750000  4.72812500  4.184615385  4.41463415  4.64550120
#> 22   0.875000000  0.921875000  0.94166667  0.854683544  0.90361446  0.96873982
#> 23   0.362500000  0.500000000  0.58750000  0.399876429  0.42682927  0.62928498
#> 24   0.495833333  0.552083333  0.62291667  0.475859912  0.56626506  0.62428174
#> 25   0.023958333  0.052083333  0.10208333  0.000000000  0.02564103  0.06795791
#> 26   0.007291667  0.031250000  0.05937500  0.012320200  0.04054054  0.08960843
#> 27   0.253125000  0.328125000  0.39791667  0.372435897  0.43373494  0.48088283
#> 28   0.357291667  0.427083333  0.49166667  0.334136546  0.39759036  0.49148304
#> 29   0.011458333  0.026041667  0.06875000  0.000000000  0.00000000  0.02321552
#> 30   0.182291667  0.229166667  0.25520833  0.139759036  0.22222222  0.31041529
#> 31   0.448958333  0.515625000  0.60104167  0.345212593  0.40243902  0.45433735
#> 32   0.958333333  0.958333333  0.98958333  0.926115859  0.97297297  0.99759036
#> 33   0.397916667  0.500000000  0.56041667  0.399876429  0.42168675  0.54186592
#> 34   0.357291667  0.427083333  0.49166667  0.334136546  0.39759036  0.49148304
#> 35   0.084375000  0.104166667  0.17291667  0.017448405  0.08641975  0.11663871
#> 36   0.527083333  0.593750000  0.65625000  0.501265823  0.56250000  0.63330244
#> 37   0.042708333  0.057291667  0.08229167  0.014884303  0.05063291  0.08960843
#> 38   0.555208333  0.625000000  0.69479167  0.524640400  0.60000000  0.66104418
#> 39   0.603125000  0.687500000  0.71770833  0.562820513  0.63291139  0.73154155
#> 40   0.011458333  0.026041667  0.06875000  0.000000000  0.00000000  0.02321552
#> 41   0.567708333  0.640625000  0.67291667  0.551251158  0.66666667  0.76162396
#> 42   0.628125000  0.687500000  0.72187500  0.574173618  0.67901235  0.77175054
#> 43   0.795833333  0.828125000  0.85208333  0.691737892  0.80722892  0.88060212
#> 44   0.567708333  0.640625000  0.67291667  0.551251158  0.66666667  0.76162396
#> 45   0.840625000  0.859375000  0.89375000  0.802162496  0.89873418  0.93461538
#> 46   0.338541667  0.380208333  0.45729167  0.400448718  0.46913580  0.52788056
#> 47   0.326041667  0.375000000  0.44479167  0.400448718  0.45945946  0.52411164
#> 48   0.687500000  0.791666667  0.88437500  0.705874716  0.72289157  0.82091463
#> 49   0.100000000  0.156250000  0.25104167  0.179085366  0.27027027  0.29412528
#> 50   0.015625000  0.031250000  0.06145833  0.000000000  0.00000000  0.02891566
#> 51   0.916666667  0.968750000  0.98437500  0.896229563  0.95121951  0.97198718
#> 52   0.951041667  0.973958333  0.99791667  0.922789456  0.96385542  0.99750000
#> 53   0.951041667  0.973958333  0.99791667  0.922789456  0.96385542  0.99750000
#> 54   0.907291667  0.947916667  0.97395833  0.903972647  0.96153846  0.99500000
#> 55   0.697916667  0.713541667  0.77812500  0.658016682  0.71951220  0.79547468
#> 56   0.984375000  0.994791667  0.99895833  0.925496581  0.98717949  1.00000000
#> 57   5.488541667  5.572916667  5.66041667  5.305307717  5.55421687  5.70458861
#> 58   0.948958333  0.958333333  0.98437500  0.922789456  0.96385542  0.99750000
#> 59   0.002083333  0.015625000  0.03333333  0.000000000  0.00000000  0.00000000
#> 60   0.001041667  0.015625000  0.03541667  0.002500000  0.03614458  0.07721054
#> 61   0.511458333  0.578125000  0.61354167  0.377423390  0.52702703  0.60463627
#> 62   0.090625000  0.125000000  0.16666667  0.103512658  0.20512821  0.28773556
#> 63   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 64   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 65   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 66   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 67   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 68   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 69   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 70   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 71   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 72   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 73   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 74   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 75   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 76   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 77   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 78   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 79   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 80   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 81   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 82   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 83   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 84   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 85   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 86   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 87   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 88   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 89   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 90   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 91   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 92  11.505208333 12.338541667 13.24895833  9.973219373 11.21794872 12.21871622
#> 93   0.442708333  0.515625000  0.56145833  0.356441149  0.44578313  0.53993243
#> 94   0.162500000  0.197916667  0.29479167  0.061601001  0.13924051  0.20132530
#> 95   0.409375000  0.447916667  0.48437500  0.290784761  0.38750000  0.46729206
#> 96   0.727560614  0.804597701  0.87440530  0.524014337  0.69444444  0.86874459
#> 97   0.011764706  0.166666667  0.41923077  0.000000000  0.22222222  0.32380952
#> 98   0.309790210  0.437500000  0.66769231  0.125252525  0.33333333  0.50000000
#> 99   0.011111111  0.100000000  0.27165775  0.000000000  0.00000000  0.00000000
#> 100  0.012500000  0.100000000  0.39893048  0.183333333  0.33333333  0.62000000
#> 101  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 102  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 103  0.000000000  0.000000000  0.06176471  0.000000000  0.00000000  0.00000000
#> 104  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 105  0.000000000  0.111111111  0.22459893  0.000000000  0.14285714  0.31666667
#> 106  0.808333333  0.875000000  0.89583333  0.782721128  0.87179487  0.90651838
#> 107  0.794003923  0.857142857  0.88391538  0.642669008  0.73913043  0.81883446
#> 108  0.008333333  0.157894737  0.28375000  0.000000000  0.13333333  0.34298643
#> 109  0.531250000  0.750000000  0.90000000  0.621719457  0.76923077  0.92666667
#> 110  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 111  0.000000000  0.041666667  0.12500000  0.008000000  0.10000000  0.17450980
#> 112  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 113  0.000000000  0.055555556  0.14500000  0.000000000  0.00000000  0.00000000
#> 114  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 115  0.000000000  0.000000000  0.02857143  0.000000000  0.00000000  0.00000000
#> 116  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 117  0.464583333  0.531250000  0.59687500  0.599316184  0.69230769  0.71795593
#> 118  0.296875000  0.338541667  0.41458333  0.422710843  0.53164557  0.60708017
#> 119  0.084375000  0.114583333  0.18333333  0.171525125  0.23076923  0.35452414
#> 120  0.005208333  0.015625000  0.02916667  0.002702703  0.06024096  0.11089233
#> 121  0.025000000  0.052083333  0.09166667  0.002469136  0.02702703  0.04852564
#> 122  0.000000000  0.005208333  0.01562500  0.000000000  0.00000000  0.00000000
#> 123  0.001041667  0.015625000  0.02604167  0.000000000  0.03797468  0.08600327
#> 124  0.011458333  0.031250000  0.05937500  0.002409639  0.02531646  0.05078174
#> 125  0.254166667  0.328125000  0.36354167  0.185565052  0.27710843  0.38814449
#> 126  0.001041667  0.005208333  0.01979167  0.000000000  0.00000000  0.02507244
#> 127  0.569791667  0.588541667  0.67708333  0.483117279  0.58750000  0.66504065
#> 128  0.672916667  0.692708333  0.77187500  0.523135580  0.63750000  0.73633521
#> 129  0.182291667  0.260416667  0.33854167  0.183037975  0.28395062  0.35180723
#> 130  0.168750000  0.255208333  0.32187500  0.183037975  0.28395062  0.35180723
#> 131  0.006250000  0.036458333  0.04166667  0.002500000  0.02409639  0.04719848
#> 132  0.015625000  0.026041667  0.07812500  0.002500000  0.06024096  0.07967968
#> 133  0.382291667  0.411458333  0.45208333  0.387702703  0.46913580  0.54704974
#> 134  0.266666667  0.286458333  0.33229167  0.177125906  0.25641026  0.29919170
#> 135  0.014583333  0.046875000  0.07187500  0.012202657  0.03750000  0.08285449
#> 136  0.016666667  0.020833333  0.04479167  0.000000000  0.01250000  0.05902127
#>      LCL.FEMALES  EST.FEMALES UCL.FEMALES         INDICATOR
#> 1    0.818646232  0.889830508  0.92483705             resp1
#> 2    0.056644494  0.093220339  0.14821201             resp2
#> 3    0.010286382  0.018181818  0.03936081             resp3
#> 4    0.000000000  0.000000000  0.01564496             resp4
#> 5   68.960501567 70.366071429 72.54332635               age
#> 6    0.000000000  0.000000000  0.00000000           ageGrp1
#> 7    0.462962963  0.568965517  0.63724138           ageGrp2
#> 8    0.149470899  0.218181818  0.27113821           ageGrp3
#> 9    0.146383516  0.189655172  0.33983229           ageGrp4
#> 10   0.000000000  0.009433962  0.05720443           ageGrp5
#> 11   0.000000000  0.000000000  0.00000000              sex1
#> 12   1.000000000  1.000000000  1.00000000              sex2
#> 13   0.008565375  0.043103448  0.06202020          marital1
#> 14   0.058045977  0.152542373  0.23429467          marital2
#> 15   0.046982759  0.100000000  0.12979831          marital3
#> 16   0.012861953  0.043103448  0.06818054          marital4
#> 17   0.577324185  0.672413793  0.74827586          marital5
#> 18   0.000000000  0.000000000  0.00000000          marital6
#> 19   0.047118644  0.133928571  0.16066411             alone
#> 20   2.486706349  2.689655172  2.77061345                MF
#> 21   4.528097063  4.660377358  4.87256442               DDS
#> 22   0.862068966  0.892857143  0.96061427              FG01
#> 23   0.528006728  0.571428571  0.64617183              FG02
#> 24   0.566021867  0.666666667  0.74840599              FG03
#> 25   0.049106003  0.081818182  0.12433642              FG04
#> 26   0.000000000  0.017857143  0.06837996              FG05
#> 27   0.206599327  0.250000000  0.30853659              FG06
#> 28   0.335956113  0.431034483  0.48051584              FG07
#> 29   0.019904682  0.050847458  0.07305236              FG08
#> 30   0.158465608  0.232758621  0.29301934              FG09
#> 31   0.430694070  0.509259259  0.63266814              FG10
#> 32   0.936283741  0.982758621  0.99837398              FG11
#> 33   0.425924765  0.520325203  0.56014027       proteinRich
#> 34   0.335956113  0.431034483  0.48051584          pProtein
#> 35   0.120686540  0.142857143  0.22251139          aProtein
#> 36   0.608213166  0.678571429  0.77068966             pVitA
#> 37   0.046533128  0.071428571  0.11125569             aVitA
#> 38   0.625642633  0.691056911  0.81206897             xVitA
#> 39   0.623569024  0.715447154  0.81206897          ironRich
#> 40   0.019904682  0.050847458  0.07305236            caRich
#> 41   0.481548822  0.551724138  0.61455811            znRich
#> 42   0.557171717  0.612068966  0.72126223             vitB1
#> 43   0.726734007  0.829268293  0.92007389             vitB2
#> 44   0.481548822  0.551724138  0.61455811             vitB3
#> 45   0.797542088  0.864406780  0.94162562             vitB6
#> 46   0.256363636  0.357723577  0.39225182            vitB12
#> 47   0.247138047  0.349593496  0.38171913       vitBcomplex
#> 48   0.687110126  0.810344828  0.86798030              HHS1
#> 49   0.104187192  0.129310345  0.19739837              HHS2
#> 50   0.005555556  0.037735849  0.12436221              HHS3
#> 51   0.974867725  0.991379310  1.00000000             ADL01
#> 52   0.991133005  1.000000000  1.00000000             ADL02
#> 53   0.991133005  1.000000000  1.00000000             ADL03
#> 54   0.917917676  0.963636364  0.98985142             ADL04
#> 55   0.675714286  0.715517241  0.79986988             ADL05
#> 56   1.000000000  1.000000000  1.00000000             ADL06
#> 57   5.564891041  5.689655172  5.78353936          scoreADL
#> 58   0.955811138  0.991379310  1.00000000         classADL1
#> 59   0.000000000  0.008620690  0.04418886         classADL2
#> 60   0.000000000  0.000000000  0.00000000         classADL3
#> 61   0.537931034  0.627272727  0.71761518           hasHelp
#> 62   0.043631436  0.098214286  0.11147335         unmetNeed
#> 63   1.000000000  1.000000000  1.00000000        wgVisionD0
#> 64   0.000000000  0.000000000  0.00000000        wgVisionD1
#> 65   0.000000000  0.000000000  0.00000000        wgVisionD2
#> 66   0.000000000  0.000000000  0.00000000        wgVisionD3
#> 67   1.000000000  1.000000000  1.00000000       wgHearingD0
#> 68   0.000000000  0.000000000  0.00000000       wgHearingD1
#> 69   0.000000000  0.000000000  0.00000000       wgHearingD2
#> 70   0.000000000  0.000000000  0.00000000       wgHearingD3
#> 71   1.000000000  1.000000000  1.00000000      wgMobilityD0
#> 72   0.000000000  0.000000000  0.00000000      wgMobilityD1
#> 73   0.000000000  0.000000000  0.00000000      wgMobilityD2
#> 74   0.000000000  0.000000000  0.00000000      wgMobilityD3
#> 75   1.000000000  1.000000000  1.00000000   wgRememberingD0
#> 76   0.000000000  0.000000000  0.00000000   wgRememberingD1
#> 77   0.000000000  0.000000000  0.00000000   wgRememberingD2
#> 78   0.000000000  0.000000000  0.00000000   wgRememberingD3
#> 79   1.000000000  1.000000000  1.00000000      wgSelfCareD0
#> 80   0.000000000  0.000000000  0.00000000      wgSelfCareD1
#> 81   0.000000000  0.000000000  0.00000000      wgSelfCareD2
#> 82   0.000000000  0.000000000  0.00000000      wgSelfCareD3
#> 83   1.000000000  1.000000000  1.00000000 wgCommunicatingD0
#> 84   0.000000000  0.000000000  0.00000000 wgCommunicatingD1
#> 85   0.000000000  0.000000000  0.00000000 wgCommunicatingD2
#> 86   0.000000000  0.000000000  0.00000000 wgCommunicatingD3
#> 87   1.000000000  1.000000000  1.00000000              wgP0
#> 88   0.000000000  0.000000000  0.00000000              wgP1
#> 89   0.000000000  0.000000000  0.00000000              wgP2
#> 90   0.000000000  0.000000000  0.00000000              wgP3
#> 91   0.000000000  0.000000000  0.00000000              wgPM
#> 92  11.901974448 12.474137931 13.46896552                K6
#> 93   0.483181522  0.517241379  0.55952978            K6Case
#> 94   0.176112853  0.237288136  0.25646401                DS
#> 95   0.361894673  0.445454545  0.55898826                H1
#> 96   0.697568389  0.780000000  0.87487685                H2
#> 97   0.000000000  0.111111111  0.49350649               H31
#> 98   0.277005348  0.428571429  0.77222222               H32
#> 99   0.000000000  0.181818182  0.38655462               H33
#> 100  0.000000000  0.000000000  0.00000000               H34
#> 101  0.000000000  0.000000000  0.00000000               H35
#> 102  0.000000000  0.000000000  0.00000000               H36
#> 103  0.000000000  0.000000000  0.12222222               H37
#> 104  0.000000000  0.000000000  0.00000000               H38
#> 105  0.000000000  0.071428571  0.37714286               H39
#> 106  0.810689655  0.896551724  0.92539199                H4
#> 107  0.789542880  0.854166667  0.90007974                H5
#> 108  0.000000000  0.071428571  0.36210526               H61
#> 109  0.392982456  0.818181818  0.92857143               H62
#> 110  0.000000000  0.000000000  0.00000000               H63
#> 111  0.000000000  0.000000000  0.00000000               H64
#> 112  0.000000000  0.000000000  0.00000000               H65
#> 113  0.000000000  0.100000000  0.24862745               H66
#> 114  0.000000000  0.000000000  0.00000000               H67
#> 115  0.000000000  0.000000000  0.11229947               H68
#> 116  0.000000000  0.000000000  0.00000000               H69
#> 117  0.415040650  0.500000000  0.54153605                M1
#> 118  0.229058032  0.324074074  0.38815047               M2A
#> 119  0.001851852  0.034482759  0.11812954               M2B
#> 120  0.000000000  0.009090909  0.01767554               M2C
#> 121  0.037759711  0.103448276  0.11976601               M2D
#> 122  0.000000000  0.008620690  0.03708523               M2E
#> 123  0.000000000  0.000000000  0.00000000               M2F
#> 124  0.000000000  0.009259259  0.04153605               M2G
#> 125  0.284993614  0.308943089  0.37758621               M2H
#> 126  0.000000000  0.000000000  0.01561129               M2I
#> 127  0.512452107  0.601694915  0.62890282                W1
#> 128  0.638646714  0.709090909  0.77190917                W2
#> 129  0.150574713  0.250000000  0.29969125                W3
#> 130  0.146871009  0.241379310  0.29085882                W4
#> 131  0.016579559  0.018867925  0.05517241            oedema
#> 132  0.012629870  0.033898305  0.06379310          screened
#> 133  0.227333894  0.327586207  0.38255406            poorVA
#> 134  0.228090833  0.327272727  0.38082992              chew
#> 135  0.018965517  0.056603774  0.07126386              food
#> 136  0.003389831  0.018867925  0.06944568              NFRI