estimateProbit

estimateProbit

estimateProbit(x, w, gam.stat = probit_gam, sam.stat = probit_sam,
  params = "MUAC", outputColumns = "MUAC", replicates = 399)

Arguments

x
Indicators dataset produced by createOP with primary sampling unit (PSU) in column named PSU

w
A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

gam.stat
A function operating on data in x to estimate GAM prevalence for older people. Fixed to probit_gam

sam.stat
A function operating on data in x to estimate SAM prevalence for older people. Fixed to probit_sam

params
Parameters (named columns in x) passed to the function specified in statistic; fixed to MUAC as indicator amenable to probit estimation

outputColumns
Names of columns in output data frame; fixed to MUAC

replicates
Number of bootstrap replicate case and non-case

Value

Dataframe of boot estimates using bootPROBIT function

Examples

  #
  test <- estimateProbit(x = indicators.ALL,
                         w = testPSU,
                         replicates = 9)

  test

#>        LCL.ALL      EST.ALL      UCL.ALL    LCL.MALES    EST.MALES   UCL.MALES
#> 1 4.733018e-03 0.0355276727 0.0675348883 1.495070e-03 9.078120e-03 0.021827037
#> 2 4.317994e-03 0.0349068520 0.0672352532 1.495070e-03 8.541328e-03 0.021801438
#> 3 4.642083e-08 0.0002966476 0.0006004115 5.700644e-18 1.674605e-05 0.001049068
#>    LCL.FEMALES EST.FEMALES UCL.FEMALES INDICATOR
#> 1 1.329008e-02 0.029063181 0.054793622       GAM
#> 2 1.294306e-02 0.025959765 0.053405073       MAM
#> 3 6.009333e-06 0.001481778 0.004421152       SAM