Function to apply bootstrap to RAM-OP indicators using a classical estimator.

Function to apply bootstrap to RAM-OP indicators using a classical estimator.

Usage

estimate_classic(
  x,
  w,
  statistic = bbw::bootClassic,
  indicators = c("demo", "food", "hunger", "adl", "disability", "mental", "dementia",
    "health", "oedema", "screening", "income", "wash", "visual", "misc"),
  params = get_variables(indicators),
  outputColumns = params,
  replicates = 399
)

Arguments

x

Indicators dataset produced by create_op_all with primary sampling unit (PSU) in column named PSU

w

A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

statistic

A function operating on data in x; fixed to bootClassic function for means

indicators

A character vector of indicator set names to estimate. Indicator set names are demo, food, hunger, disability, adl, mental, dementia, health, income, wash, visual, and misc. Default is all indicator sets.

params

Parameters (named columns in x) passed to the function specified in statistic. This is equivalent to variables corresponding to the indicator sets specified in indicators. The function get_variables is used to specify these variables.

outputColumns

Names of columns in output data frame. This defaults to values specified in params

replicates

Number of bootstrap replicates

Value

Tibble of boot estimates using bootClassic mean function

Examples

  #
  test <- estimate_classic(x = indicators.ALL,
                           w = testPSU,
                           replicates = 9)

  test
#> # A tibble: 136 × 10
#>    INDICATOR EST.ALL  LCL.ALL UCL.ALL EST.MALES LCL.MA…¹ UCL.M…² EST.F…³ LCL.F…⁴
#>    <chr>       <dbl>    <dbl>   <dbl>     <dbl>    <dbl>   <dbl>   <dbl>   <dbl>
#>  1 resp1      0.828   0.788    0.869     0.853    0.803   0.902  8.53e-1  0.821 
#>  2 resp2      0.120   0.0823   0.155     0.0641   0.0306  0.0900 1.09e-1  0.0680
#>  3 resp3      0.0365  0.0208   0.0708    0.0533   0.0291  0.0941 4.13e-2  0     
#>  4 resp4      0.0156  0.00104  0.0281    0.0260   0       0.0616 0        0     
#>  5 age       71.1    69.5     72.1      71.6     69.3    72.2    7.02e+1 69.0   
#>  6 ageGrp1    0       0        0         0        0       0      0        0     
#>  7 ageGrp2    0.542   0.462    0.620     0.506    0.401   0.593  5.77e-1  0.505 
#>  8 ageGrp3    0.219   0.167    0.240     0.247    0.159   0.400  1.83e-1  0.0959
#>  9 ageGrp4    0.240   0.115    0.292     0.154    0.125   0.277  2.19e-1  0.172 
#> 10 ageGrp5    0.0365  0.0229   0.0583    0.0533   0.0143  0.127  9.90e-3  0     
#> # … with 126 more rows, 1 more variable: UCL.FEMALES <dbl>, and abbreviated
#> #   variable names ¹​LCL.MALES, ²​UCL.MALES, ³​EST.FEMALES, ⁴​LCL.FEMALES