Estimate all standard RAM-OP indicators

Estimate all standard RAM-OP indicators

estimate_op_all(
  x,
  w,
  indicators = c("demo", "anthro", "food", "hunger", "adl", "disability", "mental",
    "dementia", "health", "oedema", "screening", "income", "wash", "visual", "misc"),
  replicates = 399
)

Arguments

x
Indicators dataset produced by create_op_all with primary sampling unit (PSU) in column named PSU

w
A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

indicators
A character vector of indicator set names to estimate. Indicator set names are demo, anthro, food, hunger, disability, adl, mental, dementia, health, income, wash, visual, and misc. Default is all indicator sets.

replicates
Number of bootstrap replicates. Default is 399.

Value

Tibble of boot estimates for all specified standard RAM-OP indicators

Examples

estimate_op_all(x = create_op_all(testSVY),
                w = testPSU,
                replicates = 9)

#> # A tibble: 139 x 13
#>    INDICATOR LCL.ALL EST.ALL UCL.ALL LCL.MALES EST.MALES UCL.MALES LCL.FEMALES
#>    <chr>       <dbl>   <dbl>   <dbl>     <dbl>     <dbl>     <dbl>       <dbl>
#>  1 resp1     8.15e-1  0.870   0.917     0.758     0.84      0.920      0.754  
#>  2 resp2     5.83e-2  0.0781  0.109     0.0248    0.0789    0.155      0.0704 
#>  3 resp3     7.29e-3  0.0365  0.0667    0.0155    0.0633    0.125      0.00859
#>  4 resp4     0.       0.0104  0.0208    0         0.0127    0.0506     0      
#>  5 age       6.90e+1 70.2    71.7      69.1      70.2      73.2       70.1    
#>  6 ageGrp1   0.       0       0         0         0         0          0      
#>  7 ageGrp2   4.80e-1  0.542   0.622     0.396     0.5       0.606      0.404  
#>  8 ageGrp3   1.95e-1  0.224   0.333     0.172     0.271     0.392      0.137  
#>  9 ageGrp4   1.15e-1  0.188   0.261     0.0686    0.184     0.263      0.181  
#> 10 ageGrp5   7.29e-3  0.0208  0.0406    0.0146    0.0533    0.0899     0      
#> # … with 129 more rows, and 5 more variables: EST.FEMALES <dbl>,
#> #   UCL.FEMALES <dbl>, GROUP <fct>, LABEL <fct>, TYPE <fct>