Function to apply bootstrap to RAM-OP indicators using a PROBIT estimator.

Function to apply bootstrap to RAM-OP indicators using a PROBIT estimator.

Usage

estimate_probit(
  x,
  w,
  gam.stat = probit_gam,
  sam.stat = probit_sam,
  params = "MUAC",
  outputColumns = "MUAC",
  replicates = 399
)

Arguments

x

Indicators dataset produced by create_op_all with primary sampling unit (PSU) in column named PSU

w

A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

gam.stat

A function operating on data in x to estimate GAM prevalence for older people. Fixed to probit_gam

sam.stat

A function operating on data in x to estimate SAM prevalence for older people. Fixed to probit_sam

params

Parameters (named columns in x) passed to the function specified in statistic; fixed to MUAC as indicator amenable to probit estimation

outputColumns

Names of columns in output data frame; fixed to MUAC

replicates

Number of bootstrap replicate case and non-case

Value

Dataframe of boot estimates using bootPROBIT function

Examples

  #
  test <- estimate_probit(x = indicators.ALL,
                          w = testPSU,
                          replicates = 3)

  test
#> # A tibble: 3 × 10
#>   INDICATOR   EST.ALL LCL.ALL UCL.ALL EST.MA…¹ LCL.MA…² UCL.MA…³ EST.F…⁴ LCL.F…⁵
#>   <chr>         <dbl>   <dbl>   <dbl>    <dbl>    <dbl>    <dbl>   <dbl>   <dbl>
#> 1 GAM         3.91e-2 2.77e-3 5.97e-2 2.24e- 3 1.16e- 4 2.44e- 3 3.74e-2 5.58e-3
#> 2 MAM         3.91e-2 2.77e-3 5.96e-2 2.24e- 3 1.16e- 4 2.44e- 3 3.63e-2 5.52e-3
#> 3 SAM         1.18e-6 6.16e-7 1.89e-4 3.91e-16 1.96e-17 4.70e-14 5.73e-4 2.87e-5
#> # … with 1 more variable: UCL.FEMALES <dbl>, and abbreviated variable names
#> #   ¹​EST.MALES, ²​LCL.MALES, ³​UCL.MALES, ⁴​EST.FEMALES, ⁵​LCL.FEMALES