get_variables

Function to get appropriate RAM-OP indicator variable names given a specified indicator set

get_variables(
  indicators = c("demo", "food", "hunger", "adl", "disability", "mental", "dementia",
    "health", "income", "wash", "anthro", "oedema", "screening", "visual", "misc")
)

Arguments

indicators
A character vector of indicator set names. Indicator set names are demo, food, hunger, disability, adl, mental, dementia, health, income, wash, anthro, screening, visual, and misc. Default is all indicator sets.

Value

A vector of variable names

Examples

get_variables(indicators = c("demo", "food"))

#>  [1] "resp1"       "resp2"       "resp3"       "resp4"       "age"        
#>  [6] "ageGrp1"     "ageGrp2"     "ageGrp3"     "ageGrp4"     "ageGrp5"    
#> [11] "sex1"        "sex2"        "marital1"    "marital2"    "marital3"   
#> [16] "marital4"    "marital5"    "marital6"    "alone"       "MF"         
#> [21] "DDS"         "FG01"        "FG02"        "FG03"        "FG04"       
#> [26] "FG05"        "FG06"        "FG07"        "FG08"        "FG09"       
#> [31] "FG10"        "FG11"        "proteinRich" "pProtein"    "aProtein"   
#> [36] "pVitA"       "aVitA"       "xVitA"       "ironRich"    "caRich"     
#> [41] "znRich"      "vitB1"       "vitB2"       "vitB3"       "vitB6"      
#> [46] "vitB12"      "vitBcomplex"