create_op_adl

Create older people indicators dataframe on acitvities of daily living from survey data collected using the standard RAM-OP questionnaire

create_op_adl(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of older people indicators on activities of daily living

Indicators

Katz "Index of Independence in Activities of Daily Living" (ADL) score

The Katz ADL score is described in:

Katz S, Ford AB, Moskowitz RW, Jackson BA, Jaffe MW (1963). Studies of illness in the aged. The Index of ADL: a standardized measure of biological and psychosocial function. JAMA, 1963, 185(12):914-9 doi:10.1001/jama.1963.03060120024016

Katz S, Down TD, Cash HR, Grotz, RC (1970). Progress in the development of the index of ADL. The Gerontologist, 10(1), 20-30 doi:10.1093/geront/10.4_Part_1.274

Katz S (1983). Assessing self-maintenance: Activities of daily living, mobility and instrumental activities of daily living. JAGS, 31(12), 721-726 doi:10.1111/j.1532-5415.1983.tb03391.x

ADL01

Bathing

ADL02

Dressing

ADL03

Toileting

ADL04

Transferring (mobility)

ADL05

Continence

ADL06

Feeding

scoreADL

ADL Score

classADL1

Severity of dependence 1

classADL2

Severity of dependence 2

classADL3

Severity of dependence 3

hasHelp

Have someone to help with everyday activities

unmetNeed

Need help but has no helper

Examples

# Create activities of daily living indicators dataset from RAM-OP survey
# data collected from Addis Ababa, Ethiopia
create_op_adl(testSVY)

#>     psu sex1 sex2 ADL01 ADL02 ADL03 ADL04 ADL05 ADL06 scoreADL classADL1
#> 1   201    0    1     1     1     1     1     1     1        6         1
#> 2   201    1    0     1     1     1     1     0     1        5         1
#> 3   201    1    0     1     1     1     1     1     1        6         1
#> 4   201    0    1     1     1     1     1     1     1        6         1
#> 5   201    0    1     1     1     1     1     0     1        5         1
#> 6   201    1    0     1     1     1     1     1     1        6         1
#> 7   201    1    0     1     1     1     1     1     1        6         1
#> 8   201    0    1     0     1     1     0     0     1        3         0
#> 9   201    1    0     1     1     1     1     0     1        5         1
#> 10  201    0    1     1     1     1     1     1     1        6         1
#> 11  201    0    1     1     1     1     1     1     1        6         1
#> 12  201    0    1     1     1     1     0     0     1        4         0
#> 13  202    1    0     1     1     1     1     1     1        6         1
#> 14  202    1    0     1     1     1     1     1     1        6         1
#> 15  202    0    1     1     1     1     1     1     1        6         1
#> 16  202    0    1     1     1     1     0     1     1        5         1
#> 17  202    0    1     1     1     1     1     1     1        6         1
#> 18  202    0    1     1     1     1     1     0     1        5         1
#> 19  202    1    0     1     1     1     0     1     1        5         1
#> 20  202    1    0     1     1     1     1     1     1        6         1
#> 21  202    1    0     1     1     1     1     1     1        6         1
#> 22  202    0    1     1     1     1     1     1     1        6         1
#> 23  202    0    1     1     1     1     1     1     1        6         1
#> 24  202    0    1     1     1     1     1     0     1        5         1
#> 25  203    0    1     1     1     1     1     1     1        6         1
#> 26  203    0    1     1     1     1     1     1     1        6         1
#> 27  203    1    0     1     1     1     1     1     1        6         1
#> 28  203    0    1     1     1     1     1     1     1        6         1
#> 29  203    0    1     1     1     1     1     1     1        6         1
#> 30  203    0    1     1     1     1     1     1     1        6         1
#> 31  203    0    1     1     1     1     1     1     1        6         1
#> 32  203    0    1     1     1     1     1     0     1        5         1
#> 33  203    1    0     1     1     1     1     1     1        6         1
#> 34  203    1    0     1     1     1     1     0     1        5         1
#> 35  203    1    0     1     1     1     1     1     1        6         1
#> 36  203    1    0     1     1     1     1     1     1        6         1
#> 37  204    0    1     1     1     1     1     0     1        5         1
#> 38  204    0    1     1     1     1     1     1     1        6         1
#> 39  204    0    1     0     0     0     1     1     1        3         0
#> 40  204    1    0     1     1     1     1     1     1        6         1
#> 41  204    0    1     0     1     1     1     1     1        5         1
#> 42  204    1    0     1     1     1     1     1     1        6         1
#> 43  204    1    0     1     1     1     1     0     1        5         1
#> 44  204    0    1     1     1     1     1     1     1        6         1
#> 45  204    0    1     1     1     1     1     1     1        6         1
#> 46  204    0    1     1     1     1     0     1     1        5         1
#> 47  204    0    1     1     1     1     0     1     1        5         1
#> 48  204    0    1     1     1     1     1     1     1        6         1
#> 49  205    0    1     1     1     1     1     1     1        6         1
#> 50  205    0    1     1     1     1     1     1     1        6         1
#> 51  205    0    1     1     1     1     1     1     1        6         1
#> 52  205    0    1     1     1     1     1     0     1        5         1
#> 53  205    0    1     1     1     1     1     1     1        6         1
#> 54  205    1    0     1     1     1     1     1     1        6         1
#> 55  205    0    1     1     1     1     1     1     1        6         1
#> 56  205    0    1     1     1     1     1     1     1        6         1
#> 57  205    0    1     1     1     1     1     1     1        6         1
#> 58  205    0    1     1     1     1     1     0     1        5         1
#> 59  205    1    0     1     1     1     1     0     1        5         1
#> 60  205    0    1     1     1     1     1     1     1        6         1
#> 61  206    1    0     1     1     1     1     1     1        6         1
#> 62  206    0    1     1     1     1     1     1     1        6         1
#> 63  206    0    1     1     1     1     1     1     1        6         1
#> 64  206    0    1     1     1     1     0     1     1        5         1
#> 65  206    1    0     1     1     1     1     1     1        6         1
#> 66  206    1    0     1     1     1     1     0     1        5         1
#> 67  206    0    1     1     1     1     1     1     1        6         1
#> 68  206    1    0     1     1     1     1     1     1        6         1
#> 69  206    1    0     1     1     1     1     1     1        6         1
#> 70  206    1    0     1     1     1     1     0     1        5         1
#> 71  206    1    0     1     1     1     1     1     1        6         1
#> 72  206    0    1     1     1     1     1     1     1        6         1
#> 73  207    1    0     1     1     1     1     1     1        6         1
#> 74  207    0    1     1     1     1     1     1     1        6         1
#> 75  207    1    0     1     1     1     1     1     1        6         1
#> 76  207    0    1     1     1     1     1     1     1        6         1
#> 77  207    1    0     1     1     1     1     1     1        6         1
#> 78  207    0    1     1     1     1     1     1     1        6         1
#> 79  207    0    1     1     1     1     1     1     1        6         1
#> 80  207    0    1     1     1     1     1     0     1        5         1
#> 81  207    1    0     1     1     1     1     1     1        6         1
#> 82  207    0    1     1     1     1     1     1     1        6         1
#> 83  207    0    1     1     1     1     1     1     1        6         1
#> 84  207    1    0     1     1     1     1     1     1        6         1
#> 85  208    0    1     1     1     1     1     1     1        6         1
#> 86  208    0    1     1     1     1     1     0     1        5         1
#> 87  208    1    0     1     1     1     1     1     1        6         1
#> 88  208    0    1     1     1     1     1     1     1        6         1
#> 89  208    0    1     1     1     1     1     0     1        5         1
#> 90  208    1    0     1     1     1     1     1     1        6         1
#> 91  208    0    1     1     1     1     1     1     1        6         1
#> 92  208    1    0     1     1     1     1     0     1        5         1
#> 93  208    0    1     1     1     1     1     0     1        5         1
#> 94  208    0    1     1     1     1     1     1     1        6         1
#> 95  208    1    0     0     1     1     1     1     1        5         1
#> 96  208    0    1     1     1     1     1     0     1        5         1
#> 97  209    0    1     1     1     1     1     0     1        5         1
#> 98  209    0    1     1     1     1     1     0     1        5         1
#> 99  209    0    1     1     1     1     1     1     1        6         1
#> 100 209    1    0     1     1     1     1     1     1        6         1
#> 101 209    1    0     1     1     1     1     1     1        6         1
#> 102 209    0    1     1     1     1     1     0     1        5         1
#> 103 209    1    0     1     1     1     1     1     1        6         1
#> 104 209    1    0     1     1     1     1     1     1        6         1
#> 105 209    1    0     1     1     1     1     1     1        6         1
#> 106 209    1    0     1     1     1     1     1     1        6         1
#> 107 209    1    0     1     1     1     1     0     1        5         1
#> 108 209    1    0     1     1     1     1     1     1        6         1
#> 109 210    0    1     1     1     1     1     1     1        6         1
#> 110 210    0    1     1     1     1     1     1     1        6         1
#> 111 210    0    1     1     1     1     1     1     1        6         1
#> 112 210    1    0     0     0     0     0     1     1        2         0
#> 113 210    0    1     1     1     1     1     1     1        6         1
#> 114 210    1    0     1     1     1     1     1     1        6         1
#> 115 210    0    1     1     1     1     1     1     1        6         1
#> 116 210    1    0     1     1     1     1     1     1        6         1
#> 117 210    0    1     1     1     1     1     0     1        5         1
#> 118 210    1    0     1     1     1     1     1     1        6         1
#> 119 210    1    0     1     1     1     1     1     1        6         1
#> 120 210    1    0     1     1     1     1     1     1        6         1
#> 121 211    1    0     1     1     1     1     1     1        6         1
#> 122 211    0    1     1     1     1     1     0     1        5         1
#> 123 211    1    0     1     1     1     1     0     1        5         1
#> 124 211    0    1     1     1     1     1     1     1        6         1
#> 125 211    0    1     1     1     1     1     1     1        6         1
#> 126 211    1    0     1     1     1     1     1     1        6         1
#> 127 211    1    0     1     1     1     1     1     1        6         1
#> 128 211    0    1     1     1     1     1     1     1        6         1
#> 129 211    0    1     1     1     1     1     1     1        6         1
#> 130 211    0    1     1     1     1     1     0     1        5         1
#> 131 211    0    1     1     1     1     1     0     1        5         1
#> 132 211    0    1     1     1     1     1     1     1        6         1
#> 133 212    0    1     1     1     1     1     1     1        6         1
#> 134 212    1    0     1     1     1     1     0     1        5         1
#> 135 212    0    1     1     1     1     1     1     1        6         1
#> 136 212    1    0     1     1     1     1     1     1        6         1
#> 137 212    0    1     1     1     1     1     1     1        6         1
#> 138 212    1    0     1     1     1     1     0     1        5         1
#> 139 212    0    1     1     1     1     1     1     1        6         1
#> 140 212    0    1     1     1     1     1     1     1        6         1
#> 141 212    1    0     1     1     1     1     1     1        6         1
#> 142 212    1    0     1     1     1     1     1     1        6         1
#> 143 212    1    0     1     1     1     1     1     1        6         1
#> 144 212    0    1     1     1     1     1     0     1        5         1
#> 145 213    1    0     1     1     1     1     1     1        6         1
#> 146 213    0    1     1     1     1     1     1     1        6         1
#> 147 213    0    1     1     1     1     1     1     1        6         1
#> 148 213    0    1     1     1     1     1     1     1        6         1
#> 149 213    0    1     1     1     1     1     0     1        5         1
#> 150 213    1    0     1     1     1     1     0     1        5         1
#> 151 213    1    0     1     1     1     1     1     1        6         1
#> 152 213    1    0     1     1     1     1     1     1        6         1
#> 153 213    0    1     1     1     1     1     1     1        6         1
#> 154 213    0    1     1     1     1     1     1     1        6         1
#> 155 213    0    1     1     1     1     1     1     1        6         1
#> 156 213    0    1     1     1     1     1     0     1        5         1
#> 157 214    0    1     1     1     1     1     1     1        6         1
#> 158 214    1    0     1     1     1     1     1     1        6         1
#> 159 214    0    1     1     1     1     1     1     1        6         1
#> 160 214    1    0     1     1     1     1     1     1        6         1
#> 161 214    1    0     1     1     1     1     1     1        6         1
#> 162 214    0    1     1     1     1     1     0     1        5         1
#> 163 214    1    0     1     1     1     1     1     1        6         1
#> 164 214    0    1     1     1     1     1     1     1        6         1
#> 165 214    1    0     1     1     1     1     1     1        6         1
#> 166 214    0    1     1     1     1     1     0     1        5         1
#> 167 214    0    1     1     1     1     1     1     1        6         1
#> 168 214    0    1     1     1     1     1     0     1        5         1
#> 169 215    0    1     1     1     1     1     1     1        6         1
#> 170 215    0    1     1     1     1     1     1     1        6         1
#> 171 215    0    1     1     1     1     1     1     1        6         1
#> 172 215    1    0     1     1     1     1     1     1        6         1
#> 173 215    1    0     1     1     1     1     0     1        5         1
#> 174 215    1    0     1     1     1     1     0     1        5         1
#> 175 215    0    1     1     1     1     1     1     1        6         1
#> 176 215    1    0     1     1     1     1     1     1        6         1
#> 177 215    0    1     1     1     1     1     0     1        5         1
#> 178 215    0    1     1     1     1     1     0     1        5         1
#> 179 215    1    0     1     1     1     1     1     1        6         1
#> 180 215    1    0     1     1     1     1     1     1        6         1
#> 181 216    1    0     1     1     1     1     0     1        5         1
#> 182 216    0    1     1     1     1     1     1     1        6         1
#> 183 216    1    0     1     1     1     1     1     1        6         1
#> 184 216    0    1     1     1     1     1     1     1        6         1
#> 185 216    0    1     1     1     1     1     1     1        6         1
#> 186 216    0    1     1     1     1     1     1     1        6         1
#> 187 216    1    0     1     1     1     1     1     1        6         1
#> 188 216    0    1     1     1     1     1     1     1        6         1
#> 189 216    1    0     0     0     0     0     0     0        0         0
#> 190 216    0    1     1     1     1     1     1     1        6         1
#> 191 216    0    1     1     1     1     1     1     1        6         1
#> 192 216    0    1     1     1     1     1     0     1        5         1
#>     classADL2 classADL3 hasHelp unmetNeed
#> 1           0         0       1         0
#> 2           0         0       1         0
#> 3           0         0       0         0
#> 4           0         0       0         0
#> 5           0         0       0         1
#> 6           0         0       1         0
#> 7           0         0       1         0
#> 8           1         0       1         0
#> 9           0         0       0         1
#> 10          0         0       1         0
#> 11          0         0       1         0
#> 12          1         0       1         0
#> 13          0         0       1         0
#> 14          0         0       1         0
#> 15          0         0       1         0
#> 16          0         0       0         1
#> 17          0         0       1         0
#> 18          0         0       1         0
#> 19          0         0       1         0
#> 20          0         0       1         0
#> 21          0         0       1         0
#> 22          0         0       0         0
#> 23          0         0       0         0
#> 24          0         0       1         0
#> 25          0         0       1         0
#> 26          0         0       0         0
#> 27          0         0       1         0
#> 28          0         0       1         0
#> 29          0         0       1         0
#> 30          0         0       0         0
#> 31          0         0       0         0
#> 32          0         0       1         0
#> 33          0         0       1         0
#> 34          0         0       1         0
#> 35          0         0       1         0
#> 36          0         0       1         0
#> 37          0         0       1         0
#> 38          0         0       1         0
#> 39          1         0       0         1
#> 40          0         0       0         0
#> 41          0         0       1         0
#> 42          0         0       1         0
#> 43          0         0       0         1
#> 44          0         0       1         0
#> 45          0         0       1         0
#> 46          0         0       1         0
#> 47          0         0       1         0
#> 48          0         0       0         0
#> 49          0         0       1         0
#> 50          0         0       0         0
#> 51          0         0       0         0
#> 52          0         0       1         0
#> 53          0         0       1         0
#> 54          0         0       1         0
#> 55          0         0       1         0
#> 56          0         0       1         0
#> 57          0         0       1         0
#> 58          0         0       1         0
#> 59          0         0       1         0
#> 60          0         0       0         0
#> 61          0         0       0         0
#> 62          0         0       0         0
#> 63          0         0       1         0
#> 64          0         0       1         0
#> 65          0         0       0         0
#> 66          0         0       1         0
#> 67          0         0       0         0
#> 68          0         0       1         0
#> 69          0         0       1         0
#> 70          0         0       0         1
#> 71          0         0       1         0
#> 72          0         0       1         0
#> 73          0         0       1         0
#> 74          0         0       1         0
#> 75          0         0       0         0
#> 76          0         0       1         0
#> 77          0         0       1         0
#> 78          0         0       1         0
#> 79          0         0       0         0
#> 80          0         0       1         0
#> 81          0         0       1         0
#> 82          0         0       1         0
#> 83          0         0       1         0
#> 84          0         0       0         0
#> 85          0         0       1         0
#> 86          0         0       1         0
#> 87          0         0       1         0
#> 88          0         0       0         0
#> 89          0         0       1         0
#> 90          0         0       0         0
#> 91          0         0       0         0
#> 92          0         0       0         1
#> 93          0         0       1         0
#> 94          0         0       1         0
#> 95          0         0       1         0
#> 96          0         0       0         1
#> 97          0         0       0         1
#> 98          0         0       1         0
#> 99          0         0       1         0
#> 100         0         0       0         0
#> 101         0         0       0         0
#> 102         0         0       1         0
#> 103         0         0       0         0
#> 104         0         0       0         0
#> 105         0         0       0         0
#> 106         0         0       1         0
#> 107         0         0       0         1
#> 108         0         0       1         0
#> 109         0         0       0         0
#> 110         0         0       1         0
#> 111         0         0       1         0
#> 112         0         1       1         0
#> 113         0         0       0         0
#> 114         0         0       1         0
#> 115         0         0       1         0
#> 116         0         0       0         0
#> 117         0         0       1         0
#> 118         0         0       0         0
#> 119         0         0       0         0
#> 120         0         0       1         0
#> 121         0         0       0         0
#> 122         0         0       1         0
#> 123         0         0       1         0
#> 124         0         0       0         0
#> 125         0         0       1         0
#> 126         0         0       1         0
#> 127         0         0       0         0
#> 128         0         0       0         0
#> 129         0         0       0         0
#> 130         0         0       0         1
#> 131         0         0       0         1
#> 132         0         0       0         0
#> 133         0         0       1         0
#> 134         0         0       0         1
#> 135         0         0       0         0
#> 136         0         0       0         0
#> 137         0         0       1         0
#> 138         0         0       1         0
#> 139         0         0       1         0
#> 140         0         0       1         0
#> 141         0         0       1         0
#> 142         0         0       1         0
#> 143         0         0       1         0
#> 144         0         0       1         0
#> 145         0         0       0         0
#> 146         0         0       0         0
#> 147         0         0       0         0
#> 148         0         0       0         0
#> 149         0         0       0         1
#> 150         0         0       0         1
#> 151         0         0       1         0
#> 152         0         0       0         0
#> 153         0         0       0         0
#> 154         0         0       1         0
#> 155         0         0       1         0
#> 156         0         0       1         0
#> 157         0         0       1         0
#> 158         0         0       1         0
#> 159         0         0       1         0
#> 160         0         0       1         0
#> 161         0         0       1         0
#> 162         0         0       0         1
#> 163         0         0       1         0
#> 164         0         0       0         0
#> 165         0         0       0         0
#> 166         0         0       1         0
#> 167         0         0       1         0
#> 168         0         0       1         0
#> 169         0         0       0         0
#> 170         0         0       0         0
#> 171         0         0       1         0
#> 172         0         0       0         0
#> 173         0         0       0         1
#> 174         0         0       0         1
#> 175         0         0       0         0
#> 176         0         0       1         0
#> 177         0         0       1         0
#> 178         0         0       0         1
#> 179         0         0       0         0
#> 180         0         0       0         0
#> 181         0         0       0         1
#> 182         0         0       1         0
#> 183         0         0       0         0
#> 184         0         0       1         0
#> 185         0         0       0         0
#> 186         0         0       1         0
#> 187         0         0       1         0
#> 188         0         0       0         0
#> 189         0         1       0         1
#> 190         0         0       1         0
#> 191         0         0       1         0
#> 192         0         0       1         0