Create older people indicators dataframe for demography and situation from survey data collected using the standard RAM-OP questionnaire

Create older people indicators dataframe for demography and situation from survey data collected using the standard RAM-OP questionnaire

Usage

create_op_demo(svy)

Arguments

svy

A dataframe collected using the standard RAM-OP questionnaire

Value

A dataframe of older people indicators on demography and situation

Demography and situation

psu

Primary sampling unit

resp1

Respondent is SUBJECT

resp2

Respondent is FAMILY CARER

resp3

Respondent is OTHER CARER

resp4

Respondent is OTHER

age

Age of respondent (years)

ageGrp1

Age of respondent is between 50 and 59 years

ageGrp2

Age of respondent is between 60 and 69 years

ageGrp3

Age of respondent is between 70 and 79 years

ageGrp4

Age of respondent is between 80 and 89 years

ageGrp5

Age of respondent is between 90 years and older

sex1

Male

sex2

Female

marital1

Marital status = SINGLE

marital2

Marital status = MARRIED

marital3

Marital status = LIVING TOGETHER

marital4

Marital status = DIVORCED

marital5

Marital status = SEPARATED

marital6

Marital status = OTHER

alone

Respondent lives alone

Author

Mark Myatt

Examples

# Create demography and situation indicators dataset from RAM-OP survey data
# collected from Addis Ababa, Ethiopia
create_op_demo(testSVY)
#> # A tibble: 192 × 20
#>      psu resp1 resp2 resp3 resp4   age ageGrp1 ageGrp2 ageGrp3 ageGrp4 ageGrp5
#>    <int> <dbl> <dbl> <dbl> <dbl> <int>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1   201     1     0     0     0    67       0       1       0       0       0
#>  2   201     1     0     0     0    74       0       0       1       0       0
#>  3   201     1     0     0     0    60       0       1       0       0       0
#>  4   201     1     0     0     0    60       0       1       0       0       0
#>  5   201     1     0     0     0    85       0       0       0       1       0
#>  6   201     0     1     0     0    86       0       0       0       1       0
#>  7   201     1     0     0     0    80       0       0       0       1       0
#>  8   201     1     0     0     0    60       0       1       0       0       0
#>  9   201     1     0     0     0    62       0       1       0       0       0
#> 10   201     1     0     0     0    72       0       0       1       0       0
#> # … with 182 more rows, and 9 more variables: sex1 <dbl>, sex2 <dbl>,
#> #   marital1 <dbl>, marital2 <dbl>, marital3 <dbl>, marital4 <dbl>,
#> #   marital5 <dbl>, marital6 <dbl>, alone <dbl>