Create older people indicators for food intake from survey data collected using the standard RAM-OP questionnaire

Create older people indicators for food intake from survey data collected using the standard RAM-OP questionnaire

create_op_food(svy)

Arguments

svy
A dataframe collected using the standard RAM-OP questionniare

Value

A dataframe of older people indicators on food intake

Indicators

Dietary intake indicators

These dietary intake indicators have been purpose-built for older people but the basic approach used is described in:

Kennedy G, Ballard T, Dop M C (2011). Guidelines for Measuring Household and Individual Dietary Diversity. Rome, FAO http://www.fao.org/docrep/014/i1983e/i1983e00.htm

and extended to include indicators of probable adequate intake of a number of nutrients / micronutrients.

MF

Meal frequenct

DDS

Dietary Diversity Score (count of 11 groups)

FG01

Cereals

FG02

Roots and tubers

FG03

Fruits and vegetables

FG04

All meat

FG05

Eggs

FG06

Fish

FG07

Legumes, nuts and seeds

FG08

Milk and milk products

FG09

Fats

FG10

Sugar

FG11

Other

proteinRich

Protein rich foods

pProtein

Protein rich plant sources of protein

aProtein

Protein rich animal sources of protein

pVitA

Plant sources of vitamin A

aVitA

Animal sources of vitamin A

xVitA

Any source of vitamin A

ironRich

Iron rich foods

caRich

Calcium rich foods

znRich

Zinc rich foods

vitB1

Vitamin B1-rich foods

vitB2

Vitamin B2-rich foods

vitB3

Vitamin B3-rich foods

vitB6

Vitamin B6-rich foods

vitB12

Vitamin B12-rich foods

vitBcomplex

Vitamin B1/B2/B3/B6/B12-rich foods

Examples

# Create food intake indicators dataset from RAM-OP survey data collected
# from Addis Ababa, Ethiopia
create_op_food(testSVY)

#>     psu sex1 sex2 MF DDS FG01 FG02 FG03 FG04 FG05 FG06 FG07 FG08 FG09 FG10 FG11
#> 1   201    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 2   201    1    0  3   7    1    1    1    0    0    0    1    0    1    1    1
#> 3   201    1    0  2   2    0    0    0    0    0    1    0    0    0    0    1
#> 4   201    0    1  3   3    1    1    0    0    0    0    1    0    0    0    0
#> 5   201    0    1  3   5    1    0    1    0    0    1    0    0    0    1    1
#> 6   201    1    0  4   5    1    1    0    0    0    1    1    0    0    0    1
#> 7   201    1    0  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 8   201    0    1  3   4    1    1    0    0    0    0    1    0    0    0    1
#> 9   201    1    0  2   4    1    0    1    0    0    0    0    0    0    1    1
#> 10  201    0    1  2   6    1    1    1    1    0    0    0    0    0    1    1
#> 11  201    0    1  2   4    1    0    1    0    0    0    0    0    1    0    1
#> 12  201    0    1  4   3    1    0    0    0    0    0    0    0    0    1    1
#> 13  202    1    0  3   4    1    0    1    0    0    0    1    0    0    0    1
#> 14  202    1    0  4   6    1    1    0    0    0    0    1    0    1    1    1
#> 15  202    0    1  2   5    1    1    0    0    0    1    0    0    0    1    1
#> 16  202    0    1  1   5    1    1    1    0    0    1    0    0    0    0    1
#> 17  202    0    1  3   6    1    1    1    0    0    0    1    0    0    1    1
#> 18  202    0    1  1   3    1    0    0    0    0    0    0    0    0    1    1
#> 19  202    1    0  2   6    1    1    1    0    0    1    1    0    0    0    1
#> 20  202    1    0  2   5    1    0    1    0    0    1    1    0    0    0    1
#> 21  202    1    0  2   5    1    0    1    0    0    1    1    0    0    0    1
#> 22  202    0    1  2   4    1    0    1    0    0    0    0    0    0    1    1
#> 23  202    0    1  3   7    1    1    1    0    0    1    0    1    1    0    1
#> 24  202    0    1  3   6    1    1    1    0    0    1    0    0    0    1    1
#> 25  203    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 26  203    0    1  5   4    1    1    0    0    0    0    1    0    0    0    1
#> 27  203    1    0  3   8    1    1    1    1    0    1    0    0    1    1    1
#> 28  203    0    1  3   4    1    0    0    0    0    0    1    0    0    1    1
#> 29  203    0    1  4   5    1    0    0    0    0    0    1    0    1    1    1
#> 30  203    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 31  203    0    1  3   5    1    0    0    0    0    0    1    0    1    1    1
#> 32  203    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 33  203    1    0  3   5    1    1    1    0    0    0    0    0    1    0    1
#> 34  203    1    0  3   7    1    1    1    0    0    1    1    0    0    1    1
#> 35  203    1    0  3   4    0    1    1    0    0    1    0    0    0    0    1
#> 36  203    1    0  2   7    1    1    1    0    1    0    1    0    0    1    1
#> 37  204    0    1  2   7    1    1    1    0    0    1    1    0    0    1    1
#> 38  204    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 39  204    0    1  2   2    0    1    0    0    0    0    0    0    0    0    1
#> 40  204    1    0  2   3    1    0    1    0    0    0    1    0    0    0    0
#> 41  204    0    1  2   5    1    1    0    0    0    1    0    0    0    1    1
#> 42  204    1    0  2   6    1    1    0    0    0    1    0    1    0    1    1
#> 43  204    1    0  4   5    1    0    0    0    0    1    0    0    1    1    1
#> 44  204    0    1  3   6    1    0    1    0    0    1    1    0    0    1    1
#> 45  204    0    1  3   3    1    0    0    0    0    1    0    0    0    0    1
#> 46  204    0    1  2   7    1    1    1    1    0    0    1    0    0    1    1
#> 47  204    0    1  2   6    1    1    1    0    0    0    1    0    0    1    1
#> 48  204    0    1  3   6    1    1    1    0    0    0    1    0    1    0    1
#> 49  205    0    1  4   7    1    1    1    0    0    0    1    0    1    1    1
#> 50  205    0    1  2   4    1    1    1    0    0    0    0    0    0    0    1
#> 51  205    0    1  2   4    1    1    0    0    0    0    0    0    1    0    1
#> 52  205    0    1  2   5    1    0    1    0    0    0    1    1    0    0    1
#> 53  205    0    1  2   3    1    0    0    0    0    0    0    0    0    1    1
#> 54  205    1    0  2   4    0    1    0    0    0    1    0    0    0    1    1
#> 55  205    0    1  4   6    1    1    1    1    0    1    0    0    0    0    1
#> 56  205    0    1  3   4    1    0    1    0    0    0    0    0    0    1    1
#> 57  205    0    1  2   6    1    1    1    0    0    0    1    0    0    1    1
#> 58  205    0    1  3   5    1    0    0    0    0    1    1    0    0    1    1
#> 59  205    1    0  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 60  205    0    1  2   6    1    1    1    0    0    0    1    0    0    1    1
#> 61  206    1    0  2   6    1    1    1    1    0    1    0    0    0    0    1
#> 62  206    0    1  4   3    0    1    1    0    0    0    0    0    0    0    1
#> 63  206    0    1  2   5    1    1    1    0    0    0    1    0    0    0    1
#> 64  206    0    1  2   4    1    0    1    0    0    0    1    0    0    0    1
#> 65  206    1    0  4   3    0    0    1    0    0    0    0    0    0    1    1
#> 66  206    1    0  3   4    1    1    0    0    0    0    0    0    0    1    1
#> 67  206    0    1  4   5    1    0    0    0    0    1    0    0    1    1    1
#> 68  206    1    0  3   6    1    1    0    0    0    0    1    0    1    1    1
#> 69  206    1    0  4   6    1    0    1    0    1    0    0    0    1    1    1
#> 70  206    1    0  1   1    0    0    0    0    0    0    0    0    0    0    1
#> 71  206    1    0  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 72  206    0    1  3   6    1    1    1    0    0    0    1    0    0    1    1
#> 73  207    1    0  3   6    1    1    0    0    0    1    1    0    1    0    1
#> 74  207    0    1  1   2    0    0    1    0    0    0    0    0    0    0    1
#> 75  207    1    0  3   4    1    0    1    0    0    1    0    0    1    0    0
#> 76  207    0    1  3   4    1    0    1    0    0    1    0    0    0    0    1
#> 77  207    1    0  4   5    1    1    1    0    0    0    0    0    0    1    1
#> 78  207    0    1  3   4    1    1    0    0    0    0    0    0    0    1    1
#> 79  207    0    1  2   4    1    0    0    0    0    1    0    0    0    1    1
#> 80  207    0    1  2   5    1    0    1    0    0    0    1    0    0    1    1
#> 81  207    1    0  3   7    1    1    1    0    0    1    1    0    0    1    1
#> 82  207    0    1  1   3    1    0    0    0    0    1    0    0    0    0    1
#> 83  207    0    1  2   3    0    1    1    0    0    0    0    0    0    0    1
#> 84  207    1    0  4   7    1    1    1    0    0    1    0    0    1    1    1
#> 85  208    0    1  3   5    1    0    0    0    0    1    0    0    1    1    1
#> 86  208    0    1  4   5    1    1    1    0    0    0    0    0    0    1    1
#> 87  208    1    0  3   5    1    1    1    0    0    0    1    0    0    0    1
#> 88  208    0    1  2   4    1    0    0    0    0    1    0    0    0    1    1
#> 89  208    0    1  3   5    1    0    0    0    0    0    1    0    1    1    1
#> 90  208    1    0  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 91  208    0    1  2   4    1    1    0    0    0    0    1    0    0    0    1
#> 92  208    1    0  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 93  208    0    1  2   5    1    0    0    0    0    1    1    0    0    1    1
#> 94  208    0    1  3   5    1    1    1    0    0    0    0    0    0    1    1
#> 95  208    1    0  3   5    1    1    0    0    0    1    1    0    0    0    1
#> 96  208    0    1  4   5    1    0    1    0    0    0    1    0    0    1    1
#> 97  209    0    1  2   6    1    1    0    0    0    1    0    0    1    1    1
#> 98  209    0    1  3   5    1    1    1    0    0    0    0    1    0    0    1
#> 99  209    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 100 209    1    0  2   4    1    0    1    0    0    0    0    0    0    1    1
#> 101 209    1    0  2   3    1    0    0    0    0    1    0    0    0    0    1
#> 102 209    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 103 209    1    0  1   4    1    1    0    0    0    0    1    0    0    0    1
#> 104 209    1    0  4   6    1    1    0    0    0    0    1    0    1    1    1
#> 105 209    1    0  1   4    1    1    0    0    0    0    1    0    0    0    1
#> 106 209    1    0  1   4    1    0    0    0    0    1    0    0    0    1    1
#> 107 209    1    0  3   5    1    0    0    0    0    1    0    0    1    1    1
#> 108 209    1    0  2   3    1    0    1    0    0    0    0    0    0    0    1
#> 109 210    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 110 210    0    1  3   4    1    0    1    0    0    0    1    0    0    0    1
#> 111 210    0    1  4   8    1    1    1    1    0    1    0    0    1    1    1
#> 112 210    1    0  2   6    1    1    1    0    0    1    1    0    0    0    1
#> 113 210    0    1  1   3    1    1    0    0    0    0    1    0    0    0    0
#> 114 210    1    0  4   5    1    1    0    1    0    0    0    0    0    1    1
#> 115 210    0    1  2   4    1    0    0    0    0    1    0    0    0    1    1
#> 116 210    1    0  2   4    1    0    0    0    0    0    1    0    0    1    1
#> 117 210    0    1  2   4    0    1    1    0    0    0    0    0    0    1    1
#> 118 210    1    0  1   4    1    0    0    0    0    0    0    0    1    1    1
#> 119 210    1    0  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 120 210    1    0  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 121 211    1    0  3   5    1    0    1    0    0    0    1    0    0    1    1
#> 122 211    0    1  3   7    1    1    1    0    0    1    1    0    0    1    1
#> 123 211    1    0  2   5    1    0    0    0    0    0    1    0    1    1    1
#> 124 211    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 125 211    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 126 211    1    0  4   7    1    1    1    0    1    1    1    0    0    0    1
#> 127 211    1    0  1   3    0    0    1    0    0    0    0    0    0    1    1
#> 128 211    0    1  2   4    1    0    0    0    1    0    1    0    0    0    1
#> 129 211    0    1  2   5    1    1    0    0    0    0    0    0    1    1    1
#> 130 211    0    1  3   6    1    0    1    0    0    0    1    0    1    1    1
#> 131 211    0    1  3   5    0    1    1    1    0    0    0    0    0    1    1
#> 132 211    0    1  2   3    0    1    1    0    0    0    0    0    0    0    1
#> 133 212    0    1  3   2    1    0    0    0    0    0    0    0    0    0    1
#> 134 212    1    0  3   4    0    1    0    0    0    1    0    0    0    1    1
#> 135 212    0    1  3   5    1    1    0    0    0    0    0    0    1    1    1
#> 136 212    1    0  4   7    1    1    1    0    0    1    1    0    0    1    1
#> 137 212    0    1  2   4    1    0    1    0    0    0    0    0    0    1    1
#> 138 212    1    0  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 139 212    0    1  4   5    1    1    1    0    0    0    0    0    0    1    1
#> 140 212    0    1  1   2    1    0    1    0    0    0    0    0    0    0    0
#> 141 212    1    0  3   4    1    0    0    0    0    0    1    0    0    1    1
#> 142 212    1    0  3   6    1    1    1    0    0    0    1    0    1    0    1
#> 143 212    1    0  1   3    1    0    0    0    0    1    0    0    0    0    1
#> 144 212    0    1  2   4    0    1    1    0    0    0    1    0    1    0    0
#> 145 213    1    0  2   4    1    1    0    0    0    0    0    0    0    1    1
#> 146 213    0    1  4   4    1    0    0    0    0    1    0    0    0    1    1
#> 147 213    0    1  3   4    1    0    1    0    0    1    0    0    0    0    1
#> 148 213    0    1  2   5    1    1    1    0    0    0    0    0    0    1    1
#> 149 213    0    1  4   4    1    1    0    0    0    0    0    0    0    1    1
#> 150 213    1    0  2   3    1    0    1    0    0    0    0    0    0    0    1
#> 151 213    1    0  4   5    1    1    1    0    0    0    0    0    0    1    1
#> 152 213    1    0  5   5    1    0    1    0    0    0    0    0    1    1    1
#> 153 213    0    1  1   2    1    0    0    0    0    0    0    0    0    0    1
#> 154 213    0    1  4   9    1    1    1    1    0    1    1    0    1    1    1
#> 155 213    0    1  2   3    1    0    0    0    0    1    0    0    0    0    1
#> 156 213    0    1  4   6    1    1    1    0    0    1    1    0    0    0    1
#> 157 214    0    1  4   7    1    1    1    1    0    1    1    0    0    0    1
#> 158 214    1    0  3   6    1    1    1    0    0    1    0    0    1    0    1
#> 159 214    0    1  2   3    0    1    0    0    0    0    0    0    0    1    1
#> 160 214    1    0  3   6    1    1    0    0    0    1    0    0    1    1    1
#> 161 214    1    0  4   5    1    1    1    0    0    1    0    0    0    0    1
#> 162 214    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 163 214    1    0  3   5    1    1    1    0    0    1    0    0    0    0    1
#> 164 214    0    1  2   2    1    0    0    0    0    0    0    0    0    0    1
#> 165 214    1    0  2   5    1    0    0    0    0    0    1    0    1    1    1
#> 166 214    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 167 214    0    1  3   6    1    0    1    0    1    0    1    0    1    0    1
#> 168 214    0    1  3   5    1    1    1    0    0    0    1    0    0    0    1
#> 169 215    0    1  1   2    0    0    0    0    0    0    0    0    0    1    1
#> 170 215    0    1  4   3    1    0    0    0    0    0    0    0    0    1    1
#> 171 215    0    1  2   4    1    1    0    0    0    0    1    0    0    0    1
#> 172 215    1    0  3   7    1    1    1    0    0    1    1    0    1    0    1
#> 173 215    1    0  2   3    1    0    0    0    0    1    0    0    0    0    1
#> 174 215    1    0  3   3    1    0    1    0    0    0    0    0    0    0    1
#> 175 215    0    1  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 176 215    1    0  4   3    1    0    0    0    0    0    0    0    0    1    1
#> 177 215    0    1  3   5    1    1    1    0    0    0    0    0    0    1    1
#> 178 215    0    1  3   3    1    0    1    0    0    0    0    0    0    0    1
#> 179 215    1    0  2   3    1    1    0    0    0    0    1    0    0    0    0
#> 180 215    1    0  3   3    1    0    1    0    0    0    0    0    0    0    1
#> 181 216    1    0  3   2    1    0    0    0    0    0    0    0    0    0    1
#> 182 216    0    1  4   7    1    1    1    0    0    1    1    0    1    0    1
#> 183 216    1    0  1   3    1    0    1    0    0    0    0    0    0    0    1
#> 184 216    0    1  3   6    1    1    1    0    0    1    1    0    0    0    1
#> 185 216    0    1  3   4    1    1    1    0    0    0    0    0    0    0    1
#> 186 216    0    1  3   4    1    0    0    0    0    0    0    0    1    1    1
#> 187 216    1    0  2   4    1    0    1    0    0    1    0    0    0    0    1
#> 188 216    0    1  5   5    1    0    1    0    0    0    1    0    0    1    1
#> 189 216    1    0  2   4    1    0    1    0    0    0    0    0    0    1    1
#> 190 216    0    1  3   6    0    1    1    0    0    0    0    1    1    1    1
#> 191 216    0    1  2   7    1    1    0    1    0    1    1    0    1    0    1
#> 192 216    0    1  2   3    1    0    1    0    0    0    0    0    0    0    1
#>     proteinRich pProtein aProtein pVitA aVitA xVitA ironRich caRich znRich
#> 1             1        1        0     1     0     1        1      0      1
#> 2             1        1        0     1     0     1        1      0      1
#> 3             0        0        0     0     0     0        1      0      1
#> 4             1        1        0     0     0     0        0      0      1
#> 5             0        0        0     1     0     1        1      0      1
#> 6             1        1        0     0     0     0        1      0      1
#> 7             1        1        0     1     0     1        1      0      1
#> 8             1        1        0     0     0     0        0      0      1
#> 9             0        0        0     1     0     1        1      0      0
#> 10            1        0        1     1     0     1        1      0      1
#> 11            0        0        0     1     0     1        1      0      0
#> 12            0        0        0     0     0     0        0      0      0
#> 13            1        1        0     1     0     1        1      0      1
#> 14            1        1        0     0     0     0        0      0      1
#> 15            0        0        0     0     0     0        1      0      1
#> 16            0        0        0     0     0     0        1      0      1
#> 17            1        1        0     1     0     1        1      0      1
#> 18            0        0        0     0     0     0        0      0      0
#> 19            1        1        0     1     0     1        1      0      1
#> 20            1        1        0     1     0     1        1      0      1
#> 21            1        1        0     1     0     1        1      0      1
#> 22            0        0        0     1     0     1        1      0      0
#> 23            1        0        1     1     1     1        1      1      1
#> 24            0        0        0     1     0     1        1      0      1
#> 25            0        0        0     1     0     1        1      0      0
#> 26            1        1        0     0     0     0        0      0      1
#> 27            1        0        1     1     0     1        1      0      1
#> 28            1        1        0     0     0     0        0      0      1
#> 29            1        1        0     0     0     0        0      0      1
#> 30            1        1        0     1     0     1        1      0      1
#> 31            1        1        0     1     0     1        0      0      1
#> 32            0        0        0     1     0     1        1      0      0
#> 33            0        0        0     1     0     1        0      0      0
#> 34            1        1        0     1     0     1        1      0      1
#> 35            0        0        0     1     0     1        1      0      1
#> 36            1        1        1     0     1     1        0      0      1
#> 37            1        1        0     1     0     1        1      0      1
#> 38            0        0        0     1     0     1        1      0      0
#> 39            0        0        0     0     0     0        0      0      0
#> 40            1        1        0     1     0     1        1      0      1
#> 41            0        0        0     0     0     0        1      0      1
#> 42            1        0        1     0     1     1        1      1      1
#> 43            0        0        0     0     0     0        1      0      1
#> 44            1        1        0     1     0     1        1      0      1
#> 45            0        0        0     0     0     0        1      0      1
#> 46            1        1        1     1     0     1        1      0      1
#> 47            1        1        0     1     0     1        1      0      1
#> 48            1        1        0     1     0     1        1      0      1
#> 49            1        1        0     1     0     1        1      0      1
#> 50            0        0        0     1     0     1        1      0      0
#> 51            0        0        0     1     0     1        0      0      0
#> 52            1        1        1     0     1     1        0      1      1
#> 53            0        0        0     0     0     0        0      0      0
#> 54            0        0        0     0     0     0        1      0      1
#> 55            1        0        1     1     0     1        1      0      1
#> 56            0        0        0     1     0     1        1      0      0
#> 57            1        1        0     1     0     1        1      0      1
#> 58            1        1        0     0     0     0        1      0      1
#> 59            0        0        0     1     0     1        0      0      0
#> 60            1        1        0     1     0     1        1      0      1
#> 61            1        0        1     0     0     0        1      0      1
#> 62            0        0        0     1     0     1        1      0      0
#> 63            1        1        0     1     0     1        1      0      1
#> 64            1        1        0     1     0     1        1      0      1
#> 65            0        0        0     1     0     1        1      0      0
#> 66            0        0        0     0     0     0        0      0      0
#> 67            0        0        0     1     0     1        1      0      1
#> 68            1        1        0     0     0     0        0      0      1
#> 69            1        0        1     1     1     1        1      0      0
#> 70            0        0        0     0     0     0        0      0      0
#> 71            0        0        0     1     0     1        1      0      0
#> 72            1        1        0     1     0     1        1      0      1
#> 73            1        1        0     0     0     0        1      0      1
#> 74            0        0        0     1     0     1        1      0      0
#> 75            0        0        0     1     0     1        1      0      1
#> 76            0        0        0     1     0     1        1      0      1
#> 77            0        0        0     1     0     1        1      0      0
#> 78            0        0        0     0     0     0        0      0      0
#> 79            0        0        0     0     0     0        1      0      1
#> 80            1        1        0     1     0     1        1      0      1
#> 81            1        1        0     1     0     1        1      0      1
#> 82            0        0        0     0     0     0        1      0      1
#> 83            0        0        0     0     0     0        0      0      0
#> 84            0        0        0     1     0     1        1      0      1
#> 85            0        0        0     0     0     0        1      0      1
#> 86            0        0        0     1     0     1        1      0      0
#> 87            1        1        0     1     0     1        1      0      1
#> 88            0        0        0     0     0     0        1      0      1
#> 89            1        1        0     1     0     1        0      0      1
#> 90            0        0        0     0     0     0        0      0      0
#> 91            1        1        0     0     0     0        0      0      1
#> 92            0        0        0     0     0     0        0      0      0
#> 93            1        1        0     0     0     0        1      0      1
#> 94            0        0        0     1     0     1        1      0      0
#> 95            1        1        0     0     0     0        1      0      1
#> 96            1        1        0     1     0     1        1      0      1
#> 97            0        0        0     0     0     0        1      0      1
#> 98            1        0        1     1     1     1        1      1      0
#> 99            0        0        0     1     0     1        0      0      0
#> 100           0        0        0     1     0     1        1      0      0
#> 101           0        0        0     0     0     0        1      0      1
#> 102           1        1        0     1     0     1        1      0      1
#> 103           1        1        0     0     0     0        0      0      1
#> 104           1        1        0     0     0     0        0      0      1
#> 105           1        1        0     0     0     0        0      0      1
#> 106           0        0        0     0     0     0        1      0      1
#> 107           0        0        0     1     0     1        1      0      1
#> 108           0        0        0     1     0     1        1      0      0
#> 109           0        0        0     1     0     1        1      0      0
#> 110           1        1        0     1     0     1        1      0      1
#> 111           1        0        1     1     0     1        1      0      1
#> 112           1        1        0     1     0     1        1      0      1
#> 113           1        1        0     0     0     0        0      0      1
#> 114           1        0        1     0     0     0        1      0      1
#> 115           0        0        0     0     0     0        1      0      1
#> 116           1        1        0     0     0     0        0      0      1
#> 117           0        0        0     1     0     1        1      0      0
#> 118           0        0        0     1     0     1        0      0      0
#> 119           0        0        0     0     0     0        0      0      0
#> 120           0        0        0     1     0     1        0      0      0
#> 121           1        1        0     1     0     1        0      0      1
#> 122           1        1        0     1     0     1        1      0      1
#> 123           1        1        0     1     0     1        0      0      1
#> 124           0        0        0     1     0     1        1      0      0
#> 125           0        0        0     1     0     1        1      0      0
#> 126           1        1        1     1     1     1        1      0      1
#> 127           0        0        0     0     0     0        0      0      0
#> 128           1        1        1     0     1     1        0      0      1
#> 129           0        0        0     1     0     1        0      0      0
#> 130           1        1        0     1     0     1        1      0      1
#> 131           1        0        1     1     0     1        1      0      1
#> 132           0        0        0     1     0     1        1      0      0
#> 133           0        0        0     0     0     0        0      0      0
#> 134           0        0        0     0     0     0        1      0      1
#> 135           0        0        0     0     0     0        0      0      0
#> 136           1        1        0     1     0     1        1      0      1
#> 137           0        0        0     1     0     1        1      0      0
#> 138           0        0        0     0     0     0        0      0      0
#> 139           0        0        0     1     0     1        1      0      0
#> 140           0        0        0     0     0     0        0      0      0
#> 141           1        1        0     0     0     0        0      0      1
#> 142           1        1        0     1     0     1        1      0      1
#> 143           0        0        0     0     0     0        1      0      1
#> 144           1        1        0     1     0     1        1      0      1
#> 145           0        0        0     0     0     0        0      0      0
#> 146           0        0        0     0     0     0        1      0      1
#> 147           0        0        0     1     0     1        1      0      1
#> 148           0        0        0     1     0     1        1      0      0
#> 149           0        0        0     0     0     0        0      0      0
#> 150           0        0        0     1     0     1        1      0      0
#> 151           0        0        0     1     0     1        1      0      0
#> 152           0        0        0     1     0     1        1      0      0
#> 153           0        0        0     0     0     0        0      0      0
#> 154           1        1        1     1     0     1        1      0      1
#> 155           0        0        0     0     0     0        1      0      1
#> 156           1        1        0     1     0     1        1      0      1
#> 157           1        1        1     1     0     1        1      0      1
#> 158           0        0        0     1     0     1        1      0      1
#> 159           0        0        0     0     0     0        0      0      0
#> 160           0        0        0     1     0     1        1      0      1
#> 161           0        0        0     1     0     1        1      0      1
#> 162           0        0        0     1     0     1        1      0      0
#> 163           0        0        0     1     0     1        1      0      1
#> 164           0        0        0     0     0     0        0      0      0
#> 165           1        1        0     0     0     0        0      0      1
#> 166           1        1        0     1     0     1        1      0      1
#> 167           1        1        1     1     1     1        1      0      1
#> 168           1        1        0     1     0     1        1      0      1
#> 169           0        0        0     0     0     0        0      0      0
#> 170           0        0        0     0     0     0        0      0      0
#> 171           1        1        0     0     0     0        0      0      1
#> 172           1        1        0     1     0     1        1      0      1
#> 173           0        0        0     0     0     0        1      0      1
#> 174           0        0        0     1     0     1        1      0      0
#> 175           0        0        0     1     0     1        1      0      0
#> 176           0        0        0     0     0     0        0      0      0
#> 177           0        0        0     1     0     1        1      0      0
#> 178           0        0        0     1     0     1        1      0      0
#> 179           1        1        0     0     0     0        0      0      1
#> 180           0        0        0     1     0     1        0      0      0
#> 181           0        0        0     0     0     0        0      0      0
#> 182           1        1        0     1     0     1        1      0      1
#> 183           0        0        0     1     0     1        1      0      0
#> 184           1        1        0     1     0     1        1      0      1
#> 185           0        0        0     1     0     1        1      0      0
#> 186           0        0        0     1     0     1        0      0      0
#> 187           0        0        0     1     0     1        1      0      1
#> 188           1        1        0     1     0     1        1      0      1
#> 189           0        0        0     1     0     1        1      0      0
#> 190           1        0        1     1     1     1        1      1      0
#> 191           1        1        1     0     0     0        1      0      1
#> 192           0        0        0     1     0     1        1      0      0
#>     vitB1 vitB2 vitB3 vitB6 vitB12 vitBcomplex
#> 1       1     1     1     1      1           1
#> 2       1     1     1     1      0           0
#> 3       1     1     1     1      1           1
#> 4       1     1     1     1      0           0
#> 5       1     1     1     1      1           1
#> 6       1     1     1     1      1           1
#> 7       1     1     1     1      1           1
#> 8       1     1     1     1      0           0
#> 9       0     1     0     1      0           0
#> 10      1     1     1     1      1           1
#> 11      0     1     0     1      0           0
#> 12      0     0     0     0      0           0
#> 13      1     1     1     1      0           0
#> 14      1     1     1     1      0           0
#> 15      1     1     1     1      1           1
#> 16      1     1     1     1      1           1
#> 17      1     1     1     1      0           0
#> 18      0     0     0     0      0           0
#> 19      1     1     1     1      1           1
#> 20      1     1     1     1      1           1
#> 21      1     1     1     1      1           1
#> 22      0     1     0     1      0           0
#> 23      1     1     1     1      1           1
#> 24      1     1     1     1      1           1
#> 25      0     1     0     1      0           0
#> 26      1     1     1     1      0           0
#> 27      1     1     1     1      1           1
#> 28      1     1     1     1      0           0
#> 29      1     1     1     1      0           0
#> 30      1     1     1     1      1           1
#> 31      1     1     1     1      0           0
#> 32      0     1     0     1      0           0
#> 33      0     0     0     1      0           0
#> 34      1     1     1     1      1           1
#> 35      1     1     1     1      1           1
#> 36      1     1     1     1      1           1
#> 37      1     1     1     1      1           1
#> 38      0     1     0     1      0           0
#> 39      0     0     0     0      0           0
#> 40      1     1     1     1      0           0
#> 41      1     1     1     1      1           1
#> 42      1     1     1     1      1           1
#> 43      1     1     1     1      1           1
#> 44      1     1     1     1      1           1
#> 45      1     1     1     1      1           1
#> 46      1     1     1     1      1           1
#> 47      1     1     1     1      0           0
#> 48      1     1     1     1      0           0
#> 49      1     1     1     1      0           0
#> 50      0     1     0     1      0           0
#> 51      1     0     0     0      0           0
#> 52      1     1     1     1      1           1
#> 53      0     0     0     0      0           0
#> 54      1     1     1     1      1           1
#> 55      1     1     1     1      1           1
#> 56      0     1     0     1      0           0
#> 57      1     1     1     1      0           0
#> 58      1     1     1     1      1           1
#> 59      0     0     0     1      0           0
#> 60      1     1     1     1      0           0
#> 61      1     1     1     1      1           1
#> 62      0     1     0     1      0           0
#> 63      1     1     1     1      0           0
#> 64      1     1     1     1      0           0
#> 65      0     1     0     1      0           0
#> 66      0     0     0     0      0           0
#> 67      1     1     1     1      1           1
#> 68      1     1     1     1      0           0
#> 69      1     1     0     1      1           0
#> 70      0     0     0     0      0           0
#> 71      0     1     0     1      0           0
#> 72      1     1     1     1      0           0
#> 73      1     1     1     1      1           1
#> 74      0     1     0     1      0           0
#> 75      1     1     1     1      1           1
#> 76      1     1     1     1      1           1
#> 77      0     1     0     1      0           0
#> 78      0     0     0     0      0           0
#> 79      1     1     1     1      1           1
#> 80      1     1     1     1      0           0
#> 81      1     1     1     1      1           1
#> 82      1     1     1     1      1           1
#> 83      0     0     0     1      0           0
#> 84      1     1     1     1      1           1
#> 85      1     1     1     1      1           1
#> 86      0     1     0     1      0           0
#> 87      1     1     1     1      0           0
#> 88      1     1     1     1      1           1
#> 89      1     1     1     1      0           0
#> 90      0     0     0     0      0           0
#> 91      1     1     1     1      0           0
#> 92      0     0     0     0      0           0
#> 93      1     1     1     1      1           1
#> 94      0     1     0     1      0           0
#> 95      1     1     1     1      1           1
#> 96      1     1     1     1      0           0
#> 97      1     1     1     1      1           1
#> 98      1     1     0     1      0           0
#> 99      0     0     0     1      0           0
#> 100     0     1     0     1      0           0
#> 101     1     1     1     1      1           1
#> 102     1     1     1     1      1           1
#> 103     1     1     1     1      0           0
#> 104     1     1     1     1      0           0
#> 105     1     1     1     1      0           0
#> 106     1     1     1     1      1           1
#> 107     1     1     1     1      1           1
#> 108     0     1     0     1      0           0
#> 109     0     1     0     1      0           0
#> 110     1     1     1     1      0           0
#> 111     1     1     1     1      1           1
#> 112     1     1     1     1      1           1
#> 113     1     1     1     1      0           0
#> 114     1     1     1     1      1           1
#> 115     1     1     1     1      1           1
#> 116     1     1     1     1      0           0
#> 117     0     1     0     1      0           0
#> 118     1     0     0     0      0           0
#> 119     0     0     0     0      0           0
#> 120     0     0     0     1      0           0
#> 121     1     1     1     1      0           0
#> 122     1     1     1     1      1           1
#> 123     1     1     1     1      0           0
#> 124     0     1     0     1      0           0
#> 125     0     1     0     1      0           0
#> 126     1     1     1     1      1           1
#> 127     0     0     0     1      0           0
#> 128     1     1     1     1      1           1
#> 129     1     0     0     0      0           0
#> 130     1     1     1     1      0           0
#> 131     1     1     1     1      1           1
#> 132     0     1     0     1      0           0
#> 133     0     0     0     0      0           0
#> 134     1     1     1     1      1           1
#> 135     0     0     0     0      0           0
#> 136     1     1     1     1      1           1
#> 137     0     1     0     1      0           0
#> 138     0     0     0     0      0           0
#> 139     0     1     0     1      0           0
#> 140     0     0     0     1      0           0
#> 141     1     1     1     1      0           0
#> 142     1     1     1     1      0           0
#> 143     1     1     1     1      1           1
#> 144     1     1     1     1      0           0
#> 145     0     0     0     0      0           0
#> 146     1     1     1     1      1           1
#> 147     1     1     1     1      1           1
#> 148     0     1     0     1      0           0
#> 149     0     0     0     0      0           0
#> 150     0     1     0     1      0           0
#> 151     0     1     0     1      0           0
#> 152     1     1     0     1      0           0
#> 153     0     0     0     0      0           0
#> 154     1     1     1     1      1           1
#> 155     1     1     1     1      1           1
#> 156     1     1     1     1      1           1
#> 157     1     1     1     1      1           1
#> 158     1     1     1     1      1           1
#> 159     0     0     0     0      0           0
#> 160     1     1     1     1      1           1
#> 161     1     1     1     1      1           1
#> 162     0     1     0     1      0           0
#> 163     1     1     1     1      1           1
#> 164     0     0     0     0      0           0
#> 165     1     1     1     1      0           0
#> 166     1     1     1     1      1           1
#> 167     1     1     1     1      1           1
#> 168     1     1     1     1      0           0
#> 169     0     0     0     0      0           0
#> 170     0     0     0     0      0           0
#> 171     1     1     1     1      0           0
#> 172     1     1     1     1      1           1
#> 173     1     1     1     1      1           1
#> 174     0     1     0     1      0           0
#> 175     0     1     0     1      0           0
#> 176     0     0     0     0      0           0
#> 177     0     1     0     1      0           0
#> 178     0     1     0     1      0           0
#> 179     1     1     1     1      0           0
#> 180     0     0     0     1      0           0
#> 181     0     0     0     0      0           0
#> 182     1     1     1     1      1           1
#> 183     0     1     0     1      0           0
#> 184     1     1     1     1      1           1
#> 185     0     1     0     1      0           0
#> 186     1     0     0     0      0           0
#> 187     1     1     1     1      1           1
#> 188     1     1     1     1      0           0
#> 189     0     1     0     1      0           0
#> 190     1     1     0     1      1           0
#> 191     1     1     1     1      1           1
#> 192     0     1     0     1      0           0