Create female older people indicators dataframe for health and health-seeking behaviours from survey data collected using the standard RAM-OP questionnaire

Create female older people indicators dataframe for health and health-seeking behaviours from survey data collected using the standard RAM-OP questionnaire

Usage

create_op_health_females(svy)

Arguments

svy

A dataframe collected using the standard RAM-OP questionnaire

Value

A tibble of female older people indicators on health and health-seeking behaviours

Examples

# Create health and health-seeking behaviours indicators dataset from RAM-OP
# survey data collected from Addis Ababa, Ethiopia
create_op_health_females(testSVY)
#> # A tibble: 113 × 25
#>      psu  sex1  sex2    H1    H2   H31   H32   H33   H34   H35   H36   H37   H38
#>    <int> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl> <dbl>
#>  1   201     0     1     1     1    NA    NA    NA    NA    NA    NA    NA    NA
#>  2   201     0     1     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  3   201     0     1     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  4   201     0     1     1     1    NA    NA    NA    NA    NA    NA    NA    NA
#>  5   201     0     1     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  6   201     0     1     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  7   201     0     1     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#>  8   202     0     1     1     1    NA    NA    NA    NA    NA    NA    NA    NA
#>  9   202     0     1     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> 10   202     0     1     0    NA    NA    NA    NA    NA    NA    NA    NA    NA
#> # ℹ 103 more rows
#> # ℹ 12 more variables: H39 <dbl>, H4 <dbl>, H5 <dbl>, H61 <dbl>, H62 <dbl>,
#> #   H63 <dbl>, H64 <dbl>, H65 <dbl>, H66 <dbl>, H67 <dbl>, H68 <dbl>, H69 <dbl>