Function to apply bootstrap to RAM-OP indicators using a classical estimator.

Function to apply bootstrap to RAM-OP indicators using a classical estimator.

estimateClassic(
  x,
  w,
  statistic = bootClassic,
  indicators = c("demo", "food", "hunger", "adl", "disability", "mental", "dementia",
    "health", "oedema", "screening", "income", "wash", "visual", "misc"),
  params = get_variables(indicators),
  outputColumns = params,
  replicates = 399
)

Arguments

x
Indicators dataset produced by createOP with primary sampling unit (PSU) in column named PSU

w
A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

statistic
A function operating on data in x; fixed to bootClassic function for means

indicators
A character vector of indicator set names to estimate. Indicator set names are demo, food, hunger, disability, adl, mental, dementia, health, income, wash, visual, and misc. Default is all indicator sets.

params
Parameters (named columns in x) passed to the function specified in statistic. This is equivalent to variables corresponding to the indicator sets specified in indicators. The function get_variables() is used to specify these variables.

outputColumns
Names of columns in output data frame. This defaults to values specified in params

replicates
Number of bootstrap replicates

Value

Dataframe of boot estimates using bootClassic mean function

Examples

  #
  test <- estimateClassic(x = indicators.ALL,
                          w = testPSU,
                          replicates = 9)

  test

#> # A tibble: 136 x 10
#>    LCL.ALL EST.ALL UCL.ALL LCL.MALES EST.MALES UCL.MALES LCL.FEMALES EST.FEMALES
#>      <dbl>   <dbl>   <dbl>     <dbl>     <dbl>     <dbl>       <dbl>       <dbl>
#>  1  0.794   0.859   0.880     0.743     0.803     0.835      0.786       0.849  
#>  2  0.075   0.0885  0.152     0.0308    0.1       0.125      0.0874      0.123  
#>  3  0.0219  0.0417  0.0729    0.0309    0.0658    0.166      0.00793     0.0169 
#>  4  0       0       0.0229    0         0.0132    0.0889     0           0.00826
#>  5 69.4    70.7    72.1      69.3      71.0      72.5       70.0        71.1    
#>  6  0       0       0         0         0         0          0           0      
#>  7  0.441   0.536   0.586     0.423     0.470     0.609      0.383       0.517  
#>  8  0.181   0.229   0.328     0.211     0.325     0.380      0.143       0.218  
#>  9  0.123   0.198   0.277     0.0890    0.153     0.268      0.129       0.272  
#> 10  0.0219  0.0365  0.0573    0.0236    0.0488    0.0747     0           0.0391 
#> # … with 126 more rows, and 2 more variables: UCL.FEMALES <dbl>,
#> #   INDICATOR <chr>