estimateClassic

Function to apply bootstrap to RAM-OP indicators using a classical estimator.

estimateClassic(
  x,
  w,
  statistic = bootClassic,
  indicators = c("demo", "food", "hunger", "adl", "disability", "mental", "dementia",
    "health", "oedema", "screening", "income", "wash", "visual", "misc"),
  params = get_variables(indicators),
  outputColumns = params,
  replicates = 399
)

Arguments

x
Indicators dataset produced by createOP with primary sampling unit (PSU) in column named PSU

w
A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

statistic
A function operating on data in x; fixed to bootClassic function for means

indicators
A character vector of indicator set names to estimate. Indicator set names are demo, food, hunger, disability, adl, mental, dementia, health, income, wash, visual, and misc. Default is all indicator sets.

params
Parameters (named columns in x) passed to the function specified in statistic. This is equivalent to variables corresponding to the indicator sets specified in indicators. The function get_variables() is used to specify these variables.

outputColumns
Names of columns in output data frame. This defaults to values specified in params

replicates
Number of bootstrap replicates

Value

Dataframe of boot estimates using bootClassic mean function

Examples

  #
  test <- estimateClassic(x = indicators.ALL,
                          w = testPSU,
                          replicates = 9)

  test

#>          LCL.ALL      EST.ALL     UCL.ALL    LCL.MALES   EST.MALES   UCL.MALES
#> 1    0.793750000  0.859375000  0.88020833  0.743319838  0.80263158  0.83450035
#> 2    0.075000000  0.088541667  0.15208333  0.030846485  0.10000000  0.12521616
#> 3    0.021875000  0.041666667  0.07291667  0.030871722  0.06578947  0.16617647
#> 4    0.000000000  0.000000000  0.02291667  0.000000000  0.01315789  0.08887408
#> 5   69.434375000 70.739583333 72.09479167 69.345251076 71.02631579 72.54055728
#> 6    0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 7    0.440625000  0.536458333  0.58645833  0.423461538  0.46987952  0.60895338
#> 8    0.181250000  0.229166667  0.32812500  0.211423279  0.32530120  0.38024133
#> 9    0.122916667  0.197916667  0.27708333  0.089009288  0.15294118  0.26772353
#> 10   0.021875000  0.036458333  0.05729167  0.023642807  0.04878049  0.07470588
#> 11   0.364583333  0.395833333  0.49895833  1.000000000  1.00000000  1.00000000
#> 12   0.501041667  0.604166667  0.63541667  0.000000000  0.00000000  0.00000000
#> 13   0.008333333  0.026041667  0.05312500  0.000000000  0.01219512  0.03787975
#> 14   0.261458333  0.333333333  0.38437500  0.407647059  0.55696203  0.65314506
#> 15   0.068750000  0.093750000  0.17083333  0.121566110  0.15789474  0.21096386
#> 16   0.052083333  0.078125000  0.11041667  0.019050319  0.07500000  0.20015480
#> 17   0.378125000  0.458333333  0.53020833  0.111039795  0.20000000  0.24306457
#> 18   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 19   0.115625000  0.145833333  0.20416667  0.051832112  0.14102564  0.29205882
#> 20   2.457291667  2.572916667  2.74375000  2.439473684  2.64705882  2.72111763
#> 21   4.330208333  4.494791667  4.86145833  4.111438696  4.52439024  4.88785425
#> 22   0.856250000  0.895833333  0.95208333  0.884829721  0.93902439  0.96226081
#> 23   0.453125000  0.531250000  0.60416667  0.367647059  0.52631579  0.63998511
#> 24   0.523958333  0.567708333  0.61250000  0.454487952  0.58823529  0.67631579
#> 25   0.030208333  0.057291667  0.10520833  0.000000000  0.05063291  0.10265460
#> 26   0.016666667  0.031250000  0.04687500  0.002500000  0.04411765  0.07727554
#> 27   0.285416667  0.348958333  0.38854167  0.300354359  0.43589744  0.54736842
#> 28   0.354166667  0.395833333  0.47395833  0.264643963  0.36764706  0.46860759
#> 29   0.005208333  0.015625000  0.05104167  0.000000000  0.00000000  0.01235294
#> 30   0.145833333  0.218750000  0.24687500  0.162948264  0.28048780  0.33374613
#> 31   0.406250000  0.463541667  0.57500000  0.312323492  0.42647059  0.49268072
#> 32   0.930208333  0.968750000  0.98333333  0.905538020  0.98684211  1.00000000
#> 33   0.414583333  0.468750000  0.52604167  0.344736842  0.40963855  0.49243671
#> 34   0.354166667  0.395833333  0.47395833  0.264643963  0.36764706  0.46860759
#> 35   0.089583333  0.098958333  0.19270833  0.028823529  0.07594937  0.18135747
#> 36   0.556250000  0.630208333  0.65833333  0.494004252  0.65000000  0.69635628
#> 37   0.036458333  0.046875000  0.09166667  0.005000000  0.04411765  0.08167183
#> 38   0.567708333  0.645833333  0.66979167  0.498710135  0.66250000  0.73653251
#> 39   0.648958333  0.697916667  0.70729167  0.548603827  0.60256410  0.74582043
#> 40   0.005208333  0.015625000  0.05104167  0.000000000  0.00000000  0.01235294
#> 41   0.576041667  0.619791667  0.67291667  0.522778172  0.64473684  0.71117647
#> 42   0.621875000  0.661458333  0.69583333  0.578080495  0.67088608  0.71117647
#> 43   0.782291667  0.807291667  0.87291667  0.678639263  0.74390244  0.88842105
#> 44   0.576041667  0.619791667  0.67291667  0.522778172  0.64473684  0.71117647
#> 45   0.819791667  0.848958333  0.90937500  0.774514529  0.86585366  0.92421053
#> 46   0.370833333  0.401041667  0.45000000  0.347647059  0.48529412  0.62105263
#> 47   0.364583333  0.395833333  0.44479167  0.340588235  0.48529412  0.58947368
#> 48   0.718750000  0.786458333  0.82604167  0.641494288  0.71794872  0.87470588
#> 49   0.140625000  0.171875000  0.24791667  0.125294118  0.28205128  0.35850571
#> 50   0.021875000  0.041666667  0.04166667  0.000000000  0.00000000  0.01052632
#> 51   0.937500000  0.968750000  0.99375000  0.916740647  0.97058824  0.98771226
#> 52   0.973958333  0.984375000  0.99895833  0.935320228  0.97368421  1.00000000
#> 53   0.973958333  0.984375000  0.99895833  0.935320228  0.97368421  1.00000000
#> 54   0.914583333  0.973958333  0.99375000  0.935320228  0.96470588  0.99705882
#> 55   0.633333333  0.718750000  0.78333333  0.675547217  0.76315789  0.85331348
#> 56   0.989583333  0.994791667  1.00000000  0.965821179  0.98717949  1.00000000
#> 57   5.460416667  5.651041667  5.71458333  5.458370323  5.63157895  5.77196575
#> 58   0.933333333  0.984375000  0.99375000  0.935320228  0.97368421  1.00000000
#> 59   0.001041667  0.005208333  0.05520833  0.000000000  0.00000000  0.00000000
#> 60   0.001041667  0.010416667  0.02500000  0.000000000  0.02631579  0.06467977
#> 61   0.481250000  0.578125000  0.66875000  0.424189383  0.64102564  0.72770898
#> 62   0.090625000  0.114583333  0.18750000  0.061019150  0.13157895  0.24102263
#> 63   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 64   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 65   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 66   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 67   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 68   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 69   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 70   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 71   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 72   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 73   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 74   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 75   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 76   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 77   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 78   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 79   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 80   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 81   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 82   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 83   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 84   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 85   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 86   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 87   1.000000000  1.000000000  1.00000000  1.000000000  1.00000000  1.00000000
#> 88   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 89   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 90   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 91   0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 92  11.371875000 11.906250000 13.17395833  9.118802099 11.91025641 13.31879518
#> 93   0.426041667  0.458333333  0.58958333  0.310462131  0.44117647  0.62764706
#> 94   0.121875000  0.197916667  0.24270833  0.096140784  0.18421053  0.31607843
#> 95   0.373958333  0.432291667  0.48854167  0.286944046  0.36842105  0.43371795
#> 96   0.706940223  0.741935484  0.79555556  0.400619195  0.67647059  0.82439560
#> 97   0.008695652  0.125000000  0.26052632  0.000000000  0.20000000  0.55000000
#> 98   0.211111111  0.315789474  0.65507246  0.147619048  0.20000000  0.42222222
#> 99   0.042857143  0.100000000  0.26262626  0.000000000  0.00000000  0.00000000
#> 100  0.008695652  0.142857143  0.26052632  0.025000000  0.27272727  0.54285714
#> 101  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 102  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 103  0.000000000  0.050000000  0.17212121  0.000000000  0.00000000  0.00000000
#> 104  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 105  0.072380952  0.136363636  0.34000000  0.128571429  0.20000000  0.42285714
#> 106  0.823958333  0.864583333  0.89062500  0.781424149  0.88157895  0.92129179
#> 107  0.794532343  0.821656051  0.85730994  0.710943396  0.78873239  0.83042142
#> 108  0.000000000  0.043478261  0.16876268  0.000000000  0.06666667  0.29059829
#> 109  0.640784314  0.848484848  0.98064516  0.647552448  0.86666667  0.98666667
#> 110  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 111  0.000000000  0.064516129  0.12503748  0.000000000  0.06666667  0.31757576
#> 112  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 113  0.000000000  0.032258065  0.13764706  0.000000000  0.00000000  0.00000000
#> 114  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 115  0.000000000  0.000000000  0.08588235  0.000000000  0.00000000  0.00000000
#> 116  0.000000000  0.000000000  0.00000000  0.000000000  0.00000000  0.00000000
#> 117  0.518750000  0.583333333  0.63541667  0.560873494  0.61538462  0.76321377
#> 118  0.328125000  0.369791667  0.51458333  0.392352941  0.47368421  0.62047497
#> 119  0.032291667  0.125000000  0.18541667  0.104608434  0.22784810  0.31681643
#> 120  0.011458333  0.041666667  0.05625000  0.019079940  0.05063291  0.09216643
#> 121  0.019791667  0.052083333  0.07604167  0.000000000  0.02500000  0.05495356
#> 122  0.000000000  0.005208333  0.01562500  0.000000000  0.00000000  0.00000000
#> 123  0.000000000  0.015625000  0.04479167  0.000000000  0.03947368  0.06347717
#> 124  0.005208333  0.010416667  0.02916667  0.000000000  0.02409639  0.04554180
#> 125  0.288541667  0.322916667  0.37708333  0.208561304  0.26923077  0.39514706
#> 126  0.000000000  0.005208333  0.02500000  0.000000000  0.01265823  0.06724309
#> 127  0.562500000  0.598958333  0.65104167  0.555294118  0.63157895  0.68529503
#> 128  0.617708333  0.697916667  0.73541667  0.642412509  0.69512195  0.73040397
#> 129  0.194791667  0.255208333  0.29687500  0.165064562  0.25000000  0.52102055
#> 130  0.189583333  0.234375000  0.29166667  0.162123386  0.25000000  0.50174344
#> 131  0.006250000  0.010416667  0.04687500  0.000000000  0.01315789  0.04862058
#> 132  0.020833333  0.036458333  0.06562500  0.002564103  0.02631579  0.06622669
#> 133  0.318750000  0.375000000  0.51354167  0.408756515  0.47368421  0.59784076
#> 134  0.219791667  0.270833333  0.34375000  0.156546855  0.25000000  0.29878543
#> 135  0.027083333  0.052083333  0.07291667  0.012941176  0.03846154  0.07727554
#> 136  0.004166667  0.036458333  0.05520833  0.000000000  0.02531646  0.06610914
#>      LCL.FEMALES  EST.FEMALES UCL.FEMALES         INDICATOR
#> 1    0.786227033  0.848739496  0.88321123             resp1
#> 2    0.087374847  0.122807018  0.18682524             resp2
#> 3    0.007930672  0.016949153  0.03573935             resp3
#> 4    0.000000000  0.008264463  0.01692067             resp4
#> 5   70.018273810 71.110169492 73.50773540               age
#> 6    0.000000000  0.000000000  0.00000000           ageGrp1
#> 7    0.383310841  0.516949153  0.58677871           ageGrp2
#> 8    0.143030303  0.218487395  0.31778846           ageGrp3
#> 9    0.129343220  0.271929825  0.31349565           ageGrp4
#> 10   0.000000000  0.039062500  0.07103052           ageGrp5
#> 11   0.000000000  0.000000000  0.00000000              sex1
#> 12   1.000000000  1.000000000  1.00000000              sex2
#> 13   0.005084746  0.035714286  0.06432113          marital1
#> 14   0.097062415  0.140625000  0.19842615          marital2
#> 15   0.028432530  0.070175439  0.09235653          marital3
#> 16   0.025191854  0.042735043  0.09989407          marital4
#> 17   0.600290557  0.714285714  0.75545524          marital5
#> 18   0.000000000  0.000000000  0.00000000          marital6
#> 19   0.075630252  0.102564103  0.16699818             alone
#> 20   2.527466426  2.640350877  2.77348273                MF
#> 21   4.579520941  4.811965812  4.97714286               DDS
#> 22   0.911428571  0.929824561  0.96940315              FG01
#> 23   0.522803020  0.603305785  0.64084821              FG02
#> 24   0.534055728  0.584745763  0.69232493              FG03
#> 25   0.045616606  0.093750000  0.13028335              FG04
#> 26   0.001562500  0.008547009  0.03383459              FG05
#> 27   0.224376693  0.312500000  0.39031250              FG06
#> 28   0.355630165  0.447368421  0.53612009              FG07
#> 29   0.018530124  0.041322314  0.08869048              FG08
#> 30   0.187000981  0.230769231  0.30089286              FG09
#> 31   0.454523810  0.554621849  0.65254615              FG10
#> 32   0.926190476  0.966942149  0.99663866              FG11
#> 33   0.451509534  0.526315789  0.59159664       proteinRich
#> 34   0.355630165  0.447368421  0.53612009          pProtein
#> 35   0.098743087  0.149122807  0.21328217          aProtein
#> 36   0.588805014  0.656250000  0.75406162             pVitA
#> 37   0.028571429  0.052631579  0.09404762             aVitA
#> 38   0.616180031  0.683760684  0.75910364             xVitA
#> 39   0.652316311  0.714285714  0.80367560          ironRich
#> 40   0.018530124  0.041322314  0.08869048            caRich
#> 41   0.549371484  0.609375000  0.69506303            znRich
#> 42   0.574092010  0.669642857  0.74812325             vitB1
#> 43   0.837260120  0.876190476  0.91018170             vitB2
#> 44   0.549371484  0.609375000  0.69506303             vitB3
#> 45   0.838955036  0.899159664  0.93549258             vitB6
#> 46   0.273245709  0.414062500  0.44756892            vitB12
#> 47   0.263104216  0.401785714  0.42766917       vitBcomplex
#> 48   0.704648760  0.771186441  0.84810745              HHS1
#> 49   0.134017094  0.177966102  0.22154664              HHS2
#> 50   0.012976190  0.050847458  0.08362603              HHS3
#> 51   0.944012605  0.975206612  1.00000000             ADL01
#> 52   0.984768908  0.992187500  1.00000000             ADL02
#> 53   0.984768908  0.992187500  1.00000000             ADL03
#> 54   0.898569954  0.950413223  0.99495798             ADL04
#> 55   0.642352941  0.722689076  0.77617510             ADL05
#> 56   1.000000000  1.000000000  1.00000000             ADL06
#> 57   5.534350582  5.616071429  5.70480950          scoreADL
#> 58   0.942414861  0.974789916  0.99830508         classADL1
#> 59   0.001694915  0.025210084  0.05758514         classADL2
#> 60   0.000000000  0.000000000  0.00000000         classADL3
#> 61   0.594369835  0.658119658  0.71047619           hasHelp
#> 62   0.023050214  0.076190476  0.09568038         unmetNeed
#> 63   1.000000000  1.000000000  1.00000000        wgVisionD0
#> 64   0.000000000  0.000000000  0.00000000        wgVisionD1
#> 65   0.000000000  0.000000000  0.00000000        wgVisionD2
#> 66   0.000000000  0.000000000  0.00000000        wgVisionD3
#> 67   1.000000000  1.000000000  1.00000000       wgHearingD0
#> 68   0.000000000  0.000000000  0.00000000       wgHearingD1
#> 69   0.000000000  0.000000000  0.00000000       wgHearingD2
#> 70   0.000000000  0.000000000  0.00000000       wgHearingD3
#> 71   1.000000000  1.000000000  1.00000000      wgMobilityD0
#> 72   0.000000000  0.000000000  0.00000000      wgMobilityD1
#> 73   0.000000000  0.000000000  0.00000000      wgMobilityD2
#> 74   0.000000000  0.000000000  0.00000000      wgMobilityD3
#> 75   1.000000000  1.000000000  1.00000000   wgRememberingD0
#> 76   0.000000000  0.000000000  0.00000000   wgRememberingD1
#> 77   0.000000000  0.000000000  0.00000000   wgRememberingD2
#> 78   0.000000000  0.000000000  0.00000000   wgRememberingD3
#> 79   1.000000000  1.000000000  1.00000000      wgSelfCareD0
#> 80   0.000000000  0.000000000  0.00000000      wgSelfCareD1
#> 81   0.000000000  0.000000000  0.00000000      wgSelfCareD2
#> 82   0.000000000  0.000000000  0.00000000      wgSelfCareD3
#> 83   1.000000000  1.000000000  1.00000000 wgCommunicatingD0
#> 84   0.000000000  0.000000000  0.00000000 wgCommunicatingD1
#> 85   0.000000000  0.000000000  0.00000000 wgCommunicatingD2
#> 86   0.000000000  0.000000000  0.00000000 wgCommunicatingD3
#> 87   1.000000000  1.000000000  1.00000000              wgP0
#> 88   0.000000000  0.000000000  0.00000000              wgP1
#> 89   0.000000000  0.000000000  0.00000000              wgP2
#> 90   0.000000000  0.000000000  0.00000000              wgP3
#> 91   0.000000000  0.000000000  0.00000000              wgPM
#> 92  11.530467394 12.605042017 13.09448622                K6
#> 93   0.449461768  0.491071429  0.57022042            K6Case
#> 94   0.198431373  0.254385965  0.30785124                DS
#> 95   0.421679198  0.476562500  0.52267471                H1
#> 96   0.613308913  0.842105263  0.92343420                H2
#> 97   0.000000000  0.111111111  0.23333333               H31
#> 98   0.066666667  0.428571429  0.91111111               H32
#> 99   0.000000000  0.240000000  0.88571429               H33
#> 100  0.000000000  0.000000000  0.00000000               H34
#> 101  0.000000000  0.000000000  0.00000000               H35
#> 102  0.000000000  0.000000000  0.00000000               H36
#> 103  0.000000000  0.080000000  0.24444444               H37
#> 104  0.000000000  0.000000000  0.00000000               H38
#> 105  0.000000000  0.000000000  0.40000000               H39
#> 106  0.816507620  0.884297521  0.94442577                H4
#> 107  0.793814433  0.845360825  0.93476982                H5
#> 108  0.000000000  0.100000000  0.17878788               H61
#> 109  0.713846154  0.842105263  0.92848485               H62
#> 110  0.000000000  0.000000000  0.00000000               H63
#> 111  0.000000000  0.000000000  0.00000000               H64
#> 112  0.000000000  0.000000000  0.00000000               H65
#> 113  0.000000000  0.052631579  0.19076923               H66
#> 114  0.000000000  0.000000000  0.00000000               H67
#> 115  0.000000000  0.000000000  0.12307692               H68
#> 116  0.000000000  0.000000000  0.00000000               H69
#> 117  0.453018162  0.483050847  0.57790063                M1
#> 118  0.200427350  0.272727273  0.34330532               M2A
#> 119  0.008489063  0.043859649  0.07436975               M2B
#> 120  0.000000000  0.008928571  0.02798535               M2C
#> 121  0.047417355  0.070175439  0.12135316               M2D
#> 122  0.000000000  0.000000000  0.02473436               M2E
#> 123  0.000000000  0.000000000  0.00000000               M2F
#> 124  0.000000000  0.017857143  0.03026316               M2G
#> 125  0.296651786  0.323809524  0.44201681               M2H
#> 126  0.000000000  0.000000000  0.01741071               M2I
#> 127  0.555954349  0.610169492  0.66144958                W1
#> 128  0.707142857  0.765625000  0.82452058                W2
#> 129  0.139226271  0.219047619  0.29598214                W3
#> 130  0.132388664  0.198347107  0.29200969                W4
#> 131  0.018518000  0.039062500  0.06958573            oedema
#> 132  0.016806723  0.042372881  0.05680731          screened
#> 133  0.267202381  0.336134454  0.53492555            poorVA
#> 134  0.205475207  0.271929825  0.43086081              chew
#> 135  0.011037851  0.050420168  0.07028509              food
#> 136  0.000000000  0.008771930  0.06286157              NFRI