estimateClassic

estimateClassic

estimateClassic(x, w, statistic = bootClassic, params = c("resp1", "resp2",
  "resp3", "resp4", "age", "ageGrp1", "ageGrp2", "ageGrp3", "ageGrp4",
  "ageGrp5", "sex1", "sex2", "marital1", "marital2", "marital3", "marital4",
  "marital5", "marital6", "alone", "MF", "DDS", "FG01", "FG02", "FG03", "FG04",
  "FG05", "FG06", "FG07", "FG08", "FG09", "FG10", "FG11", "proteinRich",
  "pProtein", "aProtein", "pVitA", "aVitA", "xVitA", "ironRich", "caRich",
  "znRich", "vitB1", "vitB2", "vitB3", "vitB6", "vitB12", "vitBcomplex", "HHS1",
  "HHS2", "HHS3", "ADL01", "ADL02", "ADL03",      "ADL04", "ADL05", "ADL06",
  "scoreADL", "classADL1", "classADL2", "classADL3", "hasHelp", "unmetNeed",
  "K6", "K6Case", "DS", "H1", "H2", "H31", "H32", "H33", "H34", "H35", "H36",
  "H37", "H38", "H39", "H4", "H5", "H61", "H62", "H63", "H64", "H65", "H66",
  "H67", "H68", "H69", "M1", "M2A", "M2B", "M2C", "M2D", "M2E", "M2F", "M2G",
  "M2H", "M2I", "W1", "W2", "W3", "W4", "MUAC", "oedema", "screened", "poorVA",
  "chew", "food", "NFRI", "wgVisionD0", "wgVisionD1", "wgVisionD2",
  "wgVisionD3", "wgHearingD0",      "wgHearingD1", "wgHearingD2", "wgHearingD3",
  "wgMobilityD0", "wgMobilityD1", "wgMobilityD2", "wgMobilityD3",
  "wgRememberingD0", "wgRememberingD1", "wgRememberingD2", "wgRememberingD3",
  "wgSelfCareD0", "wgSelfCareD1", "wgSelfCareD2", "wgSelfCareD3",
  "wgCommunicatingD0", "wgCommunicatingD1", "wgCommunicatingD2",
  "wgCommunicatingD3", "wgP0", "wgP1", "wgP2", "wgP3", "wgPM"),
  outputColumns = c("resp1", "resp2", "resp3", "resp4", "age", "ageGrp1",
  "ageGrp2", "ageGrp3", "ageGrp4", "ageGrp5", "sex1", "sex2", "marital1",
  "marital2", "marital3", "marital4", "marital5", "marital6", "alone", "MF",
  "DDS", "FG01", "FG02", "FG03", "FG04", "FG05", "FG06", "FG07", "FG08", "FG09",
  "FG10", "FG11", "proteinRich", "pProtein", "aProtein", "pVitA", "aVitA",
  "xVitA", "ironRich", "caRich", "znRich", "vitB1", "vitB2", "vitB3", "vitB6",
  "vitB12", "vitBcomplex", "HHS1", "HHS2", "HHS3", "ADL01", "ADL02", "ADL03",
    "ADL04", "ADL05", "ADL06", "scoreADL", "classADL1", "classADL2",
  "classADL3", "hasHelp", "unmetNeed", "K6", "K6Case", "DS", "H1", "H2", "H31",
  "H32", "H33", "H34", "H35", "H36", "H37", "H38", "H39", "H4", "H5", "H61",
  "H62", "H63", "H64", "H65", "H66", "H67", "H68", "H69", "M1", "M2A", "M2B",
  "M2C", "M2D", "M2E", "M2F", "M2G", "M2H", "M2I", "W1", "W2", "W3", "W4",
  "MUAC", "oedema", "screened", "poorVA", "chew", "food", "NFRI", "wgVisionD0",
  "wgVisionD1", "wgVisionD2", "wgVisionD3", "wgHearingD0",      "wgHearingD1",
  "wgHearingD2", "wgHearingD3", "wgMobilityD0", "wgMobilityD1", "wgMobilityD2",
  "wgMobilityD3", "wgRememberingD0", "wgRememberingD1", "wgRememberingD2",
  "wgRememberingD3", "wgSelfCareD0", "wgSelfCareD1", "wgSelfCareD2",
  "wgSelfCareD3", "wgCommunicatingD0", "wgCommunicatingD1", "wgCommunicatingD2",
  "wgCommunicatingD3", "wgP0", "wgP1", "wgP2", "wgP3", "wgPM"),
  replicates = 399)

Arguments

x
Indicators dataset produced by createOP with primary sampling unit (PSU) in column named PSU

w
A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

statistic
A function operating on data in x; fixed to bootClassic function for means

params
Parameters (named columns in x) passed to the function specified in statistic; defaults to all variables in indicators data frame except for MUAC

outputColumns
Names of columns in output data frame; ; defaults to all variables in indicators data frame except for MUAC

replicates
Number of bootstrap replicates

Value

Dataframe of boot estimates using bootClassic mean function

Examples

  #
  test <- estimateClassic(x = indicators.ALL,
                          w = testPSU,
                          replicates = 9)

  test

#>          LCL.ALL      EST.ALL      UCL.ALL    LCL.MALES    EST.MALES
#> 1   7.562500e-01 8.697917e-01   0.88437500 8.107108e-01   0.86250000
#> 2   7.395833e-02 1.041667e-01   0.17083333 3.225309e-02   0.08108108
#> 3   1.770833e-02 3.125000e-02   0.05312500 1.445783e-02   0.03846154
#> 4   0.000000e+00 5.208333e-03   0.04062500 0.000000e+00   0.01351351
#> 5   7.012813e+01 7.096875e+01  72.27291667 6.916154e+01  71.26923077
#> 6   0.000000e+00 0.000000e+00   0.00000000 0.000000e+00   0.00000000
#> 7   4.510417e-01 5.260417e-01   0.54479167 4.367315e-01   0.48750000
#> 8   2.114583e-01 2.447917e-01   0.28437500 1.714551e-01   0.25301205
#> 9   1.416667e-01 2.135417e-01   0.26354167 1.014245e-01   0.17073171
#> 10  9.375000e-03 4.687500e-02   0.06666667 1.469582e-02   0.04819277
#> 11  3.229167e-01 3.854167e-01   0.48854167 1.000000e+00   1.00000000
#> 12  5.114583e-01 6.145833e-01   0.67708333 0.000000e+00   0.00000000
#> 13  2.083333e-03 4.687500e-02   0.05520833 0.000000e+00   0.01282051
#> 14  2.093750e-01 2.968750e-01   0.35416667 5.182553e-01   0.60256410
#> 15  7.708333e-02 1.093750e-01   0.16458333 6.053784e-02   0.14457831
#> 16  2.500000e-02 6.770833e-02   0.08229167 5.384615e-02   0.08433735
#> 17  4.572917e-01 4.895833e-01   0.57083333 1.060394e-01   0.15662651
#> 18  0.000000e+00 0.000000e+00   0.00000000 0.000000e+00   0.00000000
#> 19  1.000000e-01 1.354167e-01   0.18333333 6.324786e-02   0.15384615
#> 20  2.429167e+00 2.567708e+00   2.84375000 2.277805e+00   2.54054054
#> 21  4.313542e+00 4.593750e+00   4.72812500 4.184615e+00   4.41463415
#> 22  8.750000e-01 9.218750e-01   0.94166667 8.546835e-01   0.90361446
#> 23  3.625000e-01 5.000000e-01   0.58750000 3.998764e-01   0.42682927
#> 24  4.958333e-01 5.520833e-01   0.62291667 4.758599e-01   0.56626506
#> 25  2.395833e-02 5.208333e-02   0.10208333 0.000000e+00   0.02564103
#> 26  7.291667e-03 3.125000e-02   0.05937500 1.232020e-02   0.04054054
#> 27  2.531250e-01 3.281250e-01   0.39791667 3.724359e-01   0.43373494
#> 28  3.572917e-01 4.270833e-01   0.49166667 3.341365e-01   0.39759036
#> 29  1.145833e-02 2.604167e-02   0.06875000 0.000000e+00   0.00000000
#> 30  1.822917e-01 2.291667e-01   0.25520833 1.397590e-01   0.22222222
#> 31  4.489583e-01 5.156250e-01   0.60104167 3.452126e-01   0.40243902
#> 32  9.583333e-01 9.583333e-01   0.98958333 9.261159e-01   0.97297297
#> 33  3.979167e-01 5.000000e-01   0.56041667 3.998764e-01   0.42168675
#> 34  3.572917e-01 4.270833e-01   0.49166667 3.341365e-01   0.39759036
#> 35  8.437500e-02 1.041667e-01   0.17291667 1.744841e-02   0.08641975
#> 36  5.270833e-01 5.937500e-01   0.65625000 5.012658e-01   0.56250000
#> 37  4.270833e-02 5.729167e-02   0.08229167 1.488430e-02   0.05063291
#> 38  5.552083e-01 6.250000e-01   0.69479167 5.246404e-01   0.60000000
#> 39  6.031250e-01 6.875000e-01   0.71770833 5.628205e-01   0.63291139
#> 40  1.145833e-02 2.604167e-02   0.06875000 0.000000e+00   0.00000000
#> 41  5.677083e-01 6.406250e-01   0.67291667 5.512512e-01   0.66666667
#> 42  6.281250e-01 6.875000e-01   0.72187500 5.741736e-01   0.67901235
#> 43  7.958333e-01 8.281250e-01   0.85208333 6.917379e-01   0.80722892
#> 44  5.677083e-01 6.406250e-01   0.67291667 5.512512e-01   0.66666667
#> 45  8.406250e-01 8.593750e-01   0.89375000 8.021625e-01   0.89873418
#> 46  3.385417e-01 3.802083e-01   0.45729167 4.004487e-01   0.46913580
#> 47  3.260417e-01 3.750000e-01   0.44479167 4.004487e-01   0.45945946
#> 48  6.875000e-01 7.916667e-01   0.88437500 7.058747e-01   0.72289157
#> 49  1.000000e-01 1.562500e-01   0.25104167 1.790854e-01   0.27027027
#> 50  1.562500e-02 3.125000e-02   0.06145833 0.000000e+00   0.00000000
#> 51  9.166667e-01 9.687500e-01   0.98437500 8.962296e-01   0.95121951
#> 52  9.510417e-01 9.739583e-01   0.99791667 9.227895e-01   0.96385542
#> 53  9.510417e-01 9.739583e-01   0.99791667 9.227895e-01   0.96385542
#> 54  9.072917e-01 9.479167e-01   0.97395833 9.039726e-01   0.96153846
#> 55  6.979167e-01 7.135417e-01   0.77812500 6.580167e-01   0.71951220
#> 56  9.843750e-01 9.947917e-01   0.99895833 9.254966e-01   0.98717949
#> 57  5.488542e+00 5.572917e+00   5.66041667 5.305308e+00   5.55421687
#> 58  9.489583e-01 9.583333e-01   0.98437500 9.227895e-01   0.96385542
#> 59  2.083333e-03 1.562500e-02   0.03333333 0.000000e+00   0.00000000
#> 60  1.041667e-03 1.562500e-02   0.03541667 2.500000e-03   0.03614458
#> 61  5.114583e-01 5.781250e-01   0.61354167 3.774234e-01   0.52702703
#> 62  9.062500e-02 1.250000e-01   0.16666667 1.035127e-01   0.20512821
#> 63  1.150521e+01 1.233854e+01  13.24895833 9.973219e+00  11.21794872
#> 64  4.427083e-01 5.156250e-01   0.56145833 3.564411e-01   0.44578313
#> 65  1.625000e-01 1.979167e-01   0.29479167 6.160100e-02   0.13924051
#> 66  4.093750e-01 4.479167e-01   0.48437500 2.907848e-01   0.38750000
#> 67  7.275606e-01 8.045977e-01   0.87440530 5.240143e-01   0.69444444
#> 68  1.176471e-02 1.666667e-01   0.41923077 0.000000e+00   0.22222222
#> 69  3.097902e-01 4.375000e-01   0.66769231 1.252525e-01   0.33333333
#> 70  1.111111e-02 1.000000e-01   0.27165775 0.000000e+00   0.00000000
#> 71  1.250000e-02 1.000000e-01   0.39893048 1.833333e-01   0.33333333
#> 72  0.000000e+00 0.000000e+00   0.00000000 0.000000e+00   0.00000000
#> 73  0.000000e+00 0.000000e+00   0.00000000 0.000000e+00   0.00000000
#> 74  0.000000e+00 0.000000e+00   0.06176471 0.000000e+00   0.00000000
#> 75  0.000000e+00 0.000000e+00   0.00000000 0.000000e+00   0.00000000
#> 76  0.000000e+00 1.111111e-01   0.22459893 0.000000e+00   0.14285714
#> 77  8.083333e-01 8.750000e-01   0.89583333 7.827211e-01   0.87179487
#> 78  7.940039e-01 8.571429e-01   0.88391538 6.426690e-01   0.73913043
#> 79  8.333333e-03 1.578947e-01   0.28375000 0.000000e+00   0.13333333
#> 80  5.312500e-01 7.500000e-01   0.90000000 6.217195e-01   0.76923077
#> 81  0.000000e+00 0.000000e+00   0.00000000 0.000000e+00   0.00000000
#> 82  0.000000e+00 4.166667e-02   0.12500000 8.000000e-03   0.10000000
#> 83  0.000000e+00 0.000000e+00   0.00000000 0.000000e+00   0.00000000
#> 84  0.000000e+00 5.555556e-02   0.14500000 0.000000e+00   0.00000000
#> 85  0.000000e+00 0.000000e+00   0.00000000 0.000000e+00   0.00000000
#> 86  0.000000e+00 0.000000e+00   0.02857143 0.000000e+00   0.00000000
#> 87  0.000000e+00 0.000000e+00   0.00000000 0.000000e+00   0.00000000
#> 88  4.645833e-01 5.312500e-01   0.59687500 5.993162e-01   0.69230769
#> 89  2.968750e-01 3.385417e-01   0.41458333 4.227108e-01   0.53164557
#> 90  8.437500e-02 1.145833e-01   0.18333333 1.715251e-01   0.23076923
#> 91  5.208333e-03 1.562500e-02   0.02916667 2.702703e-03   0.06024096
#> 92  2.500000e-02 5.208333e-02   0.09166667 2.469136e-03   0.02702703
#> 93  0.000000e+00 5.208333e-03   0.01562500 0.000000e+00   0.00000000
#> 94  1.041667e-03 1.562500e-02   0.02604167 0.000000e+00   0.03797468
#> 95  1.145833e-02 3.125000e-02   0.05937500 2.409639e-03   0.02531646
#> 96  2.541667e-01 3.281250e-01   0.36354167 1.855651e-01   0.27710843
#> 97  1.041667e-03 5.208333e-03   0.01979167 0.000000e+00   0.00000000
#> 98  5.697917e-01 5.885417e-01   0.67708333 4.831173e-01   0.58750000
#> 99  6.729167e-01 6.927083e-01   0.77187500 5.231356e-01   0.63750000
#> 100 1.822917e-01 2.604167e-01   0.33854167 1.830380e-01   0.28395062
#>        UCL.MALES  LCL.FEMALES  EST.FEMALES  UCL.FEMALES         INDICATOR
#> 1     0.90313848 8.186462e-01 8.898305e-01   0.92483705             resp1
#> 2     0.11433426 5.664449e-02 9.322034e-02   0.14821201             resp2
#> 3     0.08698475 1.028638e-02 1.818182e-02   0.03936081             resp3
#> 4     0.03475783 0.000000e+00 0.000000e+00   0.01564496             resp4
#> 5    72.53047311 6.896050e+01 7.036607e+01  72.54332635               age
#> 6     0.00000000 0.000000e+00 0.000000e+00   0.00000000           ageGrp1
#> 7     0.59773910 4.629630e-01 5.689655e-01   0.63724138           ageGrp2
#> 8     0.39036729 1.494709e-01 2.181818e-01   0.27113821           ageGrp3
#> 9     0.26076614 1.463835e-01 1.896552e-01   0.33983229           ageGrp4
#> 10    0.12250695 0.000000e+00 9.433962e-03   0.05720443           ageGrp5
#> 11    1.00000000 0.000000e+00 0.000000e+00   0.00000000              sex1
#> 12    0.00000000 1.000000e+00 1.000000e+00   1.00000000              sex2
#> 13    0.03578515 8.565375e-03 4.310345e-02   0.06202020          marital1
#> 14    0.67445394 5.804598e-02 1.525424e-01   0.23429467          marital2
#> 15    0.18999073 4.698276e-02 1.000000e-01   0.12979831          marital3
#> 16    0.12006098 1.286195e-02 4.310345e-02   0.06818054          marital4
#> 17    0.23379673 5.773242e-01 6.724138e-01   0.74827586          marital5
#> 18    0.00000000 0.000000e+00 0.000000e+00   0.00000000          marital6
#> 19    0.25800764 4.711864e-02 1.339286e-01   0.16066411             alone
#> 20    2.66734826 2.486706e+00 2.689655e+00   2.77061345                MF
#> 21    4.64550120 4.528097e+00 4.660377e+00   4.87256442               DDS
#> 22    0.96873982 8.620690e-01 8.928571e-01   0.96061427              FG01
#> 23    0.62928498 5.280067e-01 5.714286e-01   0.64617183              FG02
#> 24    0.62428174 5.660219e-01 6.666667e-01   0.74840599              FG03
#> 25    0.06795791 4.910600e-02 8.181818e-02   0.12433642              FG04
#> 26    0.08960843 0.000000e+00 1.785714e-02   0.06837996              FG05
#> 27    0.48088283 2.065993e-01 2.500000e-01   0.30853659              FG06
#> 28    0.49148304 3.359561e-01 4.310345e-01   0.48051584              FG07
#> 29    0.02321552 1.990468e-02 5.084746e-02   0.07305236              FG08
#> 30    0.31041529 1.584656e-01 2.327586e-01   0.29301934              FG09
#> 31    0.45433735 4.306941e-01 5.092593e-01   0.63266814              FG10
#> 32    0.99759036 9.362837e-01 9.827586e-01   0.99837398              FG11
#> 33    0.54186592 4.259248e-01 5.203252e-01   0.56014027       proteinRich
#> 34    0.49148304 3.359561e-01 4.310345e-01   0.48051584          pProtein
#> 35    0.11663871 1.206865e-01 1.428571e-01   0.22251139          aProtein
#> 36    0.63330244 6.082132e-01 6.785714e-01   0.77068966             pVitA
#> 37    0.08960843 4.653313e-02 7.142857e-02   0.11125569             aVitA
#> 38    0.66104418 6.256426e-01 6.910569e-01   0.81206897             xVitA
#> 39    0.73154155 6.235690e-01 7.154472e-01   0.81206897          ironRich
#> 40    0.02321552 1.990468e-02 5.084746e-02   0.07305236            caRich
#> 41    0.76162396 4.815488e-01 5.517241e-01   0.61455811            znRich
#> 42    0.77175054 5.571717e-01 6.120690e-01   0.72126223             vitB1
#> 43    0.88060212 7.267340e-01 8.292683e-01   0.92007389             vitB2
#> 44    0.76162396 4.815488e-01 5.517241e-01   0.61455811             vitB3
#> 45    0.93461538 7.975421e-01 8.644068e-01   0.94162562             vitB6
#> 46    0.52788056 2.563636e-01 3.577236e-01   0.39225182            vitB12
#> 47    0.52411164 2.471380e-01 3.495935e-01   0.38171913       vitBcomplex
#> 48    0.82091463 6.871101e-01 8.103448e-01   0.86798030              HHS1
#> 49    0.29412528 1.041872e-01 1.293103e-01   0.19739837              HHS2
#> 50    0.02891566 5.555556e-03 3.773585e-02   0.12436221              HHS3
#> 51    0.97198718 9.748677e-01 9.913793e-01   1.00000000             ADL01
#> 52    0.99750000 9.911330e-01 1.000000e+00   1.00000000             ADL02
#> 53    0.99750000 9.911330e-01 1.000000e+00   1.00000000             ADL03
#> 54    0.99500000 9.179177e-01 9.636364e-01   0.98985142             ADL04
#> 55    0.79547468 6.757143e-01 7.155172e-01   0.79986988             ADL05
#> 56    1.00000000 1.000000e+00 1.000000e+00   1.00000000             ADL06
#> 57    5.70458861 5.564891e+00 5.689655e+00   5.78353936          scoreADL
#> 58    0.99750000 9.558111e-01 9.913793e-01   1.00000000         classADL1
#> 59    0.00000000 0.000000e+00 8.620690e-03   0.04418886         classADL2
#> 60    0.07721054 0.000000e+00 0.000000e+00   0.00000000         classADL3
#> 61    0.60463627 5.379310e-01 6.272727e-01   0.71761518           hasHelp
#> 62    0.28773556 4.363144e-02 9.821429e-02   0.11147335         unmetNeed
#> 63   12.21871622 1.190197e+01 1.247414e+01  13.46896552                K6
#> 64    0.53993243 4.831815e-01 5.172414e-01   0.55952978            K6Case
#> 65    0.20132530 1.761129e-01 2.372881e-01   0.25646401                DS
#> 66    0.46729206 3.618947e-01 4.454545e-01   0.55898826                H1
#> 67    0.86874459 6.975684e-01 7.800000e-01   0.87487685                H2
#> 68    0.32380952 0.000000e+00 1.111111e-01   0.49350649               H31
#> 69    0.50000000 2.770053e-01 4.285714e-01   0.77222222               H32
#> 70    0.00000000 0.000000e+00 1.818182e-01   0.38655462               H33
#> 71    0.62000000 0.000000e+00 0.000000e+00   0.00000000               H34
#> 72    0.00000000 0.000000e+00 0.000000e+00   0.00000000               H35
#> 73    0.00000000 0.000000e+00 0.000000e+00   0.00000000               H36
#> 74    0.00000000 0.000000e+00 0.000000e+00   0.12222222               H37
#> 75    0.00000000 0.000000e+00 0.000000e+00   0.00000000               H38
#> 76    0.31666667 0.000000e+00 7.142857e-02   0.37714286               H39
#> 77    0.90651838 8.106897e-01 8.965517e-01   0.92539199                H4
#> 78    0.81883446 7.895429e-01 8.541667e-01   0.90007974                H5
#> 79    0.34298643 0.000000e+00 7.142857e-02   0.36210526               H61
#> 80    0.92666667 3.929825e-01 8.181818e-01   0.92857143               H62
#> 81    0.00000000 0.000000e+00 0.000000e+00   0.00000000               H63
#> 82    0.17450980 0.000000e+00 0.000000e+00   0.00000000               H64
#> 83    0.00000000 0.000000e+00 0.000000e+00   0.00000000               H65
#> 84    0.00000000 0.000000e+00 1.000000e-01   0.24862745               H66
#> 85    0.00000000 0.000000e+00 0.000000e+00   0.00000000               H67
#> 86    0.00000000 0.000000e+00 0.000000e+00   0.11229947               H68
#> 87    0.00000000 0.000000e+00 0.000000e+00   0.00000000               H69
#> 88    0.71795593 4.150407e-01 5.000000e-01   0.54153605                M1
#> 89    0.60708017 2.290580e-01 3.240741e-01   0.38815047               M2A
#> 90    0.35452414 1.851852e-03 3.448276e-02   0.11812954               M2B
#> 91    0.11089233 0.000000e+00 9.090909e-03   0.01767554               M2C
#> 92    0.04852564 3.775971e-02 1.034483e-01   0.11976601               M2D
#> 93    0.00000000 0.000000e+00 8.620690e-03   0.03708523               M2E
#> 94    0.08600327 0.000000e+00 0.000000e+00   0.00000000               M2F
#> 95    0.05078174 0.000000e+00 9.259259e-03   0.04153605               M2G
#> 96    0.38814449 2.849936e-01 3.089431e-01   0.37758621               M2H
#> 97    0.02507244 0.000000e+00 0.000000e+00   0.01561129               M2I
#> 98    0.66504065 5.124521e-01 6.016949e-01   0.62890282                W1
#> 99    0.73633521 6.386467e-01 7.090909e-01   0.77190917                W2
#> 100   0.35180723 1.505747e-01 2.500000e-01   0.29969125                W3
#>  [ reached getOption("max.print") -- omitted 37 rows ]