Estimate all standard RAM-OP indicators

Estimate all standard RAM-OP indicators

Usage

estimate_op_all(
  x,
  w,
  indicators = c("demo", "anthro", "food", "hunger", "adl", "disability", "mental",
    "dementia", "health", "oedema", "screening", "income", "wash", "visual", "misc"),
  replicates = 399
)

Arguments

x

Indicators dataset produced by create_op_all with primary sampling unit (PSU) in column named PSU

w

A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

indicators

A character vector of indicator set names to estimate. Indicator set names are demo, anthro, food, hunger, disability, adl, mental, dementia, health, income, wash, visual, and misc. Default is all indicator sets.

replicates

Number of bootstrap replicates. Default is 399.

Value

Tibble of boot estimates for all specified standard RAM-OP indicators

Examples

estimate_op_all(x = create_op_all(testSVY),
                w = testPSU,
                replicates = 9)
#> # A tibble: 139 × 13
#>    INDICATOR GROUP   LABEL TYPE  EST.ALL LCL.ALL UCL.ALL EST.M…¹ LCL.M…² UCL.M…³
#>    <fct>     <fct>   <fct> <fct>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>   <dbl>
#>  1 resp1     Survey  Resp… Prop…   84.4    80.8    88.6    84      73.7    87.9 
#>  2 resp2     Survey  Resp… Prop…    9.90    7.5    14.2     9.33    5.87   15.7 
#>  3 resp3     Survey  Resp… Prop…    3.65    3.23    6.25    4.88    2.37    7.83
#>  4 resp4     Survey  Resp… Prop…    1.04    0       1.98    2.67    0       7.62
#>  5 age       Demogr… Mean… Mean    70.8    69.6    73.6    72.1    70.6    73.3 
#>  6 ageGrp1   Demogr… Self… Prop…    0       0       0       0       0       0   
#>  7 ageGrp2   Demogr… Self… Prop…   53.6    40.1    62.9    45.3    36.2    56.4 
#>  8 ageGrp3   Demogr… Self… Prop…   23.4    16.5    27.5    28.2    19.5    39.7 
#>  9 ageGrp4   Demogr… Self… Prop…   20.8    12.0    36.5    20.7    11.2    28.2 
#> 10 ageGrp5   Demogr… Self… Prop…    3.12    1.67    7.19    4.71    2.52   10.1 
#> # … with 129 more rows, 3 more variables: EST.FEMALES <dbl>, LCL.FEMALES <dbl>,
#> #   UCL.FEMALES <dbl>, and abbreviated variable names ¹​EST.MALES, ²​LCL.MALES,
#> #   ³​UCL.MALES