Estimate all standard RAM-OP indicators

Estimate all standard RAM-OP indicators

estimate_op_all(
  x,
  w,
  indicators = c("demo", "anthro", "food", "hunger", "adl", "disability", "mental",
    "dementia", "health", "oedema", "screening", "income", "wash", "visual", "misc"),
  replicates = 399
)

Arguments

x
Indicators dataset produced by create_op_all with primary sampling unit (PSU) in column named PSU

w
A data frame with primary sampling unit (PSU) in column named psu and survey weight (i.e. PSU population) in column named pop

indicators
A character vector of indicator set names to estimate. Indicator set names are demo, anthro, food, hunger, disability, adl, mental, dementia, health, income, wash, visual, and misc. Default is all indicator sets.

replicates
Number of bootstrap replicates. Default is 399.

Value

Tibble of boot estimates for all specified standard RAM-OP indicators

Examples

estimate_op_all(x = create_op_all(testSVY),
                w = testPSU,
                replicates = 9)

#> # A tibble: 139 x 13
#>    INDICATOR GROUP LABEL TYPE  EST.ALL LCL.ALL UCL.ALL EST.MALES LCL.MALES
#>    <fct>     <fct> <fct> <fct>   <dbl>   <dbl>   <dbl>     <dbl>     <dbl>
#>  1 resp1     Surv… Resp… Prop…   87.0   81.5     91.7      84        75.8 
#>  2 resp2     Surv… Resp… Prop…    7.81   5.83    10.9       7.89      2.48
#>  3 resp3     Surv… Resp… Prop…    3.65   0.729    6.67      6.33      1.55
#>  4 resp4     Surv… Resp… Prop…    1.04   0        2.08      1.27      0   
#>  5 age       Demo… Mean… Mean    70.2   69.0     71.7      70.2      69.1 
#>  6 ageGrp1   Demo… Self… Prop…    0      0        0         0         0   
#>  7 ageGrp2   Demo… Self… Prop…   54.2   48.0     62.2      50        39.6 
#>  8 ageGrp3   Demo… Self… Prop…   22.4   19.5     33.3      27.1      17.2 
#>  9 ageGrp4   Demo… Self… Prop…   18.8   11.5     26.1      18.4       6.86
#> 10 ageGrp5   Demo… Self… Prop…    2.08   0.729    4.06      5.33      1.46
#> # … with 129 more rows, and 4 more variables: UCL.MALES <dbl>,
#> #   EST.FEMALES <dbl>, LCL.FEMALES <dbl>, UCL.FEMALES <dbl>