RAM-OP Survey Dataset

Dataset collected from a RAM-OP survey conducted in Addis Ababa, Ethiopia in early 2014

Usage

testSVY

Format

A data frame with 91 columns and 192 rows:

ad2

Team number

psu

PSU (cluster) number

hh

Household identifier

id

Person identifier

d1

Who is answering these questions?

d2

Age in years

d3

Sex

d4

Marital status

d5

Do you live alone?

f1

How many meals did you eat since this time yesterday?

f2a

Tinned, powdered or fresh milk?

f2b

Sweetened or flavoured water, soda drink, alcoholic drink, beer, tea or infusion, coffee, soup, or broth?

f2c

Any food made from grain such as millet, wheat, barley, sorghum, rice, maize, pasta, noodles, bread, pizza, porridge?

f2d

Any food made from fruits or vegetables that have yellow or orange flesh such as carrots, pumpkin, red sweet potatoes, mangoes, and papaya?

f2e

Any food made with red palm oil or red palm nuts?

f2f

Any dark green leafy vegetables such as cabbage, broccoli, spinach, moringa leaves, cassava leaves?

f2g

Any food made from roots or tubers such as white potatoes, white yams, false banana, cassava, manioc, onions, beets, turnips, and swedes?

f2h

Any food made from lentils, beans, peas, groundnuts, nuts, or seeds?

f2i

Any other fruits or vegetables such as banana, plantain, avocado, cauliflower, coconut?

f2j

Liver, kidney, heart, black pudding, blood, or other organ meats?

f2k

Any meat such as beef, pork, goat, lamb, mutton, veal, chicken, camel, or bush meat?

f2l

Fresh or dried fish, shellfish, or seafood?

f2m

Cheese, yoghurt, or other milk products?

f2n

Eggs?

f2o

Any food made with oil, fat, butter, or ghee?

f2p

Any mushrooms or fungi?

f2q

Grubs, snails, insects?

f2r

Sugar, honey and foods made with sugar or honey such as sweets, candies, chocolate, cakes, and biscuits?

f2s

Salt, pepper, herbs, spices, or sauces (hot sauce, soy sauce, ketchup)?

f3

In the past four weeks, how often was there ever no food to eat of any kind in your home because of lack of resources to get food?

f4

In the past four weeks, how often did you go to sleep at night hungry because there was not enough food?

f5

In the past four weeks, how often did you go a whole day and night without eating anything at all because there was not enough food?

f6

Are you or anyone in your household receiving a food ration on a regular basis?

f7

Have you or another member of your household received non-food relief items such as soap, bucket, water container, bedding, mosquito net, clothes, or plastic sheet in the previous four weeks?

a1

Have you or another member of your household received non-food relief items such as soap, bucket, water container, bedding, mosquito net, clothes, or plastic sheet in the previous four weeks?

a2

Do you need help getting dressed partially or completely (not including tying of shoes)?

a3

Do you need help going to the toilet or cleaning yourself after using the toilet or do you use a commode or bed-pan?

a4

Do you need someone (i.e. not a walking aid) to help you move from a bed to a chair?

a5

Are you partially or totally incontinent of bowel or bladder?

a6

Do you need partial or total help with eating?

a7

Is someone taking care of you or helping you with everyday activities such as shopping, cooking, bathing and dressing?

a8

Do you have problems chewing food?

k6a

About how often during the past four weeks did you feel nervous – all of the time, most of the time, some of the time, a little of the time, or none of the time?

k6b

During the past four weeks, about how often did you feel hopeless – all of the time, most of the time, some of the time, a little of the time, or none of the time?

k6c

During the past four weeks, about how often did you feel restless or fidgety – all of the time, most of the time, some of the time, a little of the time, or none of the time?

k6d

During the past four weeks, about how often did you feel so depressed that nothing could cheer you up – all of the time, most of the time, some of the time, a little of the time, or none of the time?

k6e

During the past four weeks, about how often did you feel that everything was an effort – all of the time, most of the time, some of the time, a little of the time, or none of the time?

k6f

During the past four weeks, about how often did you feel worthless – all of the time, most of the time, some of the time, a little of the time, or none of the time?

ds1

Point to nose and ask "What do you call this?"

ds2

What do you do with a hammer?

ds3

What day of the week is it?

ds4

What is the season?

ds5

Please point first to the window and then to the door.

ds6a

Child

ds6b

House

ds6c

Road

h1

Do you suffer from a long term disease that requires you to take regular medication?

h2

Do you take drugs regularly for this?

h3

Why not?

h4

Have you been ill in the past two weeks?

h5

Did you go to the pharmacy, dispensary, health centre, health post, clinic, or hospital?

h6

Why not?

m1

Do you have a personal source of income or money?

m2a

Where does your income or money come from?: Agriculture, livestock, or fishing

m2b

Where does your income or money come from?: Wages or salary

m2c

Where does your income or money come from?: Sale of charcoal, bricks, firewood, poles, etc.

m2d

Where does your income or money come from?: Trading (e.g. market, shop)

m2e

Where does your income or money come from?: Private pension, investments, interest, rents, etc.

m2f

Where does your income or money come from?: Spending savings; Sale of household goods, personal goods, or jewellery; Sale of livestock, land, or other assets

m2g

Where does your income or money come from?: Aid, gifts, charity (e.g. from church, mosque, temple), begging, borrowing, or sale of food aid or relief items

m2h

Where does your income or money come from?: Cash transfer (NGO, UNO, government); State pension, social security, benefits, welfare program

m2i

Where does your income or money come from?: Other

w1

What is your main source of drinking water?

w2

What do you usually do to the water to make it safer to drink?

w3

What kind of toilet facility do members of your household usually use?

w4

Do you share this toilet facility with other households?

as1

Mid-upper arm circumference (mm)

as2

Has someone measured your arm like this in the previous month?

as3

Bilateral pitting oedema

as4

Has someone examined your feet like this in the previous month?

va2a

Tumbling Es: first time

va2b

Tumbling Es: second time

va2c

Tumbling Es: third time

va2d

Tumbling Es: fourth time

wg1

Do you have difficulty seeing, even if wearing glasses?

wg2

Do you have difficulty hearing, even if using a hearing aid?

wg3

Do you have difficulty walking or climbing steps?

wg4

Do you have difficulty remembering or concentrating?

wg5

Do you have difficulty with self-care such as washing all over or dressing?

wg6

Using your usual (customary) language, do you have difficulty communicating, for example understanding or being understood?

Examples

testSVY
#> # A tibble: 192 × 90
#>      ad2   psu    hh    id    d1    d2    d3    d4    d5    f1   f2a   f2b   f2c
#>    <int> <int> <int> <int> <int> <int> <int> <int> <int> <int> <int> <int> <int>
#>  1     1   201     1     1     1    67     2     5     2     3     2     1     1
#>  2     1   201     2     1     1    74     1     2     2     3     2     1     1
#>  3     1   201     3     1     1    60     1     2     2     2     2     2     2
#>  4     1   201     3     2     1    60     2     2     2     3     2     2     1
#>  5     1   201     4     1     1    85     2     5     2     3     2     1     1
#>  6     1   201     5     1     2    86     1     5     1     4     2     1     1
#>  7     1   201     6     1     1    80     1     5     2     3     2     1     1
#>  8     1   201     6     2     1    60     2     5     2     3     2     2     1
#>  9     1   201     7     1     1    62     1     2     2     2     2     1     1
#> 10     1   201     8     1     1    72     2     5     2     2     2     1     1
#> # … with 182 more rows, and 77 more variables: f2d <int>, f2e <int>, f2f <int>,
#> #   f2g <int>, f2h <int>, f2i <int>, f2j <int>, f2k <int>, f2l <int>,
#> #   f2m <int>, f2n <int>, f2o <int>, f2p <int>, f2q <int>, f2r <int>,
#> #   f2s <int>, f3 <int>, f4 <int>, f5 <int>, f6 <int>, f7 <int>, a1 <int>,
#> #   a2 <int>, a3 <int>, a4 <int>, a5 <int>, a6 <int>, a7 <int>, a8 <int>,
#> #   k6a <int>, k6b <int>, k6c <int>, k6d <int>, k6e <int>, k6f <int>,
#> #   ds1 <int>, ds2 <int>, ds3 <int>, ds4 <int>, ds5 <int>, ds6a <int>, …